% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate.R
\name{accumulate}
\alias{accumulate}
\title{FUNCTION_TITLE}
\usage{
accumulate(mean, sd, months = c(1, 11, 24, 24), nsample = 1000)
}
\arguments{
\item{mean}{PARAM_DESCRIPTION}

\item{sd}{PARAM_DESCRIPTION}

\item{months}{PARAM_DESCRIPTION, Default: c(1, 11, 24, 24)}

\item{nsample}{PARAM_DESCRIPTION, Default: 1000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
