% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ace.r}
\alias{ace.r}
\title{Fitted model objects for estimating the memory of parallel_model}
\format{An object of class \code{areg.boot} of length 21.}
\usage{
ace.r
}
\description{
Fitted model objects for estimating the memory of parallel_model
}
\keyword{datasets}
