% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad2_folds.R
\name{ad2_folds}
\alias{ad2_folds}
\title{FUNCTION_TITLE}
\usage{
ad2_folds(admin_raster, shape_ident = "gaul_code", admin_shps, ss = 1,
  xy, n_folds, mask_shape, mask_raster, ...)
}
\arguments{
\item{admin_raster}{PARAM_DESCRIPTION}

\item{shape_ident}{PARAM_DESCRIPTION, Default: 'gaul_code'}

\item{admin_shps}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION, Default: 1}

\item{xy}{PARAM_DESCRIPTION}

\item{n_folds}{PARAM_DESCRIPTION}

\item{mask_shape}{PARAM_DESCRIPTION}

\item{mask_raster}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
