% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_additional_terms.R
\name{add_additional_terms}
\alias{add_additional_terms}
\title{FUNCTION_TITLE}
\usage{
add_additional_terms(fes, add_terms = NULL)
}
\arguments{
\item{fes}{PARAM_DESCRIPTION}

\item{add_terms}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
