% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gauls_regions.R
\name{add_gauls_regions}
\alias{add_gauls_regions}
\title{FUNCTION_TITLE}
\usage{
add_gauls_regions(df, simple_raster, shapefile_version = "current")
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{simple_raster}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
