% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_goal.R
\name{add_goal}
\alias{add_goal}
\title{Adds a goal to a goal object, or creates a new one}
\usage{
add_goal(goal_obj = NULL, target_year, target, target_type, abs_rel,
  baseline_year = NA, proj = TRUE, pred_type = c("cell", "admin"))
}
\arguments{
\item{goal_obj}{An existing goal object. If \code{NULL}, will create a new goal object.}

\item{target_year}{Year for which the goal is meant. Must be a future (projected) year
and the year projections must already be generated for this target_year using
\code{make_proj()}}

\item{target}{Target (e.g. 0.8 for 80\% vaccine coverage); numeric}

\item{target_type}{Type of target ('greater' or 'less'); character}

\item{abs_rel}{Is this an 'absolute' or 'relative' goal?}

\item{baseline_year}{For relative goals, what year should we compare to?}

\item{proj}{Placeholder - for now, always use projected years.  In the future this
may support modeled years as well, but still need to code}

\item{pred_type}{Type of aroc to create. Options include \code{cell}, \code{admin},
or \code{c('cell', 'admin')} for both.}
}
\value{
A "goal object" (data.table in a specific format for use in
  \code{compare_to_target()})
}
\description{
This function creates or modifies a "goal object" - which is just a data table
that contains all of the information in each row to check a given target.  This
setup is useful in case you want to check a series of different targets, or
run the same analysis for a series of target years.  (Note that all target
years must have projection objects created by first running \code{make_proj()})
}
\examples{
\dontrun{
# Define goals: start by initializing goal object
goals <- add_goal(
  target_year = 2030,
  target = 0.8,
  target_type = "greater",
  abs_rel = "absolute",
  pred_type = c("cell", "admin")
)

# Add goals to existing goal object by specifying goal_obj
goals <- add_goal(
  goal_obj = goals,
  target_year = 2020,
  target = 0.8,
  target_type = "greater",
  abs_rel = "absolute",
  pred_type = c("cell", "admin")
)
}
}
