% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabels.R
\name{addLabels}
\alias{addLabels}
\title{FUNCTION_TITLE}
\usage{
addLabels(rate, year, country, countries = NULL, col = grey(0.7),
  gap = diff(range(rate))/60, cex = 0.7, adj = 0, xpd = NA, ...)
}
\arguments{
\item{rate}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{country}{PARAM_DESCRIPTION}

\item{countries}{PARAM_DESCRIPTION, Default: NULL}

\item{col}{PARAM_DESCRIPTION, Default: grey(0.7)}

\item{gap}{PARAM_DESCRIPTION, Default: diff(range(rate))/60}

\item{cex}{PARAM_DESCRIPTION, Default: 0.7}

\item{adj}{PARAM_DESCRIPTION, Default: 0}

\item{xpd}{PARAM_DESCRIPTION, Default: NA}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[plotrix]{spreadout}}
}
