% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLines.R
\name{addLines}
\alias{addLines}
\title{FUNCTION_TITLE}
\usage{
addLines(rate, year, country, countries = NULL, col = grey(0.7),
  size = 1, border = grey(0.4))
}
\arguments{
\item{rate}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{country}{PARAM_DESCRIPTION}

\item{countries}{PARAM_DESCRIPTION, Default: NULL}

\item{col}{PARAM_DESCRIPTION, Default: grey(0.7)}

\item{size}{PARAM_DESCRIPTION, Default: 1}

\item{border}{PARAM_DESCRIPTION, Default: grey(0.4)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
