% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQuotes.R
\name{addQuotes}
\alias{addQuotes}
\title{FUNCTION_TITLE}
\usage{
addQuotes(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
