% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_equation_categorical.R
\name{adjust_equation_categorical}
\alias{adjust_equation_categorical}
\title{FUNCTION_TITLE}
\usage{
adjust_equation_categorical(effects_equation = fixed_effects,
  cat_var = "", cat_ras)
}
\arguments{
\item{effects_equation}{PARAM_DESCRIPTION, Default: fixed_effects}

\item{cat_var}{PARAM_DESCRIPTION, Default: ''}

\item{cat_ras}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
function to update the fixed effects variable
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
