% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_sex_merge.R
\name{age_sex_merge}
\alias{age_sex_merge}
\title{FUNCTION_TITLE}
\usage{
age_sex_merge(df1, df2)
}
\arguments{
\item{df1}{PARAM_DESCRIPTION}

\item{df2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
