% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_sex_spec.R
\name{age_sex_spec}
\alias{age_sex_spec}
\title{FUNCTION_TITLE}
\usage{
age_sex_spec(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
