% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agg.R
\name{Agg}
\alias{Agg}
\title{FUNCTION_TITLE}
\usage{
Agg(p_i, N_i)
}
\arguments{
\item{p_i}{PARAM_DESCRIPTION}

\item{N_i}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
