% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_vars.R
\name{agg_vars}
\alias{agg_vars}
\title{FUNCTION_TITLE}
\usage{
agg_vars(df, location_set_version_id, vars, parent_ids)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{location_set_version_id}{PARAM_DESCRIPTION}

\item{vars}{PARAM_DESCRIPTION}

\item{parent_ids}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
