% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_counts.R
\name{aggregate_counts}
\alias{aggregate_counts}
\title{Aggregate raked linked cell pred}
\usage{
aggregate_counts(linked_cell_pred, overs, raked)
}
\arguments{
\item{raked}{Boolean, raked or unraked cell draws}

\item{raked_linked_cell_pred}{a raked and linked cell pred object}

\item{year_list}{list of years in cell_pred}
}
\value{
a list with 3 data.tables for admins0-2 with draws intact
}
\description{
rake linked cell pred to admins0-2, maintaining draws
}
