% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_stackers_admin0.R
\name{aggregate_stackers_admin0}
\alias{aggregate_stackers_admin0}
\title{Create national estimates from stackes}
\usage{
aggregate_stackers_admin0(indicator, indicator_group, run_date, age,
  holdout, regions, year_list, pop_measure, stackers_logit_transform,
  predictor_logit_transform, results_file = NULL,
  shapefile_version = "current")
}
\arguments{
\item{indicator}{indicator}

\item{indicator_group}{indicator_group}

\item{run_date}{model run date}

\item{age}{age group}

\item{holdout}{model holdout number}

\item{regions}{vector of model regions}

\item{year_list}{vector of modeled years}

\item{pop_measure}{population measure to use for population-weighting}

\item{stackers_logit_transform}{logical. Were stackers logit-transformed when fitting the final model?}

\item{predictor_logit_transform}{logical. Was the linear predictor logit-transformed (eg, as is
typically the case in binomial models)?}

\item{results_file}{file path to a .csv file where results should be saved. Optional.}

\item{shapefile_version}{string specifies which shapefile version to pull}
}
\value{
A table of national-level estimates for each country, year, and stacker. If results_file
        is provided, this information is saved to the specified file. If results_file is not
        provided, this information is returned as a data.table.
}
\description{
Create national estimates from stackers by population-weighting the pixel estimates.
}
