% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_validation.R
\name{aggregate_validation}
\alias{aggregate_validation}
\title{FUNCTION_TITLE}
\usage{
aggregate_validation(holdoutlist = stratum_qt,
  cell_draws_filename = "\%s_cell_draws_eb_bin\%i_\%s_\%i\%s.RData",
  years = c(2000, 2005, 2010, 2015), indicator_group, indicator,
  run_date, reg, holdout, vallevel = "", addl_strat = c(age = 0),
  sbh_wgt = NULL, returnaggonly = TRUE, iter = i)
}
\arguments{
\item{holdoutlist}{PARAM_DESCRIPTION, Default: stratum_qt}

\item{cell_draws_filename}{PARAM_DESCRIPTION, Default: '\%s_cell_draws_eb_bin\%i_\%s_\%i\%s.RData'}

\item{years}{PARAM_DESCRIPTION, Default: \code{c(2000, 2005, 2010, 2015)}}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{reg}{PARAM_DESCRIPTION}

\item{holdout}{PARAM_DESCRIPTION}

\item{vallevel}{PARAM_DESCRIPTION, Default: ''}

\item{addl_strat}{PARAM_DESCRIPTION, Default: \code{c(age = 0)}}

\item{sbh_wgt}{PARAM_DESCRIPTION, Default: NULL}

\item{returnaggonly}{PARAM_DESCRIPTION, Default: TRUE}

\item{iter}{PARAM_DESCRIPTION, Default: i}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
