% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fractional_rf.R
\name{apply_fractional_rf}
\alias{apply_fractional_rf}
\title{Apply Fractional Raking Factors}
\usage{
apply_fractional_rf(cell_pred, link, fractional_rf, overs)
}
\arguments{
\item{cell_pred}{a cell pred with deaths/cases}

\item{link}{link table generated by `build_link_table()`}

\item{fractional_rf}{a data.table generated by `get_fractional_rf`}

\item{overs}{a vector of the column names in the cell pred corresponding to draws
(typically V1:Vndraws)}
}
\value{
a linked cell pred with the rf applied
}
\description{
Apply raking factors considering cases where a pixel is in
multiple admin units for deaths or cases
}
