% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_raking_factors.R
\name{apply_raking_factors}
\alias{apply_raking_factors}
\title{Apply raking factors}
\usage{
apply_raking_factors(cell_pred, simple_raster, rake_dt, rake_method,
  force_simp_ras_dt_match = T)
}
\arguments{
\item{cell_pred}{Matrix. Cell_pred object}

\item{simple_raster}{Raster. Should be the rasterized version of a shapefile or otherwise denoting the gaul/admin codes specified in rake_dt}

\item{rake_dt}{data.table Data.table output of calculate_raking_factors. Or at the very least, a three column data.table with columns for loc, year and raking_factor.}

\item{rake_method}{character string. Determines whether logit raking or linear scaling occurs}

\item{force_simp_ras_dt_match}{logical. Determines whether the function should break if it detects a difference between number of country years models and number of country years
supplied to rake.}
}
\value{
\code{cell_pred}
}
\description{
A function to apply raking factors to a cell pred
}
