% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_match.R
\name{approx_match}
\alias{approx_match}
\title{FUNCTION_TITLE}
\usage{
approx_match(country)
}
\arguments{
\item{country}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
