% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_ratio_plot.R
\name{aspect_ratio_plot}
\alias{aspect_ratio_plot}
\title{FUNCTION_TITLE}
\usage{
aspect_ratio_plot(plt, aspect_ratio = 1, expand = 0.05)
}
\arguments{
\item{plt}{PARAM_DESCRIPTION}

\item{aspect_ratio}{PARAM_DESCRIPTION, Default: 1}

\item{expand}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
