% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brt_covs.R
\name{brt_covs}
\alias{brt_covs}
\title{BRT Covariates}
\usage{
brt_covs(df, indicator_family = indicator_family, lcovs = cov_layers,
  years = c(2000, 2005, 2010, 2015), weight = NULL, tc = 4,
  lr = 0.005, bf = 0.75, return_only_raster = TRUE)
}
\arguments{
\item{df}{data.table with variables "latitude", "longitude", "N", and specified indicator}

\item{lcovs}{list of raster layers or bricks (for temporally varying) output of load_and_crop_covariates()}

\item{years}{analysis years. defaults to 2000, 2005, 2010, 2015}

\item{weight}{character of weight variable name in df, defaults to null}

\item{tc}{tree.complexity for BRT, defaults to 4}

\item{lr}{learning.rate for BRT, defaults to 0.005}

\item{bf}{bagging.fraction for BRT, defaults to 0.75}

\item{return_only_raster:}{if TRUE, only returns raster of results, otherwise returns a list of BRT model object and prediction raster}
}
\value{
A raster or a brick of rasters (see above)
}
\description{
Takes Covariates and Data and returns Boosted Regression Trees Outputs
}
\note{
depends on \code{gbm} and \code{dismo} packages
}
