% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fixed_effects_logic.R
\name{build_fixed_effects_logic}
\alias{build_fixed_effects_logic}
\title{Build Fixed Effects Logic for MBG}
\usage{
build_fixed_effects_logic(dagobject, admin_codes = NULL)
}
\arguments{
\item{dagobject}{The MBGDag object which contains the MBGPipeline object
along with the config parameters.}

\item{admin_codes}{A vector of GADM codes to be used if we want country
fixed effects in the MBG. Default: \code{NULL}}
}
\value{
A string with the formula for the fixed effects going into the MBG
}
\description{
Check variaous config parameters and determine the fixed effects
to be used in the MBG run
}
