% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_link_polygon.R
\name{build_link_polygon}
\alias{build_link_polygon}
\title{Build Link Polygon}
\usage{
build_link_polygon(region, simple_raster = NULL,
  shapefile_version = "current")
}
\arguments{
\item{region}{character string of region to be pulled from \link{get_adm0_codes()}}

\item{simple_raster}{default \code{NULL}. provides option to pass in simple raster rather
than making it using a region name to save time.}

\item{shapefile_version}{string to identify version of shapefile to build simple_raster}
}
\value{
returns an sf polygon
}
\description{
Builds a polygon where there is a layer for each pixel in a raster,
with an ID field referring to the id of the pixel in the raster
}
\examples{
\dontrun{
idpoly <- build_link_polygon(region)
}

}
