% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_link_table.R
\name{build_link_table}
\alias{build_link_table}
\title{Build Link Table}
\usage{
build_link_table(shapefile_version, cores, region = "stage1+stage2",
  custom_shapefile_path = NULL, custom_shapefile_field = NULL)
}
\arguments{
\item{shapefile_version}{version of admin file to base link table on}

\item{cores}{number of cores to run mclapply over}

\item{region}{the region used to generate the simple raster}

\item{custom_shapefile_path}{A non-admin shapefile to build a link table for}

\item{custom_shapefile_field}{the field in custom_shapefile with admin identifiers}
}
\value{
returns the link data.table described above, and the id raster and polygon used in the calculation
}
\description{
: Builds a link table where each row represents an admin2-pixel, and includes
the admin0-2 levels, the id of the pixel (based on simple raster), and the percent of the
pixel covered by a country. Used for fractional raking where a pixel is in multiple
administrative units.
}
\examples{
\dontrun{
link <- build_link_table(shapefile_version = "2018_08_01", cores = 10, region = "stage1+stage2")

link <- build_link_table(shapefile_version = NULL, cores = 10, region, region = NULL, custom_shapefile_path = "C:/shapefile.shp", custom_shapefile_field = "ADM_CODE")
}

}
