% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mbg_data_stack.R
\name{build_mbg_data_stack}
\alias{build_mbg_data_stack}
\title{Build an INLA data stack for modeling}
\usage{
build_mbg_data_stack(df, fixed_effects, mesh_s, mesh_t, exclude_cs = "",
  usematernnew = F, sig0 = 0.5, rho0 = 0.3, coefs.sum1 = FALSE,
  use_ctry_res = FALSE, use_subnat_res = FALSE,
  remove_non_subnats = FALSE, use_nugget = FALSE,
  stacker_names = child_model_names, yl = year_list, zl = z_list,
  zcol = zcol, scale_gaussian_variance_N = TRUE, tmb = FALSE,
  shapefile_version = "current")
}
\arguments{
\item{df}{data.frame/table containing observational data. must have
latitude and longitude columns as well as extracted values for
all covariates that will be used in modeling}

\item{fixed_effects}{string containing fixed effects used in
res_fit separated by " + "}

\item{mesh_s}{An 'inla.mesh' object used to fit the spatial SPDE GP
approximation in res_fit}

\item{mesh_t}{An 'inla.mesh' object used to fit the temporal
correlation structure in res_fit}

\item{exclude_cs}{vector of strings detailing covariates that are
not center-scaled}

\item{usematernnew}{Logical. if TRUE, used penalized complexity
priors. Not currently suggested...}

\item{sig0}{if usematernnew, median prior value for spde rho}

\item{rho0}{if usematernnew, median prior value for spde sigma}

\item{coefs.sum1}{Logical. If TRUE, add a constraint to ensure
covariate coefs sum to 1 in fitted model}

\item{use_ctry_res}{Logical. If TRUE, include country random
effects}

\item{use_subnat_res}{Logical. If TRUE, include subnational (admin-1) random
effects}

\item{remove_non_subnats}{Logical. If TRUE, NA out all points of the subnational
random effect which are not associated with the country}

\item{use_nugget}{Logical. If TRUE, include a nugget effect}

\item{stacker_names}{string vector of child model names}

\item{yl}{numeric vector of years in model}

\item{zl}{numeric vector of third dimension in kronecker
product. Only implemented in TMB}

\item{zcol}{column name of z values associated with zl. Only
implemented in TMB}

\item{scale_gaussian_variance_N}{Logical. Do you want to scale
gaussian variance by sample size? Only implemented in TMB.}

\item{tmb}{Logical. use tmb?}
}
\value{
List containing 1) 'inla.stack' object, 2) 'inla.spde'
  object, 3) cs_df: a center-scale dataframe containing info on
  mean and SD used to scale covaraites in design matrix
}
\description{
Generates an INLA stack that contains information
  mapping data to various INLA effects (GP, random effects, ...)
}
