% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mbg_data_stack_tmb.R
\name{build_mbg_data_stack_tmb}
\alias{build_mbg_data_stack_tmb}
\title{Build Stack for TMB MBG model}
\usage{
build_mbg_data_stack_tmb(d = df, yl = year_list, zl = z_list,
  fes = all_fixed_effects, indic = indicator,
  country_re = use_inla_country_res, nid_re = use_nid_res,
  exclude_cs = "", nugget = FALSE, zcol = NULL,
  shapefile_version = "current", scale_gaussian_variance_N = TRUE,
  mesh = mesh_s)
}
\arguments{
\item{d}{prepped model frame with no NAs}

\item{yl}{a vector of years for analysis (i.e. c(2001,2002,2003))}

\item{zl}{a vector of zcol for analysis (i.e. ages c(1,2,3,4,5). Must be integers starting with 1)}

\item{fes}{a string of model fixed effects of form 'var1 + var2 + var3' or as string vector
corresponding to column names in d}

\item{indic}{indicator name}

\item{country_re}{TRUE/FALSE include country re. If true and there is a zcol then there will be random slope as well (todo)}

\item{nid_re}{TRUE/FALSE include re on survey. Need nid column in the data frame. defaults to use_nid_res from config.}

\item{exclude_cs}{character vector of covariates to exclude from centrescaling}

\item{mesh}{an inla mesh object. NOTE! this gets remade inside this function... Roy said there was a good reason}

\item{indic}{indicator name, corresponding to the response variable in d}

\item{s2mesh}{Logical. Should the mesh be created on the surface of
a sphere? If TRUE, s2params is used to specify mesh parameters
instead of max_edge and mesh_offset}

\item{s2params}{string of 3 element numeric vector in R
notation. e.g. "c(25, 500, 1000)". The entries describe the
minimum triangle edge length allowed, hos far to extend the mesh
beyond the 'simple' boundary, and the maximum allowed triangle
edge length, respectively. Units are in kilometers. Used only if
s2mesh=TRUE.}
}
\value{
returns a named list with Data and Parameters to be passed into fit_mbg_tmb()

TODO: Allow users to set priors, allow different data model (gaussian to start), country random effects
}
\description{
Organize Data and Parameter Stacks to fit a TMB MBG model
}
\author{
Roy Burstein
}
