% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mbg_formula.R
\name{build_mbg_formula}
\alias{build_mbg_formula}
\title{FUNCTION_TITLE}
\usage{
build_mbg_formula(fixed_effects, positive_constrained_variables = NULL,
  interact_with_year = NULL, int = TRUE, nullmodel = FALSE,
  add_nugget = FALSE, add_ctry_res = FALSE)
}
\arguments{
\item{fixed_effects}{PARAM_DESCRIPTION}

\item{positive_constrained_variables}{PARAM_DESCRIPTION, Default: NULL}

\item{interact_with_year}{PARAM_DESCRIPTION, Default: NULL}

\item{int}{PARAM_DESCRIPTION, Default: TRUE}

\item{nullmodel}{PARAM_DESCRIPTION, Default: FALSE}

\item{add_nugget}{PARAM_DESCRIPTION, Default: FALSE}

\item{add_ctry_res}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\note{
The \code{formula} object was created at the very end of the function after dealing with everything before that in strings.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
