% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mbg_formula_with_priors.R
\name{build_mbg_formula_with_priors}
\alias{build_mbg_formula_with_priors}
\title{Build an INLA formula for mbg}
\usage{
build_mbg_formula_with_priors(fixed_effects,
  positive_constrained_variables = NULL, interact_with_year = NULL,
  int = TRUE, nullmodel = FALSE, add_nugget = FALSE,
  nugget_prior = "list(prior = 'loggamma', param = c(2, 1))",
  add_ctry_res = FALSE,
  ctry_re_prior = "list(prior = 'loggamma', param = c(2, 1))",
  temporal_model_type = "'ar1'",
  temporal_model_theta_prior = "list(prior = 'loggamma', param = c(1, 0.00005))",
  temporal_model_theta1_prior = "list(prior = 'normal', param = c(0, 1/(2.58^2)))",
  no_gp = FALSE, stacker_names = child_model_names,
  coefs.sum1 = FALSE, subnat_RE = FALSE,
  subnat_re_prior = "list(prior = 'loggamma', param = c(1, 5e-5))",
  use_space_only_gp = FALSE, use_time_only_gmrf = FALSE,
  time_only_gmrf_type = "rw2")
}
\arguments{
\item{fixed_effects}{PARAM_DESCRIPTION}

\item{positive_constrained_variables}{PARAM_DESCRIPTION, Default: NULL}

\item{interact_with_year}{PARAM_DESCRIPTION, Default: NULL}

\item{int}{PARAM_DESCRIPTION, Default: TRUE}

\item{nullmodel}{PARAM_DESCRIPTION, Default: FALSE}

\item{add_nugget}{PARAM_DESCRIPTION, Default: FALSE}

\item{nugget_prior}{PARAM_DESCRIPTION, Default: 'list(prior = 'loggamma', param = c(2, 1))'}

\item{add_ctry_res}{PARAM_DESCRIPTION, Default: FALSE}

\item{ctry_re_prior}{PARAM_DESCRIPTION, Default: 'list(prior = 'loggamma', param = c(2, 1))'}

\item{temporal_model_type}{PARAM_DESCRIPTION, Default: ''ar1''}

\item{temporal_model_theta_prior}{PARAM_DESCRIPTION, Default: 'list(prior = 'loggamma', param = c(1, 0.00005))'}

\item{temporal_model_theta1_prior}{PARAM_DESCRIPTION, Default: 'list(prior = 'normal', param = c(0, 1/(2.58^2)))'}

\item{no_gp}{PARAM_DESCRIPTION, Default: FALSE}

\item{stacker_names}{PARAM_DESCRIPTION, Default: child_model_names}

\item{coefs.sum1}{PARAM_DESCRIPTION, Default: FALSE}

\item{subnat_RE}{Additional admin-1 random effect. Default: FALSE}

\item{use_space_only_gp}{Logical. include a space only (time stationary) gp. Default: FALSE}

\item{use_time_only_gmrf}{Logical. include a time only gp. Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
construct a formula object for use with R-INLA::inla() model fitting in conjunction with the prepped data object from build_mbg_data_stack()
}
\details{
DETAILS
}
\note{
The \code{formula} object was created at the very end of the function after dealing with everything before that in strings.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[base]{strsplit}}
}
