% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_raking_link_table.R
\name{build_raking_link_table}
\alias{build_raking_link_table}
\title{Aggregates link_table pixel ownership for geographies at distinct admin levels.}
\usage{
build_raking_link_table(shapefile_version = "current",
  force_adm0 = TRUE, countries_not_to_subnat_rake = NULL)
}
\arguments{
\item{shapefile_version}{Shapefile version. Default: "current"}

\item{force_adm0}{Force link table aggregation to ADM0? This produces output consistent
with \code{build_simple_raster_pop}. It is necessary to ensure pixels are consistently
assigned between different rasters. Set this to FALSE with caution. Default: TRUE}

\item{countries_not_to_subnat_rake}{as it sounds. Replaces subnational metadata with national metadata. Default: NULL}
}
\value{
a link table with names "pixel_id", "area_fraction", and "loc_id".
}
\description{
Normally aggregation of link table values from e.g., admin 2 to
admin 0 is straight-forward. This function handles aggregating link table values
to heterogeneous administrative levels by providing a information about which
geographies need to be aggregated to a particular level. In addition, it also
translates from Local Burden of Disease administrative codes to GBD location_ids.
Finally, if desired, because this is only used for creating the raking raster
at ADM0 level, we sub out all ADM1 and ADM2 values to use their parent
location ID by mapping with the GBD location hierarchy
}
