% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_simple_raster_pop.R
\name{build_simple_raster_pop}
\alias{build_simple_raster_pop}
\title{FUNCTION_TITLE}
\usage{
build_simple_raster_pop(subset_shape, u5m = FALSE, field = NULL,
  raking = F, link_table = modeling_shapefile_version)
}
\arguments{
\item{subset_shape}{PARAM_DESCRIPTION}

\item{u5m}{PARAM_DESCRIPTION, Default: FALSE}

\item{field}{PARAM_DESCRIPTION, Default: NULL}

\item{raking}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{mask}}
}
