% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_space_mesh.R
\name{build_space_mesh}
\alias{build_space_mesh}
\title{Buils spatial FEM mesh}
\usage{
build_space_mesh(d, simple, max_edge, mesh_offset, plot_mesh = F,
  s2mesh = FALSE, s2params = NULL)
}
\arguments{
\item{d}{data.frame or data.table of observational data with
'longitude' and 'latitude' columns}

\item{simple}{simple_polygon defining modeling domain}

\item{max_edge}{string of 2 element numeric vector in R
notation. e.g. "c(0.25, 5)". First entry decribes the maximum
allowed triangle edge length within the simple_polygon modeling
boundary. Second entry describes the maximum allowed triaqngle
edge length in the extended buffer region. Units are in
lat-long distances. Used if s2mesh=FALSE.}

\item{mesh_offset}{string of 2 element numeric vector in R
notation. e.g. "c(1, 5)". Describes the automatic extension
distance from the 'simple' boundary. The entries are the inner
and outer extension distances, respectively. Units are in
lat-long distance. Used if s2mesh=FALSE.}

\item{plot_mesh}{Logical. Should a plot of the mesh be generated?
Not currently implemented for s2mesh}

\item{s2mesh}{Logical. Should the mesh be created on the surface of
a sphere? If TRUE, s2params is used to specify mesh parameters
instead of max_edge and mesh_offset}

\item{s2params}{string of 3 element numeric vector in R
notation. e.g. "c(25, 500, 1000)". The entries describe the
minimum triangle edge length allowed, hos far to extend the mesh
beyond the 'simple' boundary, and the maximum allowed triangle
edge length, respectively. Units are in kilometers. Used only if
s2mesh=TRUE.}
}
\value{
an 'inla.mesh' object
}
\description{
Build a finite-elements triangulation mesh to use in
  SPDE GP approximation
}
