% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_time_mesh.R
\name{build_time_mesh}
\alias{build_time_mesh}
\title{FUNCTION_TITLE}
\usage{
build_time_mesh(periods = 1:4)
}
\arguments{
\item{periods}{PARAM_DESCRIPTION, Default: 1:4}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
