% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fractional_rfs.R
\name{calculate_fractional_rfs}
\alias{calculate_fractional_rfs}
\title{Calculate fractional raking factor}
\usage{
calculate_fractional_rfs(ndraws = ndraws, cell_pred = cell_pred,
  gbd = gbd, sharedir = sharedir, run_date = run_date,
  indicator = indicator, shapefile_version = shapefile_version,
  stratum = stratum, countries_not_to_rake = NULL,
  custom_output_folder = NULL, countries_not_to_subnat_rake = NULL,
  MaxJump = 10, MaxIter = 80, FunTol = 1e-05, approx_0_1 = F,
  zero_heuristic = F, iterate = F, rake_method = "linear")
}
\arguments{
\item{ndraws}{the number of draws used in this model run, can be calculated from the number of columns in a raw cell_pred}

\item{cell_pred}{the preped and linked cell_pred that has also had its population raked.}

\item{gbd}{a data frame of the gbd targets that we are raking to}

\item{sharedir}{the share directory for the indicator group, used to save the raking factors}

\item{run_date}{your run date}

\item{indicator}{the modeled indicator}

\item{shapefile_version}{the shapefile version}

\item{stratum}{the region you are modeling over, used for saving the raking factors}

\item{countries_not_to_rake}{countries (vector or summed string) to not rake to GBD (we set rake factor to 1 for those)}

\item{custom_output_folder}{Output the rake factors to custom folder path if specified. Default: NULL}

\item{countries_not_to_subnat_rake}{countries (vector or summed string) to not rake to subnationals (we set rake factor to 1 for those)}

\item{MaxJump}{default `10`. Maximum size of a jump to the answer (for logit raking) - this will likely never need to be changed}

\item{MaxIter}{default `80`. Number of jumps towards the solution (for logit raking) - this will likely never need to be changed}

\item{FunTol}{default `1e-5`. Maximum allowed difference between the raking target and raked results (for logit raking) - this will likely never need to be changed}

\item{approx_0_1}{default `F`. If logit raking, any values of zero will be replaced with 1e-10 and values of 1 will be replaced with (1-(1e-10)).  Otherwise, logit transformations will fail in `NewFindK`. Useful if some areas have very low or high predicted values in `cell_pred`,
such that some draws are either 0 or 1}

\item{zero_heuristic}{default `F`.  If logit raking, this will automatically set rf = -9999 for any country-year with a target value of 0.  This produces a raked value of 0 for all pixels.  Raking to a target of zero in logit space is very time-consuming and the algorithm
can only approach zero asymptotically.  For situations where the target is truly zero (most useful for, say, an intervention pre-introduction) this will both speed up the process and ensure that zeros are returned.}

\item{iterate}{default `F`. If logit raking for a location-year fails, try again with `MaxJump` and `MaxIter` times 10. If that fails, try again times 100. For circumstances where raking target is very far from estimate and raking does not converge.}

\item{rake_method}{if set to "logit" creates raking factors in logit space, otherwise assumes linear raking}
}
\value{
a table of raking factors for each of the raking geographies used.  This is so that
the fractionally aggregated mbg estimates rasters for a given geography year will equal
the GBD estimates for that geography in that year.  This also assumes that the GBD is internally consistent
in converting rates to counts, namely rate*pop=count
}
\description{
This calculates the raking factors, it takes a rates cell_pred which is the main difference between it and the
get_fractional_rf function above (see Michael Collison's work).
}
