% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pop_scalars.R
\name{calculate_pop_scalars}
\alias{calculate_pop_scalars}
\title{Calculate population scalars}
\usage{
calculate_pop_scalars(cell_pred = cell_pred, age_group = age_group,
  connector = connector, sex = sex_id, sharedir = sharedir,
  run_date = run_date, indicator = indicator, stratum = stratum,
  gbd_pops = gbd_pops, custom_output_folder = NULL)
}
\arguments{
\item{cell_pred}{the cell pred object that has been prepped and linked}

\item{age_group}{the GBD age_group you are modeling}

\item{connector}{the list of raking geographies}

\item{sex}{the GBD sex_id you are modeling}

\item{sharedir}{the share directory for the indicator group, used to save the population raking factors}

\item{run_date}{your run date}

\item{indicator}{the modeled indicator}

\item{stratum}{the region you are modeling over, used for saving the population raking factors}

\item{gbd_pops}{output from central code "get_population" function, passed from main function}

\item{custom_output_folder}{Output the rake factors to custom folder path if specified. Default: NULL}
}
\value{
a table of population raking facotrs for each of the raking geographies used.  This is so that
the fractionally aggregated worldpop rasters for a given age and sex group in a geography year will equal
the GBD population for that age and sex group in that geography in that year.
}
\description{
This calculates the population raking factors needed so that prevalence rates and counts will align with GBD
}
