% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_raking_factors.R
\name{calculate_raking_factors}
\alias{calculate_raking_factors}
\title{Calculate raking factors}
\usage{
a = calculate_raking_factors(cell_pred, aaa[[1]], year_list = 2000:2015, simple_raster = simple_raster, weight_brick = pop_raster_annual, rake_method = 'linear')
}
\arguments{
\item{cell_pred}{matrix. draws from an mbg model in the cell pred form}

\item{rake_targets}{data.table with at least 3 columns (specified by lyv) with gaul code, year and val to rake towards}

\item{lyv}{character vector of length 3. Denotes the gaul_code column, year column and value column-- in that order}

\item{year_list}{numeric vector. Vector of years implied by the cell pred.}

\item{simple_raster}{raster. Should be the rasterized version of a shapefile or otherwise denoting the gaul/admin codes specified in rake targets}

\item{weight_brick}{rasterbrick. Raster brick where the values are the spatial weightings (usually population).}

\item{rake_method}{character. One of 'linear' or 'logit'. Possibly more in the future}

\item{if_no_gbd}{default `return_na`, other option `return_unraked`. If return_na, any location-years without gbd estimates will return NA raking factors. if return_unraked, will any location-years without gbd estimates will return 1 for linear raking, 0 for logit raking.}

\item{MaxJump}{default `10`. Maximum size of a jump to the answer for logit raking.}

\item{MaxIter}{default `80`. Number of jumps towards the solution}

\item{FunTol}{default `1e-5`. Maximum allowed difference between the raking target and raked results}

\item{iterate}{default `F`. If logit raking for a location-year fails, try again with `MaxJump` and `MaxIter` times 10. If that fails, try again times 100. For circumstances where raking target is very far from estimate and raking does not converge.}

\item{zero_heuristic}{default `F`.  If logit raking, this will automatically set rf = -9999 for any country-year with a target value of 0.  This produces a raked value of 0 for all pixels.  Raking to a target of zero in logit space is very time-consuming and the algorithm
can only approach zero asymptotically.  For situations where the target is truly zero (most useful for, say, an intervention pre-introduction) this will both speed up the process and ensure that zeros are returned.}

\item{approx_0_1}{default `F`. If logit raking, any values of zero will be replaced with 1e-10 and values of 1 will be replaced with (1-(1e-10)).  Otherwise, logit transformations will fail in `NewFindK`. Useful if some areas have very low or high predicted values in `cell_pred`,
such that some draws are either 0 or 1}
}
\value{
a data.table with raking factors
}
\description{
A function to logit rake MBG results
}
