% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_quadtree_leaf.R
\name{cell.quadtree.leaf}
\alias{cell.quadtree.leaf}
\title{FUNCTION_TITLE}
\usage{
cell.quadtree.leaf(q, xylim)
}
\arguments{
\item{q}{PARAM_DESCRIPTION}

\item{xylim}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
