% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_pred_to_raster_brick.R
\name{cell_pred_to_raster_brick}
\alias{cell_pred_to_raster_brick}
\title{Convert cell pred into raster brick}
\usage{
cell_pred_to_raster_brick(cell_pred, simple_raster, year_list, func)
}
\arguments{
\item{cell_pred}{a cell pred object in matrix format}

\item{simple_raster}{the simple raster corresponding to the cell_pred_object}

\item{year_list}{list of years in cell_pred}

\item{func}{function to summarize cell pred. `mean`, `upper`, and `lower` are acceptable}
}
\value{
a rasterBrick with a layer for each year in the cell_pred
}
\description{
Insert cell pred object into a series of raster layers for each year
then combine into rasterBrick
}
