% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centreScale.R
\name{centreScale}
\alias{centreScale}
\title{FUNCTION_TITLE}
\usage{
centreScale(x, df, inverse = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION}

\item{inverse}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
