% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb.R
\name{cfb}
\alias{cfb}
\title{FUNCTION_TITLE}
\usage{
cfb(v)
}
\arguments{
\item{v}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
