% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_custom_regions.R
\name{check_custom_regions}
\alias{check_custom_regions}
\title{Check custom regions}
\usage{
check_custom_regions(region_list, shapefile_version = "current")
}
\arguments{
\item{region_list}{vector of region names (defined in `pull_custom_modeling_regions()`)}
}
\value{
list of two tables
        [[1]] gaul_codes in custom, but not default regions,
          [[2]] gaul_codes in default, but not custom regions, and
      [[3]] duplicated gaul_codes
}
\description{
Check custom region list against base MBG list and see what's added
or missing in comparison
}
\examples{
\dontrun{
region_list <- c(
  "vax_soas", "vax_seas", "vax_eaas", "vax_caeu",
  "vax_crbn", "vax_ctam", "vax_ansa", "vax_trsa",
  "vax_name", "vax_cssa", "vax_essa", "vax_sssa", "vax_wssa"
)
check_list <- check_custom_regions(region_list)
}
}
