% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_cov_issues.R
\name{check_for_cov_issues}
\alias{check_for_cov_issues}
\title{Checks for covariate issues}
\usage{
check_for_cov_issues(cc = cs_covs, afe = all_fixed_effects,
  afeb = all_fixed_effects_brt, fe = fixed_effects, tc = the_covs,
  acl = all_cov_layers, check_uniform = TRUE,
  check_pixelcount = TRUE, check_pixelcount_thresh = 0.95)
}
\arguments{
\item{cc}{object to change: cs_covs}

\item{afe}{object to change: all_fixed_effects}

\item{afeb}{object to change: all_fixed_effects_brt}

\item{tc}{object to change: the_covs}

\item{acl}{object to change: all_cov_layers}

\item{check_uniform}{boolean, do a check for uniform covariates, defaults to TRUE}

\item{check_pixelcount}{boolean, do a check for covariates with too few pixel
(ie too low of geographic coverage in the region), defaults to TRUE}

\item{check_pixelcount_thresh}{for a pixelcount check, what proportion of the maximum observed
pixel coverage is needed to keep the covariate in? must be between 0 and 1. defaults to 0.95}
}
\value{
the necessary objects as a named list that will later get assigned
to the environment within the parallel_model.R script
}
\description{
This function looks for issues with covariates that will cause issues down the line and tries to fix them.
The two issues looked for here are uniform covariates and pixel coverage. Uniform Covariates: If one of the extracted variables
does not vary over the data area. Pixel Coverage: If one of the covariates is missing in large parts of your model region,
these areas will be NA in your results later on. This often happens if you are using a modeled covariate in a new or
partially new region which was not previously modelled. This function takes in all parallel model objects that would need to
change if a covariate was removed for the code below in the parallel mode to work.
}
