% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_holdouts.R
\name{check_for_holdouts}
\alias{check_for_holdouts}
\title{FUNCTION_TITLE}
\usage{
check_for_holdouts(dir)
}
\arguments{
\item{dir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
