% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_lock.R
\name{check_for_lock}
\alias{check_for_lock}
\title{FUNCTION_TITLE}
\usage{
check_for_lock(shp,
  fast_shapefile_dir = "/share/geospatial/rds_shapefiles/")
}
\arguments{
\item{shp}{PARAM_DESCRIPTION}

\item{fast_shapefile_dir}{PARAM_DESCRIPTION, Default: '/share/geospatial/rds_shapefiles/'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
