% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_in_spoly.R
\name{check_if_in_spoly}
\alias{check_if_in_spoly}
\title{FUNCTION_TITLE}
\usage{
check_if_in_spoly(the_poly, compare_to,
  the_proj = projection(master_shape))
}
\arguments{
\item{the_poly}{PARAM_DESCRIPTION}

\item{compare_to}{PARAM_DESCRIPTION}

\item{the_proj}{PARAM_DESCRIPTION, Default: projection(master_shape)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[rgeos]{gIntersection}}
 \code{\link[raster]{area}}
}
