% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_nonzero_exit.R
\name{check_if_nonzero_exit}
\alias{check_if_nonzero_exit}
\title{FUNCTION_TITLE}
\usage{
check_if_nonzero_exit(lv, notification)
}
\arguments{
\item{lv}{PARAM_DESCRIPTION}

\item{notification}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
