% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_hierarchy_fix.R
\name{china_hierarchy_fix}
\alias{china_hierarchy_fix}
\title{FUNCTION_TITLE}
\usage{
china_hierarchy_fix(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
