% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono_folds.R
\name{chrono_folds}
\alias{chrono_folds}
\title{FUNCTION_TITLE}
\usage{
chrono_folds(yr, ss = NULL, n_folds = NULL, ...)
}
\arguments{
\item{yr}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION, Default: NULL}

\item{n_folds}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
