% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_after_postest.R
\name{clean_after_postest}
\alias{clean_after_postest}
\title{FUNCTION_TITLE}
\usage{
clean_after_postest(indicator, indicator_group, run_date, strata,
  delete_region_rasters = F)
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{strata}{PARAM_DESCRIPTION}

\item{delete_region_rasters}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
