% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_results_table.R
\name{clean_model_results_table}
\alias{clean_model_results_table}
\title{FUNCTION_TITLE}
\usage{
clean_model_results_table(rd = run_date, regs = Regions, ages = 0,
  nm = "", indic = indicator, ig = indicator_group,
  stackers = stacked_fixed_effects,
  coefs.sum1 = as.logical(coefs_sum1))
}
\arguments{
\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{regs}{PARAM_DESCRIPTION, Default: Regions}

\item{ages}{PARAM_DESCRIPTION, Default: 0}

\item{nm}{PARAM_DESCRIPTION, Default: ''}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}

\item{ig}{PARAM_DESCRIPTION, Default: indicator_group}

\item{stackers}{PARAM_DESCRIPTION, Default: stacked_fixed_effects}

\item{coefs.sum1}{PARAM_DESCRIPTION, Default: as.logical(coefs_sum1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\note{
Adapted by Jon Mosser from Roy Burstein's `clean_model_results()` function
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_match}}
}
