% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_path.R
\name{clean_path}
\alias{clean_path}
\title{Clean up path}
\usage{
clean_path(messy_path)
}
\arguments{
\item{messy_path}{a file path (character vector). This assumes that the path
is constructed with one or more '/' as a separator and not '\'}
}
\value{
a 'clean' path with a single '/' as a separator and no trailing '/'
}
\description{
Sometimes extra '/''s are added to file paths here and there when they are
constructed which makes it difficult to compare a file path (string) to
another. This function will clean up an existing file path removing
additional '/''s to make it possbile to compare them reliably
}
