% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_inla_scratch.R
\name{cleanup_inla_scratch}
\alias{cleanup_inla_scratch}
\title{FUNCTION_TITLE}
\usage{
cleanup_inla_scratch(run_date)
}
\arguments{
\item{run_date}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
