% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanv.R
\name{cleanv}
\alias{cleanv}
\title{FUNCTION_TITLE}
\usage{
cleanv(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
