% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip1.R
\name{clip1}
\alias{clip1}
\title{FUNCTION_TITLE}
\usage{
clip1(xylim.clip, i, upper)
}
\arguments{
\item{xylim.clip}{PARAM_DESCRIPTION}

\item{i}{PARAM_DESCRIPTION}

\item{upper}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
