% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip2.R
\name{clip2}
\alias{clip2}
\title{FUNCTION_TITLE}
\usage{
clip2(xylim_clip, i, upper)
}
\arguments{
\item{xylim_clip}{PARAM_DESCRIPTION}

\item{i}{PARAM_DESCRIPTION}

\item{upper}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
