% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_by.R
\name{collapse_by}
\alias{collapse_by}
\title{FUNCTION_TITLE}
\usage{
collapse_by(df, var, by_vars, ...)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}

\item{by_vars}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
