% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_list.R
\name{collapse_list}
\alias{collapse_list}
\title{FUNCTION_TITLE}
\usage{
collapse_list(df, vars)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{vars}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
