% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_aggregation.R
\name{combine_aggregation}
\alias{combine_aggregation}
\title{Combine aggregation}
\usage{
combine_aggregation(rd = run_date, indic = indicator,
  ig = indicator_group, ages, regions, holdouts, raked,
  dir_to_search = NULL, delete_region_files = T,
  merge_hierarchy_list = F, check_for_dupes = F)
}
\arguments{
\item{run_date, }{indicator, indicator_group for this run}

\item{ages:}{single value or vector of ages}

\item{regions:}{vector of regions used}

\item{holdouts:}{vector of holdouts used, e.g. just 0 or c(1,2,3,4,5,0)}

\item{raked:}{vector of raked values, e.g. just T, just F, or c(T,F)}

\item{dir_to_search:}{which directory to search in (defaults to share directory)}

\item{delete_region_files:}{logical. Should we delete the region-specific intermediate files?}

\item{merge_hierarchy_list:}{logical. Do you want to merge the sp_hierarchy_list onto your admin tables?}
}
\value{
rdata files for each combo of age/holdout/raked
  each with admin_0, admin_1, admin_2 data table objects & the sp_hierarchy_list object
  that maps them to names of admin units
}
\description{
Combine aggregation objects across region
}
