% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_region_image_history.R
\name{combine_region_image_history}
\alias{combine_region_image_history}
\title{FUNCTION_TITLE}
\usage{
combine_region_image_history(indicator, indicator_group, run_date,
  fixed_effects)
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{fixed_effects}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{merge}}
}
