% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_admin_estimates.R
\name{compare_admin_estimates}
\alias{compare_admin_estimates}
\title{FUNCTION_TITLE}
\usage{
compare_admin_estimates(nid_list, results_raster, compare_source_title,
  raked_title, outcome_title, compare_data, compare_spdf, master_list,
  shapefile_version, by_source = FALSE)
}
\arguments{
\item{nid_list}{PARAM_DESCRIPTION}

\item{results_raster}{PARAM_DESCRIPTION}

\item{compare_source_title}{PARAM_DESCRIPTION}

\item{raked_title}{PARAM_DESCRIPTION}

\item{outcome_title}{PARAM_DESCRIPTION}

\item{compare_data}{PARAM_DESCRIPTION}

\item{compare_spdf}{PARAM_DESCRIPTION}

\item{master_list}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION}

\item{by_source}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
