% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_all_admin_target.R
\name{compare_all_admin_target}
\alias{compare_all_admin_target}
\title{Runs comparison of projected years against a goal that is meant to be
applied to all admins within a given country.}
\usage{
compare_all_admin_target(ind_gp, ind, rd, measure, target_year, target,
  target_type, admin_level, uselogit, proj,
  shapefile_version = "current")
}
\arguments{
\item{ind_gp}{indicator group}

\item{ind}{indicator}

\item{rd}{run_date}

\item{measure}{prevalence, incidence, mortality, etc}

\item{target_year}{Year for which the goal is meant. Can be either a future (projected) year
with the year projections  already  generated for this target_year using \code{make_proj()},
or can be a modeled year (depending on the \code{proj} parameter below.}

\item{target}{Target (e.g. 0.8 for 80\% vaccine coverage); numeric}

\item{target_type}{Type of target ('greater' or 'less'); character}

\item{admin_level}{What admin level should the target be evaluated at? For instance, use
\code{1} if the goal applies to all admin1s within a given country, or \code{2} if
all the goal applies to all admin2s within a given country}

\item{uselogit}{Should this be done in logit space?}

\item{proj}{Is \code{target_year} projected (\code{proj = T}) or modeled (\code{proj = F})?}

\item{shapefile_version}{character string indicating version of shapefile to pull}
}
\value{
A data table by country with probability of meeting the specified goal
}
\description{
This is somewhat of a special use scenario but does have some specific applications.
For instance, the GVAP goal for geographic parity in vaccine coverage is "All
districts with coverage > 80\% within a country by 2020.  The proper way to
calculate the probability of achieving this goal is to calculate by draw
whether the country achieved the goal in the given draw, then
see in how many draws the country achieved the goal.  That is what this
function does.
}
\examples{
\dontrun{
# Compare against all-admin2 > 80\% target for GVAP
compare_all_admin_target(
  ind_gp = indicator_group,
  ind = indicator,
  rd = run_date,
  measure = "prevalence",
  target_year = 2020,
  target = 0.8,
  target_type = "greater",
  admin_level = 2,
  uselogit = T,
  proj = T
)
}
}
