% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_target.R
\name{compare_to_target}
\alias{compare_to_target}
\title{Compare to target}
\usage{
compare_to_target(ind_gp, ind, rd, goal_obj, measure,
  year_list = c(2000:2015), uselogit, raked, skip_cols = NULL,
  matrix_pred_name = NULL, shapefile_version = "current")
}
\arguments{
\item{ind_gp}{indicator group}

\item{ind}{indicator}

\item{rd}{run_date}

\item{goal_obj}{An existing goal object made with \code{add_goal()}}

\item{measure}{prevalence, incidence, mortality, etc}

\item{year_list}{Vector (integer) of years included in the model run / cell pred object}

\item{uselogit}{Should this be done in logit space?}

\item{skip_cols}{columns to skip when reading in the cell preds
For example, if the first two columns store non-pred information in your
file format, \code{skip_cols = 2} will read in all columns from 3 onwards}

\item{matrix_pred_name}{In \code{sprintf} notation. The one object passed into
the string should will be a region name. this allows different regions to be
passed to different named matrix_preds (pixel level, ad0, ad1, ad2, ...)
e.g. 'had_diarrhea_cell_draws_eb_bin0_\%s_diarrhea2_0.RData' which
will be passed to sprintf('had_diarrhea_cell_draws_eb_bin0_\%s_0.RData', reg)}

\item{shapefile_version}{character string indicating version of shapefile to pull}
}
\value{
generates cell- and admin- level probabilities of meeting the specified goal,
  according to what is specified in \code{goal_obj}.  Objects are written to
  standard directories and formats in the 'pred_derivatives' folder of the model run.
  cell-level objects are in the cell_pred indexed format, but are no longer
  wide by draw (just a single column as they are probabilities).  admin objects are
  saved both as rds files and as .csv files with admin hierarchy appended.
}
\description{
Runs comparison of projected years against a set of goals defined in a goal object
}
