% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_polygons.R
\name{compile_polygons}
\alias{compile_polygons}
\title{FUNCTION_TITLE}
\usage{
compile_polygons(this_nid, estimates_raster)
}
\arguments{
\item{this_nid}{PARAM_DESCRIPTION}

\item{estimates_raster}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
