% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_results_table.R
\name{compile_results_table}
\alias{compile_results_table}
\title{FUNCTION_TITLE}
\usage{
compile_results_table(indicator_group, indicator, run_date,
  measure = "mortality", baseline_year = 2000, year_to_map = 2015,
  goal_threshold = 0.001, metric = "sdgprob",
  shapefile_version = "current")
}
\arguments{
\item{indicator_group}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{measure}{PARAM_DESCRIPTION, Default: 'mortality'}

\item{baseline_year}{PARAM_DESCRIPTION, Default: 2000}

\item{year_to_map}{PARAM_DESCRIPTION, Default: 2015}

\item{goal_threshold}{PARAM_DESCRIPTION, Default: 0.001}

\item{metric}{PARAM_DESCRIPTION, Default: 'sdgprob'}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{merge}}
}
