% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_runtime_string.R
\name{complete_runtime_string}
\alias{complete_runtime_string}
\title{Complete user-supplied runtime string into DHMS format}
\usage{
complete_runtime_string(run_time)
}
\arguments{
\item{run_time}{String. User supplied run-time in format \code{[DD]:[HH]:[MM]:[SS]}}
}
\value{
The same run-time supplied by user once validations have passed,
but with days appended if not supplied, in format \code{DD:HH:MM:SS}
}
\description{
Check the input string of runtime and convert its values into a DD:HH:MM:SS format
}
