% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSim.R
\name{condSim}
\alias{condSim}
\title{FUNCTION_TITLE}
\usage{
condSim(vals, weights = NULL, group = NULL, fun = NULL, ...)
}
\arguments{
\item{vals}{PARAM_DESCRIPTION}

\item{weights}{PARAM_DESCRIPTION, Default: NULL}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{fun}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
