% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config_description}
\alias{config_description}
\title{Configuration data.table: Description}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 95 rows and 3 columns.}
\usage{
config_description
}
\description{
A data.table with default config values and description for MBG
}
\examples{
\dontrun{
data(config_description)
}
}
\keyword{datasets}
