% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config_tests}
\alias{config_tests}
\title{Tests for config files}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 96 rows and 4 columns.}
\usage{
config_tests
}
\description{
A list of tests for checking types (and some values) for all the MBG configuration variables
}
\examples{
\dontrun{
data(config_tests)
}
}
\keyword{datasets}
