% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config_values}
\alias{config_values}
\title{Configuration data.table}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 97 columns.}
\usage{
config_values
}
\description{
A data.table with default config values for MBG
}
\examples{
\dontrun{
data(config_values)
}
}
\keyword{datasets}
