% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_iso3.R
\name{convert_to_iso3}
\alias{convert_to_iso3}
\title{FUNCTION_TITLE}
\usage{
convert_to_iso3(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
