% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariatePathHelper.R
\docType{data}
\name{CovariatePathHelper}
\alias{CovariatePathHelper}
\title{Helper object for dealing with covariate paths}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CovariatePathHelper
}
\description{
CovariatePathHelper turns covariate/measure/release data into paths via \code{covariate_paths()}
 and converts it back via \code{covariate_metadata_from_path}. This is used internally by
 \code{MbgStandardCovariateLoader}.
}
\examples{
\dontrun{

covariates <- c("access", "evi")
measures <- c("mean", "median")
releases <- c("2019_06_10", "2019_06_10")
helper <- CovariatePathHelper$new()
paths <- helper$covariate_paths(covariates = covariates,
                                measures = measures,
                                releases = releases)
metadata <- helper$covariate_metadata_from_path(paths[1])
}
}
\keyword{datasets}
