% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_categorical_raster.R
\name{create_categorical_raster}
\alias{create_categorical_raster}
\title{FUNCTION_TITLE}
\usage{
create_categorical_raster(ras)
}
\arguments{
\item{ras}{a rasterlayer or brick}
}
\value{
one layer or brick per non-NA value in the input raster. Don't use with floats.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
