% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_set_mask.R
\name{crop_set_mask}
\alias{crop_set_mask}
\title{Crop and mark}
\usage{
crop_set_mask(raster_object, template_raster)
}
\arguments{
\item{raster_object}{A raster object (raster, Brick, etc) of a specific region}

\item{simple_raster}{A simple raster object that serves as the template for that region}
}
\value{
A raster/Brick/Stack, with the extents and masks of the simple raster.
}
\description{
Crop the raster to the extent of a simple_raster object,
set the extent, and then mask it based on the simple_raster object.
}
\author{
Rebecca Stubbs
}
