% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_cell_pred_add_NA.R
\name{crosswalk_cell_pred_add_NA}
\alias{crosswalk_cell_pred_add_NA}
\title{@title Crosswalk Cell Pred Object}
\usage{
crosswalk_cell_pred_add_NA(simple_raster, new_simple_raster, cell_pred,
  year_list)
}
\arguments{
\item{simple_raster}{Simple raster used to create the cell pred object (created by build_simple_raster_pop).}

\item{new_simple_raster}{Simple raster to be crosswalked to (created by build_simple_raster_pop).}

\item{cell_pred}{Cell pred object that matches with simple raster.}

\item{year_list}{List of years}
}
\value{
Returns crosswalked cell pred object
}
\description{
This version of the crosswalk adds NAs to the cell pred where there are pixels in the new simple
raster but not in the old. Reversible.
}
\note{
This function loops through each pixel in the rasters and compares them. pixels where both are NA or both have values are kept the same. Pixels where the original raster has a value and the new is na have the row in the cell pred object deleted. For pixels where the original is na and the new has a value, an NA row is added to the cell pred object.
}
\examples{
\dontrun{
new_pred_object <- crosswalk_cell_pred(simple_raster, new_simple_raster, cell_pred, years)
}
}
