% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_cell_pred_alter_raster.R
\name{crosswalk_cell_pred_alter_raster}
\alias{crosswalk_cell_pred_alter_raster}
\title{Crosswalk Cell Pred Object}
\usage{
crosswalk_cell_pred_alter_raster(simple_raster, new_simple_raster,
  cell_pred, year_list)
}
\arguments{
\item{simple_raster}{Simple raster used to create the cell pred object (created by build_simple_raster_pop).}

\item{new_simple_raster}{Simple raster to be crosswalked to (created by build_simple_raster_pop).}

\item{cell_pred}{Cell pred object that matches with simple raster.}

\item{year_list}{List of years}
}
\value{
Returns crosswalked cell pred object and altered new simple raster
}
\description{
This version of the crosswalk edits the new simple raster to match the cell pred by making any
pixel not in the original simple raster na. Nonreversible.
}
\note{
This function loops through each pixel in the rasters and compares them. pixels where both are NA or both have values are kept the same. Pixels where the original raster has a value and the new is na have the row in the cell pred object deleted. For pixels where the original is na and the new has a value, that pixel in the new raster is changed to an NA because there is no corresponding value in the cell pred object for that pixel.
}
\examples{
\dontrun{
new_pred_object <- crosswalk_cell_pred(simple_raster, new_simple_raster, cell_pred, years)
}
}
