% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_aggregate.R
\name{custom_aggregate}
\alias{custom_aggregate}
\title{Custom Aggregation Function}
\usage{
custom_aggregate(cell_pred, shapefile_path, field, reg, yl,
  pop_meas = pop_measure, ss = NULL, return_shapefile = T,
  verbose = T, modeling_shapefile_version = "current")
}
\arguments{
\item{cell_pred}{Cell pred object to be aggregated}

\item{field}{Field in shapefile that should be used for aggregation (ADM2_CODE, etc).}

\item{reg}{Region used to produce cell pred object.}

\item{yl}{List of years}

\item{ss}{Summary statistics}

\item{return_shapefile}{Boolean, should the function return the shapefile?}

\item{verbose}{Should the function be talkative?}

\item{modeling_shapefile_version}{string used to specify which shapefile version to use for merging adm codes to pixels. should be the version used to generate the cell_preds}

\item{custom_shapefile_path}{Path to shapefile that will be used for aggregation}

\item{custom_shapefile}{Alternatively can specify a shapefile object directly}
}
\value{
Returns a list containing:
         \code{draws}: data.table with rows for each aggregated area & year,
                  and columns representing [field], year, and draws[V1...Vn]
         \code{shapefile_data}: contents of the shapefile defined by \code{shapefile_path},
                           which can be merged on to "draws" to add on admin names, etc
         \code{summarized}: data table with row for each aggregated area & year,
                       and columns for [field], year, and all summary stats in \code{ss}
                       (if \code{ss}=NULL, then this is also NULL)
         \code{shapefile}: the shapefile itself (if return_shapefile = T; otherwise will be NULL)
}
\description{
Create custom aggregates
}
\examples{
\dontrun{
custom_agg_output <- custom_aggregate(
  cell_pred = cell_pred,
  shapefile_path = "/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory/ETH_adm3_2007.shp",
  field = "ID_3",
  reg = "essa",
  yl = c(2000:2016),
  pop_meas = pop_measure,
  ss = c("mean", "upper", "lower", "cirange"),
  verbose = T
)
}
}
