% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_rake.R
\name{custom_rake}
\alias{custom_rake}
\title{Custom Raking Function}
\usage{
custom_rake(cell_pred, shapefile_path, field, rake_to, reg, year_list,
  modeling_shapefile_version = "current")
}
\arguments{
\item{cell_pred}{Cell pred object to be raked.}

\item{shapefile_path}{Path to shapefile that will be used for
raking.}

\item{field}{Field in shapefile that has admin identifiers that
match with rake_to.}

\item{rake_to}{Df with name, year, and mean columns. values in name
must match up with values in field.}

\item{reg}{Region used to produce cell pred object.}

\item{year_list}{List of years}

\item{modeling_shapefile_version}{string indicating which version of
shapefile to use for matching cell_pred pixels to adm codes}
}
\value{
Returns a new cell pred object, simple raster, raking factors, pre-raking aggregate numbers, and rasters of mean, lower, upper, and cirange for years in year list
}
\description{
Builds a new simple polygon and raster, crosswalks the cell pred object to match the new raster. Generates population weights based on the shapefile and field passed in, calculates raking factors using the rake_to df
}
\examples{
\dontrun{
custom_rake_output <- custom_rake(
  cell_pred = cell_draws,
  shapefile_path = "/snfs1/WORK/11_geospatial/05_survey shapefile library/Shapefile directory/IND_for_collaborator.shp",
  field = "gbd_loc_id",
  rake_to = rake_to,
  reg = "south_asia",
  year_list = c(2000:2015)
)
}
}
