% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\title{FUNCTION_TITLE}
\usage{
db(machine = c("local"), user = "lina1864", warnings = FALSE)
}
\arguments{
\item{machine}{PARAM_DESCRIPTION, Default: c("local")}

\item{user}{PARAM_DESCRIPTION, Default: 'lina1864'}

\item{warnings}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
