% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupe_linked_cell_pred.R
\name{dedupe_linked_cell_pred}
\alias{dedupe_linked_cell_pred}
\title{Deduplicate linked cell preds}
\usage{
dedupe_linked_cell_pred(linked_cell_pred, overs)
}
\arguments{
\item{linked_cell_pred}{a cell pred merged to a link table generated by `get_link_table()`}

\item{overs}{a vector of the column names in the cell pred corresponding to draws
(typically V1:Vndraws)}
}
\value{
a cell pred object in original matrix form which corresponds to simple raster
}
\description{
Collapse rows of cell preds with multiple admin units in one pixel back into
single rows
}
