% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultOptions.R
\name{defaultOptions}
\alias{defaultOptions}
\title{FUNCTION_TITLE}
\usage{
defaultOptions(resolution = 5, location = loc(),
  cores = switch(loc(), oxford = 60, seattle = 60), spacetime = TRUE,
  start = Sys.time())
}
\arguments{
\item{resolution}{PARAM_DESCRIPTION, Default: 5}

\item{location}{PARAM_DESCRIPTION, Default: loc()}

\item{cores}{PARAM_DESCRIPTION, Default: switch(loc(), oxford = 60, seattle = 60)}

\item{spacetime}{PARAM_DESCRIPTION, Default: TRUE}

\item{start}{PARAM_DESCRIPTION, Default: Sys.time()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
