% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_model_outputs.R
\name{delete_model_outputs}
\alias{delete_model_outputs}
\title{Deletes model outputs from share directories}
\usage{
delete_model_outputs(indicator_group, indicator, run_date,
  dryrun = FALSE, ...)
}
\arguments{
\item{indicator_group}{The indicator group level example "u5m"}

\item{indicator}{The indicator level example "died"}

\item{run_date}{The name of your model but often its just a run date}

\item{dryrun}{Only print the directories that would be deleted dont actually}

\item{...}{Additional arguments passed to unlink function}
}
\value{
None
}
\description{
Quickly delete the model outputs of a number of model runs will create
 a unique combination for each indicator, indicator_group, run_date combination
}
