% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_adm_shapefile_date_type.R
\name{detect_adm_shapefile_date_type}
\alias{detect_adm_shapefile_date_type}
\title{Detect admin shapefile date and type}
\usage{

  detect_adm_shapefile_date_type(shpfile_path = get_admin_shapefile(version
  = modeling_shapefile_version))
}
\arguments{
\item{shpfile_path}{path to admin shapefile we want to learn about}
}
\value{
two element named list:
  1) list element 'shpfile_type' contains string: either 'gadm' or 'gaul'
  2) list element 'shpfile_date' contains the actual date of the
  shapefile, even if version='current' was used in shpfile_path
}
\description{
Detects whether the admin shapefile you are using is gaul or gadm
and returns the version/date of the shapefile even if the 'current'
version was specified
}
\examples{
\dontrun{
detect_shapefile_type(get_admin_shapefile(version = "current"))
}
}
