% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_demographics.R
\name{detect_demographics}
\alias{detect_demographics}
\title{FUNCTION_TITLE}
\usage{
detect_demographics(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
