% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhms_to_hours.R
\name{dhms_to_hours}
\alias{dhms_to_hours}
\title{Run-time to hours}
\usage{
dhms_to_hours(run_time)
}
\arguments{
\item{run_time}{String. Run-time in format \code{DD:HH:MM:SS}.}
}
\value{
number of hours
}
\description{
Convert a validated \code{DD:HH:MM:SS} runtime string to hours
}
