% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dircos.R
\name{dircos}
\alias{dircos}
\title{FUNCTION_TITLE}
\usage{
dircos(x1)
}
\arguments{
\item{x1}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
