% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupRecords.R
\name{dupRecords}
\alias{dupRecords}
\title{FUNCTION_TITLE}
\usage{
dupRecords(j)
}
\arguments{
\item{j}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
