% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elogit.R
\name{elogit}
\alias{elogit}
\title{FUNCTION_TITLE}
\usage{
elogit(y, n)
}
\arguments{
\item{y}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
