% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplogit.R
\name{emplogit}
\alias{emplogit}
\title{FUNCTION_TITLE}
\usage{
emplogit(success, N, epsilon = NULL)
}
\arguments{
\item{success}{PARAM_DESCRIPTION}

\item{N}{PARAM_DESCRIPTION}

\item{epsilon}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
