% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_world_raster.R
\name{empty_world_raster}
\alias{empty_world_raster}
\title{Create raster representing world}
\usage{
empty_world_raster(whole_world = FALSE)
}
\arguments{
\item{whole_world}{Logical if TRUE return raster for entire world. If FALSE,
return raster for stage1+stage2}
}
\description{
Creates a raster file consistent with LBD's world raster.
}
