% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvalDiff.R
\name{EvalDiff}
\alias{EvalDiff}
\title{FUNCTION_TITLE}
\usage{
EvalDiff(gbdval, vals, weightval)
}
\arguments{
\item{gbdval}{PARAM_DESCRIPTION}

\item{vals}{PARAM_DESCRIPTION}

\item{weightval}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
