% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvalK.R
\name{EvalK}
\alias{EvalK}
\title{FUNCTION_TITLE}
\usage{
EvalK(K, p_i, p_N, N_i, Mult = FALSE)
}
\arguments{
\item{K}{PARAM_DESCRIPTION}

\item{p_i}{PARAM_DESCRIPTION}

\item{p_N}{PARAM_DESCRIPTION}

\item{N_i}{PARAM_DESCRIPTION}

\item{Mult}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
