% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_covariates}
\alias{extract_covariates}
\title{FUNCTION_TITLE}
\usage{
extract_covariates(df, covariate_list, id_col = NULL,
  reconcile_timevarying = T, period_var = NULL,
  return_only_results = F, centre_scale = F, period_map = NULL)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{covariate_list}{PARAM_DESCRIPTION}

\item{id_col}{PARAM_DESCRIPTION, Default: NULL}

\item{reconcile_timevarying}{PARAM_DESCRIPTION, Default: T}

\item{period_var}{PARAM_DESCRIPTION, Default: NULL}

\item{return_only_results}{PARAM_DESCRIPTION, Default: F}

\item{centre_scale}{PARAM_DESCRIPTION, Default: F}

\item{period_map}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
