% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_brick.R
\name{extract_from_brick}
\alias{extract_from_brick}
\title{Extract from brick}
\usage{
extract_from_brick(data, yearcol = "original_year", rb,
  varname = "extracted_values", raster_years = 2000:2015)
}
\arguments{
\item{data}{A data.table or data.frame (that will be converted to a data.table) with
columns for latitude, longitude, and year.}

\item{yearcol}{A string name of a column that defines the year of the data within the data object}

\item{rb}{A rasterbrick with the bands that correspond to the raster_years.}

\item{varname}{A string name of what you want the column of extracted values to be}

\item{raster_years}{A vector that describes the years that serve as bands in the rasterbrick, in order.}
}
\value{
Returns the data object, sorted based on year, with a new column with the extracted values (name
as specified based on the varname parameter).
}
\description{
Extract values from a raster or rasterbrick to lat/lon points by relevant year of data.
}
\author{
Rebecca Stubbs
}
