% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raked_and_unraked.R
\name{extract_raked_and_unraked}
\alias{extract_raked_and_unraked}
\title{FUNCTION_TITLE}
\usage{
extract_raked_and_unraked(gaul, results, results_raked, master_list,
  admin0, load_simple_raster = FALSE, shapefile_version = "current")
}
\arguments{
\item{gaul}{PARAM_DESCRIPTION}

\item{results}{PARAM_DESCRIPTION}

\item{results_raked}{PARAM_DESCRIPTION}

\item{master_list}{PARAM_DESCRIPTION}

\item{admin0}{PARAM_DESCRIPTION}

\item{load_simple_raster}{PARAM_DESCRIPTION, Default: FALSE}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
