% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_region.R
\name{extract_region}
\alias{extract_region}
\title{FUNCTION_TITLE}
\usage{
extract_region(region)
}
\arguments{
\item{region}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
