% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_preds.R
\name{extract_year_preds}
\alias{extract_year_preds}
\title{FUNCTION_TITLE}
\usage{
extract_year_preds(country_year)
}
\arguments{
\item{country_year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
