% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_evparse.R
\name{ez_evparse}
\alias{ez_evparse}
\title{Easy eval-parse}
\usage{
ez_evparse(data, column)
}
\arguments{
\item{data}{The data.table}

\item{column}{The column with the string call}
}
\value{
Evaluated call
}
\description{
Allows for easily eval-parsing through a config dataset
}
