% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_load_shapefile.R
\name{fast_load_shapefile}
\alias{fast_load_shapefile}
\title{FUNCTION_TITLE}
\usage{
fast_load_shapefile(shape,
  fast_shapefile_dir = "/share/geospatial/rds_shapefiles/")
}
\arguments{
\item{shape}{PARAM_DESCRIPTION}

\item{fast_shapefile_dir}{PARAM_DESCRIPTION, Default: '/share/geospatial/rds_shapefiles/'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
