% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_covariate_layer.R
\name{fetch_covariate_layer}
\alias{fetch_covariate_layer}
\title{FUNCTION_TITLE}
\usage{
fetch_covariate_layer(ras, period = 1)
}
\arguments{
\item{ras}{PARAM_DESCRIPTION}

\item{period}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
