% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_rdata.R
\name{fetch_from_rdata}
\alias{fetch_from_rdata}
\title{fetch_from_rdata}
\usage{
fetch_from_rdata(file_location, item_name, use_grep = F)
}
\arguments{
\item{file_location}{where to look for the Rdata object}

\item{item_name}{the item to pull from the Rdata object}

\item{use_grep}{Weather or not to use grep to look for similar file names}
}
\value{
loads the specific item from an Rdata object
}
\description{
This pulls out pieces of an Rdata object
}
