% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_gbd_cov.R
\name{fetch_gbd_cov}
\alias{fetch_gbd_cov}
\title{Fetch GBD covariates}
\usage{
fetch_gbd_cov(name, measure, age_ids, afras, year_ids, shapefile_version)
}
\arguments{
\item{name}{Name of covariate}

\item{measure}{Measure of covariate}

\item{age_ids}{Age group IDs}

\item{afras}{PARAM_DESCRIPTION}

\item{year_ids}{Year IDs}

\item{shapefile_version}{Shapefile version}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get GBD covariates
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{getValues}}
}
