% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindK.R
\name{FindK}
\alias{FindK}
\title{FUNCTION_TITLE}
\usage{
FindK(p_i, p_N, N_i, a, Mult = FALSE)
}
\arguments{
\item{p_i}{PARAM_DESCRIPTION}

\item{p_N}{PARAM_DESCRIPTION}

\item{N_i}{PARAM_DESCRIPTION}

\item{a}{PARAM_DESCRIPTION}

\item{Mult}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
