% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_earth.R
\name{fit_earth}
\alias{fit_earth}
\title{FUNCTION_TITLE}
\usage{
fit_earth(df, covariates = all_fixed_effects, additional_terms = NULL,
  weight_column = NULL, indicator, indicator_family = "binomial")
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: all_fixed_effects}

\item{additional_terms}{PARAM_DESCRIPTION, Default: NULL}

\item{weight_column}{PARAM_DESCRIPTION, Default: NULL}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
