% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{FUNCTION_TITLE}
\usage{
fit_gam(df, covariates = all_fixed_effects, additional_terms = NULL,
  weight_column = NULL, bam = F, spline_args = list(),
  auto_model_select = F, indicator, indicator_family = "binomial",
  cores = "auto")
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: all_fixed_effects}

\item{additional_terms}{PARAM_DESCRIPTION, Default: NULL}

\item{weight_column}{PARAM_DESCRIPTION, Default: NULL}

\item{bam}{PARAM_DESCRIPTION, Default: F}

\item{spline_args}{PARAM_DESCRIPTION, Default: list()}

\item{auto_model_select}{PARAM_DESCRIPTION, Default: F}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}

\item{cores}{PARAM_DESCRIPTION, Default: 'auto'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[mgcv]{gam}},\code{\link[mgcv]{bam}}
}
