% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam_child_model.R
\name{fit_gam_child_model}
\alias{fit_gam_child_model}
\title{FUNCTION_TITLE}
\usage{
fit_gam_child_model(df, model_name = "gam", fold_id_col = "fold_id",
  covariates = all_fixed_effects, additional_terms = NULL,
  weight_column = NULL, bam = F, spline_args = list(bs = "ts", k =
  3), auto_model_select = T, indicator = indicator,
  indicator_family = "binomial", cores = "auto")
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{model_name}{PARAM_DESCRIPTION, Default: 'gam'}

\item{fold_id_col}{PARAM_DESCRIPTION, Default: 'fold_id'}

\item{covariates}{PARAM_DESCRIPTION, Default: all_fixed_effects}

\item{additional_terms}{PARAM_DESCRIPTION, Default: NULL}

\item{weight_column}{PARAM_DESCRIPTION, Default: NULL}

\item{bam}{PARAM_DESCRIPTION, Default: F}

\item{spline_args}{PARAM_DESCRIPTION, Default: list(bs = "ts", k = 3)}

\item{auto_model_select}{PARAM_DESCRIPTION, Default: T}

\item{indicator}{PARAM_DESCRIPTION, Default: indicator}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}

\item{cores}{PARAM_DESCRIPTION, Default: 'auto'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
