% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gbm.R
\name{fit_gbm}
\alias{fit_gbm}
\title{FUNCTION_TITLE}
\usage{
fit_gbm(df, covariates = all_fixed_effects, additional_terms = NULL,
  weight_column = NULL, tc = 4, lr = 0.005, bf = 0.75, indicator,
  indicator_family = "binomial", plot.main = F)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: all_fixed_effects}

\item{additional_terms}{PARAM_DESCRIPTION, Default: NULL}

\item{weight_column}{PARAM_DESCRIPTION, Default: NULL}

\item{tc}{PARAM_DESCRIPTION, Default: 4}

\item{lr}{PARAM_DESCRIPTION, Default: 0.005}

\item{bf}{PARAM_DESCRIPTION, Default: 0.75}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}

\item{plot.main}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
