% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glmnet.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{FUNCTION_TITLE}
\usage{
fit_glmnet(df, covariates = all_fixed_effects, additional_terms = NULL,
  weight_column = NULL, alpha = 1, indicator,
  indicator_family = "binomial", parallel = FALSE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: all_fixed_effects}

\item{additional_terms}{PARAM_DESCRIPTION, Default: NULL}

\item{weight_column}{PARAM_DESCRIPTION, Default: NULL}

\item{alpha}{PARAM_DESCRIPTION, Default: 1}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}

\item{parallel}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
