% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mbg.R
\name{fit_mbg}
\alias{fit_mbg}
\title{Fit MBG model in INLA}
\usage{
fit_mbg(indicator_family, stack.obs, spde, cov, N, int_prior_mn,
  int_prior_prec = 1, f_mbg, run_date, keep_inla_files, cores,
  verbose_output = FALSE, wgts = 0, intstrat = "eb",
  fe_mean_prior = 0, fe_sd_prior = 1, omp_strat = "huge",
  blas_cores = 1, pardiso_license = NULL)
}
\arguments{
\item{indicator_family}{PARAM_DESCRIPTION}

\item{stack.obs}{PARAM_DESCRIPTION}

\item{spde}{PARAM_DESCRIPTION}

\item{cov}{PARAM_DESCRIPTION}

\item{N}{PARAM_DESCRIPTION}

\item{int_prior_mn}{PARAM_DESCRIPTION}

\item{int_prior_prec}{PARAM_DESCRIPTION, Default: 1}

\item{f_mbg}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{keep_inla_files}{PARAM_DESCRIPTION}

\item{cores}{PARAM_DESCRIPTION}

\item{verbose_output}{PARAM_DESCRIPTION, Default: FALSE}

\item{wgts}{PARAM_DESCRIPTION, Default: 0}

\item{intstrat}{PARAM_DESCRIPTION, Default: 'eb'}

\item{fe_mean_prior}{PARAM_DESCRIPTION, Default: 0}

\item{fe_sd_prior}{PARAM_DESCRIPTION, Default: 1}

\item{blas_cores}{Number of BLAS threads to use. Default: 1.}

\item{pardiso_license}{Path to PARDISO license. Must be provided if using a Pardiso startegy in \code{omp_strat}. Default: NULL}

\item{omp_start}{OpenMP strategy to be used, one of \code{c('small', 'medium', 'large', 'huge', 'default', 'pardiso.serial', 'pardiso.parallel')}.
Default: \code{'huge'}, which is the INLA default for TAUCS solver.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Fit MBG model in INLA
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
