% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mbg_tmb.R
\name{fit_mbg_tmb}
\alias{fit_mbg_tmb}
\title{Fit a TMB MBG model}
\usage{
fit_mbg_tmb(lbdcorerepo = core_repo, cpp_template = "mbg_tmb_model",
  tmb_input_stack = input_data, ADmap_list = NULL,
  control_list = NULL, optimizer = "nlminb")
}
\arguments{
\item{lbdcorerepo}{core repo location}

\item{cpp_template}{name of cpp template file within ./<lbdcorerepo>/mbg_central/}

\item{tmb_input_stack}{object that results from build_mbg_data_stack_tmb() or build_mbg_data_stack(...,tmb=TRUE)}

\item{ADmap_list}{map parameter for ignoring parameters}

\item{control_list}{pass control list to nlminb()}

\item{optimizer}{which software to use for optimization (optim, or nlminb)}
}
\value{
list of tmb model objects: ADfun, opt, sdrep

# @useDynLib mbg_tmb_model
}
\description{
Fit a TMB MBG model and pull jointPrecision report
}
\note{
TODO support for other data models, sum to one constraint, country random effects
}
\author{
Roy Burstein
}
