% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rf.R
\name{fit_rf}
\alias{fit_rf}
\title{FUNCTION_TITLE}
\usage{
fit_rf(df, covariates = all_fixed_effects, additional_terms = NULL,
  ntree = 1000, indicator, indicator_family = "binomial")
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: all_fixed_effects}

\item{additional_terms}{PARAM_DESCRIPTION, Default: NULL}

\item{ntree}{PARAM_DESCRIPTION, Default: 1000}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
