% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{fit_stats}
\alias{fit_stats}
\title{FUNCTION_TITLE}
\usage{
fit_stats(is_data, oos_data, indicator, indicator_group, run_date,
  pathaddin)
}
\arguments{
\item{is_data}{PARAM_DESCRIPTION}

\item{oos_data}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{pathaddin}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
