% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats_ho_id.R
\name{fit_stats_ho_id}
\alias{fit_stats_ho_id}
\title{FUNCTION_TITLE}
\usage{
fit_stats_ho_id(draws, draw_prefix, observed)
}
\arguments{
\item{draws}{PARAM_DESCRIPTION}

\item{draw_prefix}{PARAM_DESCRIPTION}

\item{observed}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
