% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_param_table_tmb.R
\name{fitted_param_table_tmb}
\alias{fitted_param_table_tmb}
\title{Get fitted model parameters from TMB object}
\usage{
fitted_param_table_tmb(model_fit, exp_fixed_effects = TRUE,
  transform_hyperparams = TRUE, draws = 1000, calculate_range = TRUE,
  calculate_nom_var = TRUE)
}
\arguments{
\item{model_fit}{fitted TMB object that comes out of fit_mbg_tmb()}

\item{exp_fixed_effects}{Boolean, should the fixed effects be exponentiated (if model was fit with logit link). Defaults to TRUE}

\item{transform_hyperparams}{Boolean, should hyperparmeters be transformed from fitted to more natural space. Defaults to TRUE}

\item{draws}{Integer, number of draws to use. Defaults to 1000}

\item{calculate_range}{Boolean, should we calculate and report range using kappa in draws space. Defaults to TRUE}

\item{calculate_nom_var}{Boolean, should we calculate and report nominal variance using kappa and tau in draws space. Defaults to TRUE}
}
\value{
formatted data.table object
}
\description{
Make a nice table of fitted model parameters from a geostat TMB object
}
\author{
Roy Burstein
}
