% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_diacritics.R
\name{fix_diacritics}
\alias{fix_diacritics}
\title{Fixes diacritic encoding}
\usage{
fix_diacritics(x)
}
\arguments{
\item{x}{String or character vector to fix}
}
\value{
String or character vector with replaced diacritics
}
\description{
Fixes diacritic encoding to UTF-8
}
\examples{
\dontrun{
fix_diacritics(c("Côte", "Sør-Trøndela", "Löwenbräu"))
}

}
