% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_link.R
\name{fix_link}
\alias{fix_link}
\title{Fix area fractions in link table}
\usage{
fix_link(link)
}
\arguments{
\item{link}{link table}
}
\value{
a link table with adjusted area fractions
}
\description{
Scales area fractions for pixels with water. Pixels with 1 admin unit and
water have the area fraction set to 1, as all values will occur on the land. Pixels
containing multiple admin units and water are scaled proportionally. Internal function
used in \code{build_link_table()}
}
