% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_covariates.R
\name{format_covariates}
\alias{format_covariates}
\title{FUNCTION_TITLE}
\usage{
format_covariates(covariates = covariates)
}
\arguments{
\item{covariates}{PARAM_DESCRIPTION, Default: covariates}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
