% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_plot_obj.R, R/format_plot_obj_scatter.R
\name{format_plot_obj}
\alias{format_plot_obj}
\alias{format_plot_obj_scatter}
\title{FUNCTION_TITLE}
\usage{
format_plot_obj(plot_obj)

format_plot_obj_scatter(map, scatter)
}
\arguments{
\item{plot_obj}{PARAM_DESCRIPTION}

\item{map}{PARAM_DESCRIPTION}

\item{scatter}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
