% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_time.R
\name{format_time}
\alias{format_time}
\title{FUNCTION_TITLE}
\usage{
format_time(run_time)
}
\arguments{
\item{run_time}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
