% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional_agg_rates.R
\name{fractional_agg_rates}
\alias{fractional_agg_rates}
\title{Fractional aggregation of a rates cell pred}
\usage{
fractional_agg_rates(cell_pred = cell_pred,
  simple_raster = simple_raster, simple_polygon = simple_polygon,
  pixel_id = pixel_id, shapefile_version = shapefile_version,
  reg = reg, pop_measure = pop_measure, year_list = year_list,
  use_intermediate_years = TRUE, interval_mo = interval_mo,
  rake_subnational = rake_subnational, sharedir = sharedir,
  run_date = run_date, indicator = indicator, main_dir = main_dir,
  rake_method = "linear", age = 0, holdout = 0,
  return_objects = FALSE,
  countries_not_to_subnat_rake = countries_not_to_subnat_rake,
  custom_output_folder = NULL)
}
\arguments{
\item{cell_pred}{cell_pred object from mbg models.  Each cell must have a rate in it.  This needs to be unraked}

\item{simple_raster}{the simple raster that the cell_pred is based on}

\item{simple_polygon}{the simple polygon that the cell_pred is based on}

\item{pixel_id}{list of the pixels in the simple raster that have non na values}

\item{shapefile_version}{which shapefile geographies are being used}

\item{reg}{the modeling region}

\item{pop_measure}{the worldpop agegroup on which the model is built}

\item{year_list}{the modeled years}

\item{use_intermediate_years}{Boolean to indicate whether or not to rake to intermediate years. Default: TRUE}

\item{interval_mo}{the time in months between the modeled years}

\item{rake_subnational}{a logical value indicating the use of subnational raking targets or not}

\item{sharedir}{Path to sharedir e.g.: \code{sharedir <- sprintf('/share/geospatial/mbg/\%s/\%s',indicator_group,indicator)}}

\item{run_date}{model run date string}

\item{indicator}{modeled indicator}

\item{main_dir}{Path to main_dir e.g.: \code{main_dir <- paste0('/share/geospatial/mbg/', indicator_group, '/', indicator, '/output/', run_date, '/')}}

\item{rake_method}{if set to "logit" creates raking factors in logit space, otherwise assumes linear raking}

\item{age}{age holdout group, assumed to be 0 but left flexible}

\item{holdout}{n fold hold out group, assumed to be 0 but left flexible}

\item{return_objects}{Return cell pred and raking factors into environments? Default: FALSE}

\item{countries_not_to_subnat_rake}{character vector of iso3 codes for countries not to subnationally rake}

\item{custom_output_folder}{Get the rake factors from this folder, and also save the Rdata files here. Default: NULL}
}
\value{
automatically saves unraked and raked versions admin aggregations of both rates and counts,
as well as raked cell_pred objects of both rates and counts. If desired, one can also return raked_cell_pred
and the raking factors into environment.
}
\description{
This function:
\enumerate{ 
 \item Takes a rates cell_pred 
 \item Links it to a fractional raking and aggregation link table 
 \item Rakes the rates based on the raking factors above 
 \item Saves raked rates cell_pred 
 \item Adds population per fractional cell 
 \item Rakes that population to ensure matching to GBD using pop raking factors 
 \item Uses that population to create a counts cell_pred 
 \item Saves that raked counts cell_pred 
 \item Aggregates the cell_pred to ADM 0, 1, 2 levels 
 \item Saves aggregations at ADM 0,1,2 levels, raked and unraked, rates and counts
}
}
\details{
Fractional aggregation of a rates cell pred
}
\note{
the raked cell_pred objects cannot be used to create raked aggregation

you need to start with an unraked cell_pred, link it, rake it, then aggregate it.
}
