% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractionally_rake_counts.R
\name{fractionally_rake_counts}
\alias{fractionally_rake_counts}
\title{Fractionally Rake Counts}
\usage{
fractionally_rake_counts(count_cell_pred, rake_to, reg, year_list,
  rake_subnational, countries_not_to_subnat_rake,
  countries_not_to_rake = c("ESH+GUF"), simple_raster = NULL,
  modeling_shapefile_version = "current",
  raking_shapefile_version = "current")
}
\arguments{
\item{count_cell_pred}{Cell pred matrix object with counts to be raked.}

\item{rake_to}{Df with `name`, `year`, and `mean` columns. Values in name must be gbd loc_ids.}

\item{reg}{character string - Region used to produce cell pred object.}

\item{year_list}{integer vector - vector of years}

\item{rake_subnational}{If true, rakes to ADM1 units not in `countries_not_to_subnat_rake`}

\item{countries_not_to_subnat_rake}{default set in config, character vector of subnational country iso3 codes NOT to rake to}

\item{countries_not_to_rake}{default c("GUF", "ESH), french guiana and western sahara. Any country iso3s passed here will have raking factor set to 1.}

\item{simple_raster}{default `NULL`, option to pass in simple raster to function if it's been loaded already.}

\item{modeling_shapefile_version}{default `current`, string identifying version of of shapefile used in modeling}

\item{raking_shapefile_version}{default `current`, string identifying version of of shapefile to use in raking}
}
\value{
Returns a named list with a raked cell pred object, raking factors, and adm0, 1, and 2 aggregated draws raked and unraked
}
\description{
A function used to rake mbg count ouputs to GBD estimates using a fractional methodolgy. Counts in each pixel are assigned to admin units and split by area fraction for pixels with multiple admin units. Counts are then summed to ADM0 (except for those specified as subnational) and raking factors for those admin units are calculated. From there, the raking factors are applied to the linked cell pred, which is then processed in 2 ways: aggregated up to adm0/1/2, and deduplicated (the linked cell pred has 1 row for each pixel-adm unit-year which is reduced to 1 row per each pixel-year). N.B. If you rake subnationally and are missing raking targets for one or more subnationals, the final sum of counts will not match GBD at ADM0.
}
\examples{
\dontrun{
output <- fractionally_rake_counts(
  count_cell_pred = count_cell_pred,
  rake_to = gbd,
  reg = "soas",
  year_list,
  rake_subnational = T,
  countries_not_to_subnat_rake = c("IND", "IDN", "BRA", "MEX", "CHN"),
  countries_not_to_rake = c("GUF+ESH"),
  simple_raster = simple_raster,
  modeling_shapefile_version = shapefile_version,
  raking_shapefile_version = shapefile_version
)
}
}
