% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fss}
\alias{fss}
\title{Fitted model objects for estimating the memory of parallel_model}
\format{An object of class \code{list} of length 3.}
\usage{
fss
}
\description{
Based off of profiled model runs from 2018, AOZ
performed some nonlinear regression to estimate the relationship
between peak RAM usage and number of pixels, number of years, and
the number of posterior samples in prediction. The resulting model fit
objects are stored in this data object to allow RAM use prediction with the
function estimate_parallel_model_mem().
}
\keyword{datasets}
