% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.R
\name{G}
\alias{G}
\title{FUNCTION_TITLE}
\usage{
G(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
