% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLegend.R, R/g_legend.R, R/get_legend.R
\name{gLegend}
\alias{gLegend}
\alias{g_legend}
\alias{get_legend}
\title{Pulls legend out of a ggplot object for use later}
\usage{
gLegend(a.gplot)

g_legend(a.gplot)

get_legend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot() object}
}
\value{
grob containing legend
}
\description{
Pulls legend out of a ggplot object for use later, based off of function originally written by Hadley Wickham
}
\examples{
\dontrun{
the_legend <- g_legend(my_gplot)
}
}
