% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_covs.R
\name{gam_covs}
\alias{gam_covs}
\title{GAM Covariates}
\usage{
gam_covs(df, indicator_family, lcovs)
}
\arguments{
\item{df}{data.table with variables "latitude", "longitude", "N", and specified indicator}

\item{indicator_family}{the type of likelihood function, either \code{"binomial"} or \code{"gaussian"}}

\item{lcovs}{list of raster layers or bricks (for temporally varying). Currently assumes
four layers named like U5M periods.}
}
\value{
Two-item list (bricks of time-varying and non-time-varying covariates)
}
\description{
Take input data and covariate layers and fit GAM models
}
\note{
This function currently won't work if you have anything other than *four* distinct periods in your data and covariates.
          This will be made more flexible.
}
