% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_stacker.R
\name{gam_stacker}
\alias{gam_stacker}
\title{FUNCTION_TITLE}
\usage{
gam_stacker(df, model_names = c("gam", "gbm"), weight_column = NULL,
  bam = F, spline_args = list(bs = "ts", k = 3),
  indicator = indicator, indicator_family = indicator_family,
  cores = "auto")
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{model_names}{PARAM_DESCRIPTION, Default: c("gam", "gbm")}

\item{weight_column}{PARAM_DESCRIPTION, Default: NULL}

\item{bam}{PARAM_DESCRIPTION, Default: F}

\item{spline_args}{PARAM_DESCRIPTION, Default: list(bs = "ts", k = 3)}

\item{indicator}{PARAM_DESCRIPTION, Default: indicator}

\item{indicator_family}{PARAM_DESCRIPTION, Default: indicator_family}

\item{cores}{PARAM_DESCRIPTION, Default: 'auto'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
