% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaul_convert.R
\name{gaul_convert}
\alias{gaul_convert}
\title{FUNCTION_TITLE}
\usage{
gaul_convert(countries, from = "iso3", verbose = F,
  shapefile_version = "current")
}
\arguments{
\item{countries}{PARAM_DESCRIPTION}

\item{from}{PARAM_DESCRIPTION, Default: 'iso3'}

\item{verbose}{PARAM_DESCRIPTION, Default: F}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_match}}
}
