% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_age.R
\name{gbd_age}
\alias{gbd_age}
\title{FUNCTION_TITLE}
\usage{
gbd_age(df, age_yr)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{age_yr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
