% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qsub_command.R
\name{generate_qsub_command}
\alias{generate_qsub_command}
\title{Generates a qsub command that should work on any IHME cluster.}
\usage{
generate_qsub_command(..., stderr_log, stdout_log, project, resources,
  job_name, singularity_str, queue, cores = NULL, priority = 0,
  hold_jid = NULL)
}
\arguments{
\item{...}{The command + arguments to be run as a hibh e.g., the
singularity shell, the parallel script and its 8 arguments.}

\item{stderr_log}{string indicating directory or file to log stderr to.}

\item{stdout_log}{string indicating directory or file to log stdout to.}

\item{project}{the project to run the job under.}

\item{resources}{named vector of resources.
@seealso \code{\link{get_resources}}}

\item{job_name}{string name for the job.}

\item{cores}{numeric number of cores to use}

\item{priority}{Job priority that can be deprioritized if needed, and can only be used for values in [-1023,0]. Default = 0.
This value will get bounded to 0 or -1023 if the user supplies a value outside those bounds.}

\item{hold_jid}{Job names or job ID to hold on (uses \code{-hold_jid the_held_job_names}). Must be a vector if multiple jobs specified}

\item{singularity_opts}{string options to pass to singularity.
@seealso \code{\link{qsub_sing_envs}}}
}
\description{
Generates a qsub command that should work on any IHME cluster.
}
