% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_time_log.R
\name{generate_time_log}
\alias{generate_time_log}
\title{FUNCTION_TITLE}
\usage{
generate_time_log(ticlog)
}
\arguments{
\item{ticlog}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
