% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_polygon_quiet.R
\name{geom_polygon_quiet}
\alias{geom_polygon_quiet}
\title{FUNCTION_TITLE}
\usage{
geom_polygon_quiet(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[ggplot2]{geom_polygon}}
}
