% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aroc.R
\name{get.aroc}
\alias{get.aroc}
\title{FUNCTION_TITLE}
\usage{
get.aroc(pred.rast, year.map = NULL, pow = 1, uselogit = FALSE)
}
\arguments{
\item{pred.rast}{PARAM_DESCRIPTION}

\item{year.map}{PARAM_DESCRIPTION, Default: NULL}

\item{pow}{PARAM_DESCRIPTION, Default: 1}

\item{uselogit}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
