% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adm0_codes.R
\name{get_adm0_codes}
\alias{get_adm0_codes}
\title{Get Admin0 (Country) Codes}
\usage{
get_adm0_codes(adm0s, strict = FALSE, lookup_table = NULL,
  core_repo = NULL, adm0_type = "detect", shapefile_version = NULL,
  subnational_raking = FALSE)
}
\arguments{
\item{adm0s}{Vector of ISO-3 codes, four-letter modeling region names, region
group names as defined in get_region_groups, or 'all' (returns all Admin0
codes)}

\item{strict}{(default `FALSE`) Causes the function to fail if an ISO code or
region name is not included in the full list.}

\item{lookup_table}{(default `NULL`) Sets a data.frame or data.table to be
used as the lookup_table. If `NULL`, the lookup table will be loaded using
the `load_adm0_lookup_table()` function.}

\item{core_repo}{(default NULL) THIS ARGUMENT IS DEPRECATED AND WILL BE
REMOVED IN THE FUTURE. Please remove it from your function calls.}

\item{adm0_type}{(default 'detect') Which class of admin0 codes
should be pulled? Must be one of 'gaul', 'gadm', or 'detect'. If
'gaul' or 'gadm', it will simply use what was specified. If
'detect', the function will detect the adm0_type directly from an
admin_shapefile. It will check the adm0_type specified by the
shapefile_version argument if it is specified, otherwise if will
detect the adm0_type using 'current'.}

\item{shapefile_version}{string specifying shapefile_version to be
used in adm0_type determination if adm0_type is set to 'detect'}

\item{subnational_raking}{Logical. set to true if you want to use raking version of shapefile}
}
\value{
Vector of numeric ADM0 codes. Warns if not all regions align with a
  ADM0 code.
}
\description{
Pull Admin0 Codes for the specified countries or regions,
  optionally excluding certain Admin0 codes.
}
