% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adm_codes_subnat.R
\name{get_adm_codes_subnat}
\alias{get_adm_codes_subnat}
\title{Get subnational administrative codes (adm1 and adm2)}
\usage{
get_adm_codes_subnat(adm0_list, admin_level,
  shapefile_version = "current")
}
\arguments{
\item{adm0_list}{A numeric vector of Admin0 codes pulled using
get_adm0_codes().}

\item{admin_level}{A subnational administrative level: 1 and 2 are currently
supported.}

\item{shapefile_version}{A character string indicating which shapefile version to pull}
}
\value{
list of administrative codes
}
\description{
Pulls subnational administrative codes worldwide for a given
  administrative level (1 or 2) and a given set of country-level geographies,
  represented by a vector of standard Admin0 codes pulled using
  get_adm0_codes(). This function generalizes get_gaul_codes_subnat(), which
  will be deprecated.
}
