% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_preds.R
\name{get_admin_preds}
\alias{get_admin_preds}
\title{FUNCTION_TITLE}
\usage{
get_admin_preds(pull_year, year_list)
}
\arguments{
\item{pull_year}{PARAM_DESCRIPTION}

\item{year_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
