% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_pv.R
\name{get_admin_pv}
\alias{get_admin_pv}
\title{get_pv_admin}
\usage{
get_admin_pv(indicator, indicator_group, run_date, input_file = NULL,
  shapefile_version = "current", samp_col = "sum_of_sample_weights",
  strata = strata, indicator_family = "binomial", nid_col = "nid",
  admin0_file = NULL, admin1_file = NULL, admin2_file = NULL)
}
\arguments{
\item{indicator}{indicator name used in file structure for mbg}

\item{indicator_group}{indicator group}

\item{run_date}{model run date}

\item{input_file}{If your team does not use the standard `/share/geospatial/mbg/input_data/{indicator}.csv` input data filepath, supply the filepath to your data here}

\item{shapefile_version}{String indicating version of shapefile to pull}

\item{samp_col}{This is the name of the column that contains the sum of the sample weights for the collapsed points.}

\item{strata}{Regions specified from your model}

\item{indicator_family}{If specified as Gaussian, this makes sure to not divide the prevalence by N which is required for binomial indicators}

\item{admin0_file}{If your team does not use the standard `'/share/geospatial/mbg/',indicator_group, '/', indicator, '/output/',run_date, '/pred_derivatives/admin_summaries/',indicator, '_admin_0_unraked_summary.csv'` directory to save unraked aggregated admin summaries, supply the correct filepath here}

\item{admin1_file}{If your team does not use the standard `'/share/geospatial/mbg/',indicator_group, '/', indicator, '/output/',run_date, '/pred_derivatives/admin_summaries/',indicator, '_admin_1_unraked_summary.csv'` directory to save unraked aggregated admin summaries, supply the correct filepath here}

\item{admin2_file}{If your team does not use the standard `'/share/geospatial/mbg/',indicator_group, '/', indicator, '/output/',run_date, '/pred_derivatives/admin_summaries/',indicator, '_admin_2_unraked_summary.csv'` directory to save unraked aggregated admin summaries, supply the correct filepath here}

\item{input_data}{If specified, provides the preloaded input data so the function does not look in your model directory}

\item{nic_col}{This is the name of the unique survey variable, usually labeled "nid" but other teams might use different terminology}
}
\value{
a table with bias, rmse, mae, and se for each admin level
}
\description{
This function collapses input data to admin 0/admin 1/admin 2 using the `input_aggregate_admin` function
See `input_aggregate_admin` function documentation for input data requirements

To use this function, your input dataset must have a column that is the sum of the sample weights

This function pulls the input data from the model directory and aggregated admin summaries created in the `aggregate_results.R` script

This function returns average RMSE, Bias, Mean Absolute Error, and Standard Error for admin 0 /admin 1 / admin 2
}
\author{
Lauren Woyczynski, \email{lpw3@uw.edu}
}
