% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_shape_dir.R
\name{get_admin_shape_dir}
\alias{get_admin_shape_dir}
\title{Return path to admin shapes directory}
\usage{
get_admin_shape_dir(version = "current")
}
\arguments{
\item{version}{Admin shapefile version to pull}
}
\description{
Returns path to the official LBD administrative shape file directory. This
actually includes non-shape data that is important for mapping, notably
the standard link table and standard id raster.
}
