% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_shapefile.R
\name{get_admin_shapefile}
\alias{get_admin_shapefile}
\title{Return path to world admin shapefile at specified admin level}
\usage{
get_admin_shapefile(admin_level = 0, suffix = ".shp", type = "admin",
  version = "current", raking = F)
}
\arguments{
\item{admin_level}{Valid admin level we have a shapefile for. Current 0/1/2.}

\item{suffix}{'.shp' by default, provide any other suffix to e.g., get the .dbf file
associated with the admin shapefile.}

\item{raking}{boolean, default F. If TRUE pulls subnational raking shapefile.}
}
\description{
Returns path to world admin shapefile given \code{admin_level}. stop()s if
no file exists at that admin_level. Defaults to returning the ".shp" file
path, but will substitute any other file \code{suffix} provided.
}
\examples{
\dontrun{
get_admin_shapefile(2)
get_admin_shapefile(2, suffix = ".dbf")
}
}
