% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_group_from_worldpop.R
\name{get_age_group_from_worldpop}
\alias{get_age_group_from_worldpop}
\title{Get GBD age group for the given WorldPop measure}
\usage{
get_age_group_from_worldpop(pop_measure)
}
\arguments{
\item{pop_measure}{The WorldPop measure, e.g. \code{a0004t}. Default: \code{a0004t}}
}
\value{
The \code{age_group_id} for the associated \code{pop_measure} input.
}
\description{
This function will return the age group from the GBD age group ID table
which corresponds to the WorldPop measures used
}
\note{
Currently only taking in measures starting with "a" and for 12 month interval (ending with "t"). And \code{total}.
}
