% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ages.R
\name{get_ages}
\alias{get_ages}
\title{FUNCTION_TITLE}
\usage{
get_ages()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
