% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aggs.R
\name{get_aggs}
\alias{get_aggs}
\title{FUNCTION_TITLE}
\usage{
get_aggs(agg_geo_est = cond_sim_adm0, gaul_list = gaul_list,
  rake_to = gbd)
}
\arguments{
\item{agg_geo_est}{PARAM_DESCRIPTION, Default: cond_sim_adm0}

\item{gaul_list}{PARAM_DESCRIPTION, Default: gaul_list}

\item{rake_to}{PARAM_DESCRIPTION, Default: gbd}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
