% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_pred_for_aroc.R
\name{get_cell_pred_for_aroc}
\alias{get_cell_pred_for_aroc}
\title{Pull cell pred object and format for aroc or projection calculations}
\usage{
get_cell_pred_for_aroc(ind_gp, ind, rd, reg, measure,
  matrix_pred_name = NULL, skip_cols = NULL, rk = T,
  shapefile_version = "current")
}
\arguments{
\item{ind_gp}{indicator group}

\item{ind}{indicator}

\item{rd}{run_date}

\item{reg}{region}

\item{measure}{prevalence, incidence, mortality, etc}

\item{matrix_pred_name}{In \code{sprintf} notation. The one object passed into
the string should will be a region name. this allows different regions to be
passed to different named matrix_preds (pixel level, ad0, ad1, ad2, ...)
e.g. 'had_diarrhea_cell_draws_eb_bin0_\%s_diarrhea2_0.RData' which
will be passed to sprintf('had_diarrhea_cell_draws_eb_bin0_\%s_0.RData', reg)}

\item{skip_cols}{columns to skip when reading in the cell preds
For example, if the first two columns store non-pred information in your
file format, \code{skip_cols = 2} will read in all columns from 3 onwards}

\item{rk}{raked? (logical)}

\item{shapefile_version}{string which shapefile version to use for indexing cell_preds}
}
\value{
a formated \code{cell_pred} object
}
\description{
A helper function that pulls a cell pred object and then formats it for use in other functions
}
