% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_scheme.R
\name{get_color_scheme}
\alias{get_color_scheme}
\title{Get color scheme}
\usage{
get_color_scheme(theme)
}
\arguments{
\item{theme}{Name of theme}
}
\value{
A vector of color hex codes
}
\description{
Grabs a color scheme from a set of defaults
}
\examples{
\dontrun{
get_color_scheme("carto_discrete")
}
}
