% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_con.R
\name{get_con}
\alias{get_con}
\title{FUNCTION_TITLE}
\usage{
get_con(dbname, host)
}
\arguments{
\item{dbname}{PARAM_DESCRIPTION}

\item{host}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
