% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov_estimates.R
\name{get_cov_estimates}
\alias{get_cov_estimates}
\title{FUNCTION_TITLE}
\usage{
get_cov_estimates(covariate_name_short, filters = list())
}
\arguments{
\item{covariate_name_short}{PARAM_DESCRIPTION}

\item{filters}{PARAM_DESCRIPTION, Default: list()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[RMySQL]{MySQLDriver-class}}
}
