% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariate_metadata.R
\name{get_covariate_metadata}
\alias{get_covariate_metadata}
\title{FUNCTION_TITLE}
\usage{
get_covariate_metadata(list = NULL)
}
\arguments{
\item{list}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
