% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariate_version.R
\name{get_covariate_version}
\alias{get_covariate_version}
\title{FUNCTION_TITLE}
\usage{
get_covariate_version(list)
}
\arguments{
\item{list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
