% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R
\name{get_covariates}
\alias{get_covariates}
\title{FUNCTION_TITLE}
\usage{
get_covariates(list, ci = FALSE)
}
\arguments{
\item{list}{PARAM_DESCRIPTION}

\item{ci}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
