% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fhs_population.R
\name{get_fhs_population}
\alias{get_fhs_population}
\title{Get pre-saved FHS population outputs}
\usage{

  get_fhs_population(population_version = "20190403_test_new_cluster_1000d_rerun_fix_draw_squeezed_agg_ordered",
  pop_measure = "a0004t", sex_ids = 3, scenarios = 0,
  year_ids = NULL, gbd_regions = NULL)
}
\arguments{
\item{population_version}{A version tag found from looking into: \code{/share/scratch/users/sadatnfs/geospatial_shared/population_fbd}.
Default: \code{"20190403_test_new_cluster_1000d_rerun_fix_draw_squeezed_agg_ordered"}}

\item{pop_measure}{WorldPop measure; only one allowed for now! Default: a0004t}

\item{gbd_regions}{Regions to query (GBD Location IDs). Default: NULL (get all regions)}

\item{sex_id}{Sex ID. Default: 3}

\item{scenario}{FHS scenario. Default: 0 (reference).}

\item{year_ids.}{Year_id to query. Default: NULL (get all years)}
}
\value{
A data.table with location_id, year_id, age_group_id = pop_measure, sex_id, run_id, population
}
\description{
Access pre-saved RDS population outputs from FHS (saved out by Nafis Sadat in \code{/share/scratch/users/sadatnfs/geospatial_shared/population_fbd})
}
