% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fractional_rf.R
\name{get_fractional_rf}
\alias{get_fractional_rf}
\title{Calculate Fractional Raking Factors}
\usage{
get_fractional_rf(cell_pred, link, gbd, region, overs, shapefile_version)
}
\arguments{
\item{cell_pred}{a cell pred with deaths/cases in data.table format with a pixel_id column from `get_link_table()`}

\item{link}{link table gotten from `get_link_table()`}

\item{gbd}{data.table with a columns - `gaul`, `year_id`, and `gbd_field (see below)`
with estimates for the indicator being raked to for each year-admin0}

\item{region}{the region used to make the simple raster}

\item{overs}{a vector of the column names in the cell pred corresponding to draws
(typically V1:Vndraws)}

\item{shapefile_version}{string indicating which shapefile version to match ADM0 codes against}
}
\value{
a \code{data.table} with raking factor by adm0 gaul code and year
}
\description{
Calculate raking factors considering cases where a pixel is in
multiple admin units for deaths or cases
}
\details{
In using 5x5 pixels, some pixels are split between several admin
units. Fractional raking is a way of raking counts which takes
this into consideration. For any pixel that is split between
admin units, the fractional area of that admin unit within the
pixel is multiplied by the count so that counts are properly
attributed to their respective admin units.

The key to fractional raking is a [link table] which has a row
for each pixel-admin unit, and includes the gaul codes for admins
0-2, as well as the fractional area of the admin unit in that pixel
(a pixel split between two admin 2 units would have two rows in the
link table). There is a global link table generated from the stage 1
and stage 2 simple raster, where every pixel gets a unique id.
Practically, when using a link table, the global link table and raster
are subsetted based on the simple raster of the region being raked
using get_link_table()

Build_link_polygon() and build_link_table() are functions used to
create the link table, which will only need to be used when
there is a change in the shapefile used to generate simple rasters.

Many of the functions require the parameter overs, which is a vector
of the column names of the draws, typically named V1 to Vndraws
when the cell pred is converted to a data.frame or data.table.


For additional questions, ask Michael Collison
}
