% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gaul_codes.R
\name{get_gaul_codes}
\alias{get_gaul_codes}
\title{Get GAUL Codes (TO BE DEPRECATED)}
\usage{
get_gaul_codes(adm0s, strict = FALSE, lookup_table = NULL,
  core_repo = NULL, adm0_type = "gaul", shapefile_version = NULL)
}
\description{
Pull Admin0 codes for the specified countries or regions,
  optionally excluding certain Admin0 codes. The only difference from the
  standard get_adm0_codes() function is that the default adm0_type is 'gaul'
  rather than 'gadm'.
}
\note{
THIS FUNCTION IS BEING DEPRECATED - ALWAYS USE get_adm0_codes() INSTEAD.
}
