% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gaul_codes_subnat.R
\name{get_gaul_codes_subnat}
\alias{get_gaul_codes_subnat}
\title{Get subnational GAUL codes}
\usage{
get_gaul_codes_subnat(gaul_list, admin_level, shapefile_version)
}
\description{
Pulls subnational administrative codes worldwide for a given
  administrative level (1 or 2) and a given set of country-level geographies,
  represented by a vector of standard Admin0 codes pulled using
  get_adm0_codes(). This function is a simple wrapper for
  get_adm_codes_subnat() and has been kept for backwards-compatibility.
}
\note{
THIS FUNCTION IS BEING DEPRECATED. ALWAYS USE \code{get_adm_codes_subnat()}
  INSTEAD.
}
