% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbd_estimates.R
\name{get_gbd_estimates}
\alias{get_gbd_estimates}
\title{Get GBD estimates}
\usage{
get_gbd_estimates(gbd_type, gbd_name, region, measure_id = 6,
  age_group_id = 1, metric_id = 3, year_ids = c(2000:2017),
  shapefile_version = "current", rake_subnational = TRUE,
  gbd_round_id = 5)
}
\arguments{
\item{gbd_name}{= GBD cov_name_short if "covariate" and GBD cause_id if "output"}

\item{region}{= name of region you want to pull results for. e.g. 'africa'}

\item{measure_id}{= if "output", which measure you want to pull (defaults to incidence)}

\item{age_group_id}{= if "output", which age group you want to pull (defaults to Under 5)}

\item{metric_id}{= if "output", which metric you want to pull (defaults to rate)}

\item{year_ids}{= numeric vector of years to pull}

\item{shapefile_version}{string of dated shapefile version to use when generating list of ihme loc ids and ADM code}

\item{rake_subnational}{Logical. do you want subnational estimates or just national ones}

\item{gbd_round_id}{numeric gbd round id to pull from}
}
\value{
Returns 3-column data.table where "name" = ihme loc id, "year" = year, and
     "mean" = value. Consistent with expected input for rake_cell_pred and calculate_raking_factors.
}
\description{
Loads national or subnational gbd estimates to be used for raking.
}
