% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbd_locs.R
\name{get_gbd_locs}
\alias{get_gbd_locs}
\title{Get GBD locations}
\usage{
get_gbd_locs(reg, rake_subnational = T,
  shapefile_version = raking_shapefile_version)
}
\arguments{
\item{reg}{region you want to pull codes for. e.g. 'africa'}

\item{rake_subnational}{logical - do you want to pull location codes for subnational raking or just national raking}

\item{shapefile_version}{string specifying version date of shapefile to use}
}
\value{
returns a dataframe containing, at a minimum, an ADM_CODE column and an loc_id (ihme loc id) column. If rake_subnational=T, it will also contain ADM0*, ADM1* and rak_level (admin level for raking) columns
}
\description{
Loads in the shapefiles to get a complete set of codes for GBD
locations. this is particularly useful if you want to get
subnational GBD location codes which are a combination of different
administrative levels
}
