% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ihme_lc_id.R
\name{get_ihme_lc_id}
\alias{get_ihme_lc_id}
\title{Generate location ids}
\usage{
get_ihme_lc_id(iso_lookup, loc_id_hierarchy)
}
\arguments{
\item{iso_lookup}{a data.table with loc_id and iso_name columns.}

\item{loc_id_hierarchy}{a comma-delimited string of loc_id values
representing the logical path from EARTH through each admin unit to the
desired location.}
}
\value{
String representing the location id.
}
\description{
Generate \code{ihme_lc_id} for a national or subnational location
}
