% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_dir.R
\name{get_indicator_dir}
\alias{get_indicator_dir}
\title{get_indicator_dir}
\usage{
get_indicator_dir(indicator_group, indicator)
}
\arguments{
\item{indicator_group}{the indicator group e.g., "hiv" or "u5m".}

\item{indicator}{the indicator e.g., "wasting_mod_b" or "male_circumcision".}
}
\value{
filesystem path.
}
\description{
Compute the directory for outputs for a given indicator
}
