% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_link_table.R
\name{get_link_table}
\alias{get_link_table}
\title{Subset global link table to regional table}
\usage{
get_link_table(simple_raster, shapefile_version)
}
\arguments{
\item{simple_raster}{Simple raster used for modelling region}
}
\value{
list of 2 objects:
   - a link table
   - a vector of pixel ids used to link one year in a cell pred to the link table.
}
\description{
Reads in the global link table and id raster and subsets using
a regional simple raster.
}
