% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_code_mapping.R
\name{get_location_code_mapping}
\alias{get_location_code_mapping}
\title{Return location metadata}
\usage{
get_location_code_mapping(shapefile_version, remove_diacritics = T)
}
\arguments{
\item{shapefile_version}{string indicating version of shapefile to use. 'gaul' or 'gadm' is inferred from this}

\item{fix_diacritics}{(default TRUE) whether to strip all diacritic characters from result.}
}
\value{
data.frame with above columns
}
\description{
Get location metadata for:
 IHME location ids (loc_id),
 Admin codes (ADM_CODE),
 IHME loc ids (ihme_lc_id),
 Long location names (loc_name),
 Short location names (loc_nm_sh)
}
\note{
: includes a \code{GAUL_CODE} column (same values as \code{ADM_CODE}) for compatibility.
}
