% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_code_mapping_GAUL.R
\name{get_location_code_mapping_GAUL}
\alias{get_location_code_mapping_GAUL}
\title{FUNCTION_TITLE}
\usage{
get_location_code_mapping_GAUL(remove_diacritics)
}
\arguments{
\item{remove_diacritics}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[DBI]{dbDisconnect}}
}
