% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_hierarchy.R
\name{get_location_hierarchy}
\alias{get_location_hierarchy}
\title{FUNCTION_TITLE}
\usage{
get_location_hierarchy(location_set_version_id, china.fix = FALSE)
}
\arguments{
\item{location_set_version_id}{PARAM_DESCRIPTION}

\item{china.fix}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
