% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_runtime_by_queue.R
\name{get_max_runtime_by_queue}
\alias{get_max_runtime_by_queue}
\title{Get max runtime of queue}
\usage{
get_max_runtime_by_queue(queue)
}
\arguments{
\item{queue}{String. User supplied queue}
}
\value{
a string of run-time in format \code{DD:HH:MM:SS}
}
\description{
Check the input string of queue and the maximum runtime available for that queue
}
