% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mkl_threads.R
\name{get_mkl_threads}
\alias{get_mkl_threads}
\title{Finds number of threads set for MKL operations}
\usage{
get_mkl_threads()
}
\value{
number of threads assigned to MKL
}
\description{
\code{get_mkl_threads()} Uses environmental variable "MKL_NUM_THREADS" used
in LBD Singualirty images to determine how many threads have been assigned
for MKL operations.
}
\details{
The OpenMP function \code{omp_get_num_threads()} will report how many threads
have been set for OpenMP operations. Unfortunately, there is no
"mkl_get_num_threads()" function in the MKL library, so we have to rely on
our MKL_NUM_THREADS environmental variable to find out how many threads have
been assigned for MKL operations. Fortunately, we can guarantee that
MKL_NUM_THREADS has been set in LBD Singularity containers spun up by either
"shell_sing.sh" or "singR.sh".
}
\seealso{
This function is used by:
\code{\link{get_total_threads()}}

Other Mutlti-threading Functions: \code{\link{get_max_forked_threads}},
  \code{\link{get_omp_threads}},
  \code{\link{get_total_threads}},
  \code{\link{setmkldynamic}}, \code{\link{setmklthreads}},
  \code{\link{setompnested}}, \code{\link{setompthreads}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{Mutlti-threading Functions}
