% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_output_dir.R
\name{get_model_output_dir}
\alias{get_model_output_dir}
\title{Get model output directory}
\usage{
get_model_output_dir(indicator_group, indicator, run_date)
}
\arguments{
\item{indicator_group}{the indicator group e.g., "hiv" or "u5m".}

\item{indicator}{the indicator e.g., "wasting_mod_b" or "male_circumcision".}

\item{run_date}{string run date in YYYY_MM_DD_HH_MM_SS format.}
}
\value{
the path to the models output directory.
}
\description{
return path to model output directory.
}
