% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_newest_worldpop_release.R
\name{get_newest_worldpop_release}
\alias{get_newest_worldpop_release}
\title{Returns newest worldpop measure release}
\usage{
get_newest_worldpop_release(config)
}
\arguments{
\item{config}{a \code{data.table} with the existing configuration. This MUST have two columns named V1 and V2. V1 should contain configuration variable names one of which is "pop_measure". The V2 should contain configuration values e.g., "a0004t".}
}
\value{
character for the newest release e.g., "2017_04_27"
}
\description{
Checks the existing \code{config} object to determine which worldpop measure to check, and returns thew newest release for that measure.
}
