% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omp_threads.R
\name{get_omp_threads}
\alias{get_omp_threads}
\title{Finds number of threads set for OpenMP operations}
\usage{
get_omp_threads()
}
\value{
number of threads assigned to OpenMP
}
\description{
\code{get_omp_threads()} Uses environmental variable "OMP_NUM_THREADS" used
in LBD Singualirty images to determine how many threads have been assigned
for OpenMP operations.
}
\details{
The OpenMP function \code{omp_get_num_threads()} will report how many threads
have been set for OpenMP operations. Unfortunately, there is no
"mkl_get_num_threads()" function in the MKL library, so we have to rely on
our OMP_NUM_THREADS environmental variable to find out how many threads have
been assigned for OpenMP operations as we do for MKL. Fortunately, we can
guarantee that OMP_NUM_THREADS has been set in LBD Singularity containers
spun up by either "shell_sing.sh" or "singR.sh".
}
\seealso{
This function is used by:
\code{\link{get_total_threads()}}

Other Mutlti-threading Functions: \code{\link{get_max_forked_threads}},
  \code{\link{get_mkl_threads}},
  \code{\link{get_total_threads}},
  \code{\link{setmkldynamic}}, \code{\link{setmklthreads}},
  \code{\link{setompnested}}, \code{\link{setompthreads}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{Mutlti-threading Functions}
