% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_target.R
\name{get_p_target}
\alias{get_p_target}
\title{FUNCTION_TITLE}
\usage{
get_p_target(draws, target_type, target)
}
\arguments{
\item{draws}{PARAM_DESCRIPTION}

\item{target_type}{PARAM_DESCRIPTION}

\item{target}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
