% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentile.R
\name{get_percentile}
\alias{get_percentile}
\title{FUNCTION_TITLE}
\usage{
get_percentile(x, percentile)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{percentile}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
