% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_data.R
\name{get_population_data}
\alias{get_population_data}
\title{FUNCTION_TITLE}
\usage{
get_population_data(simple_raster)
}
\arguments{
\item{simple_raster}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
