% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_project}
\alias{get_project}
\title{Get project}
\usage{
get_project(proj, use_geo_nodes)
}
\arguments{
\item{proj}{str project name (may be NULL)}

\item{use_geo_nodes}{logical indicating whether to use geospatial nodes.}
}
\value{
proj (if not null) or an appropriate default project.
}
\description{
Return default project if not set.
}
