% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qa_wrapper.R
\name{get_qa_wrapper}
\alias{get_qa_wrapper}
\title{FUNCTION_TITLE}
\usage{
get_qa_wrapper(jobids)
}
\arguments{
\item{jobids}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
