% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qacct_exit_status.R
\name{get_qacct_exit_status}
\alias{get_qacct_exit_status}
\title{FUNCTION_TITLE}
\usage{
get_qacct_exit_status(jobid)
}
\arguments{
\item{jobid}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
