% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_queue.R
\name{get_queue}
\alias{get_queue}
\title{Get queue}
\usage{
get_queue(use_geo_nodes, use_c2_nodes, queue = NULL, run_time = NULL)
}
\arguments{
\item{use_geo_nodes}{Boolean. Use geo nodes?}

\item{use_c2_nodes}{Boolean. Use c2 nodes? Only relevant to cluster-prod.}

\item{queue}{String. User supplied queue.}

\item{run_time}{String. User supplied run-time}
}
\value{
A queue string.
}
\description{
Get queue based on user supplied queue, run-time, and whether to use geos nodes or c2 nodes
If queue is already supplied by the user, then we just validate that string and return it
}
