% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_str.R
\name{get_resource_str}
\alias{get_resource_str}
\title{Get resource string}
\usage{
get_resource_str(resources)
}
\arguments{
\item{resources}{named vector of resources.}
}
\value{
String of resources
}
\description{
Compute a string of resources to request for a qsub command.
}
\seealso{
\code{\link{get_resources}}
}
