% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources.R
\name{get_resources}
\alias{get_resources}
\title{Get Resources}
\usage{
get_resources(..., use_geo_nodes, cores, ram_gb, runtime = NULL)
}
\arguments{
\item{use_geo_nodes}{logical in dicating if code should run on LBD nodes.}

\item{cores}{numeric indicating the number of cores to request.}

\item{ram_gb}{numeric indicating the number of GB of RAM to request.}

\item{runtime}{(default NULL) runtime in 'DD:HH:MM:SS' format.}
}
\value{
named vector with resource name and string values.
}
\description{
Get list of resources to request for new cluster.
}
