% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_time.R
\name{get_run_time}
\alias{get_run_time}
\title{Get run-time}
\usage{
get_run_time(use_geo_nodes, use_c2_nodes, queue = NULL,
  run_time = NULL)
}
\arguments{
\item{use_geo_nodes}{Boolean. Use geo nodes?}

\item{use_c2_nodes}{Boolean. Use c2 nodes? Only relevant to cluster-prod.}

\item{queue}{String. User supplied queue.}

\item{run_time}{String. User supplied run-time}
}
\value{
A run-time string in format \code{DD:HH:MM:SS} if using fair cluster. Returns \code{NULL} otherwise.
If run-time is empty, then it return the max runtime
}
\description{
Get run-time based on user supplied queue, run-time, and whether to use geos nodes or c2 nodes
If run-time is already supplied by the user, then we just validate that string and return it
}
