% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_counts.R
\name{get_sample_counts}
\alias{get_sample_counts}
\title{FUNCTION_TITLE}
\usage{
get_sample_counts(ss = 1, xy, shapes, shape_ident = "gaul_code", ...)
}
\arguments{
\item{ss}{PARAM_DESCRIPTION, Default: 1}

\item{xy}{PARAM_DESCRIPTION}

\item{shapes}{PARAM_DESCRIPTION}

\item{shape_ident}{PARAM_DESCRIPTION, Default: 'gaul_code'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
