% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sexes.R
\name{get_sexes}
\alias{get_sexes}
\title{FUNCTION_TITLE}
\usage{
get_sexes()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
