% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sge_output_dir.R
\name{get_sge_output_dir}
\alias{get_sge_output_dir}
\title{SGE output directory}
\usage{
get_sge_output_dir(user)
}
\arguments{
\item{user}{the name of the user.}
}
\value{
the root directory for sge error/output logs.
}
\description{
return the directory where sgeoutput is usually kept.
}
