% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shared_db_conn.R
\name{get_shared_db_conn}
\alias{get_shared_db_conn}
\title{FUNCTION_TITLE}
\usage{
get_shared_db_conn()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[RMySQL]{character(0)}},\code{\link[RMySQL]{MySQLDriver-class}},\code{\link[RMySQL]{constants}}
}
