% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_hierarchy.R
\name{get_sp_hierarchy}
\alias{get_sp_hierarchy}
\title{Pull the hierarchy of admin codes}
\usage{
get_sp_hierarchy(shapefile_version = "current")
}
\arguments{
\item{shapefile_version}{character string indicating version of shapefile to pull}
}
\value{
a list of 3 data tables (adm0, adm1, adm2). Each data table includes
 the location codes of its administrative level and columns for the higher-
 order administrative levels as well.
}
\description{
Takes the admin2 shapefile and pulls a list of
admin level codes to be used in merges, etc.
}
