% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strat_combos.R
\name{get_strat_combos}
\alias{get_strat_combos}
\title{FUNCTION_TITLE}
\usage{
get_strat_combos(data = data, strat_cols = strat_cols, ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION, Default: data}

\item{strat_cols}{PARAM_DESCRIPTION, Default: strat_cols}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
