% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strat_rows.R
\name{get_strat_rows}
\alias{get_strat_rows}
\title{FUNCTION_TITLE}
\usage{
get_strat_rows(data = data, strata, ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION, Default: data}

\item{strata}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
