% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template_raster.R
\name{get_template_raster}
\alias{get_template_raster}
\title{FUNCTION_TITLE}
\usage{
get_template_raster(simple)
}
\arguments{
\item{simple}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
