% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weighted_correlation_raw_estimated.R
\name{get_weighted_correlation_raw_estimated}
\alias{get_weighted_correlation_raw_estimated}
\title{Get weight correlation between raw data and unraked estimates}
\usage{
get_weighted_correlation_raw_estimated(indicator_group, indicator, rundate)
}
\arguments{
\item{indicator_group}{The string of the indicator group.}

\item{indicator}{The string name of the indicator.}

\item{rundate}{The string name of the rundate.}
}
\value{
Returns the output from a call to the weights::wtd.cors() function
between the raw data points and the estimates.
}
\description{
Returns the weighted correlation between the raw data points input into a model run,
and the unraked estimates.
}
\author{
Rebecca Stubbs
}
