% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBranches.R
\name{getBranches}
\alias{getBranches}
\title{FUNCTION_TITLE}
\usage{
getBranches(x, parent = "")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
