% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentreScale.R
\name{getCentreScale}
\alias{getCentreScale}
\title{FUNCTION_TITLE}
\usage{
getCentreScale(x, exclude = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{exclude}{PARAM_DESCRIPTION, Default: NULL}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
