% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCols.R
\name{getCols}
\alias{getCols}
\title{FUNCTION_TITLE}
\usage{
getCols(df, period = 1)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{period}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
