% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEst.R
\name{getEst}
\alias{getEst}
\title{FUNCTION_TITLE}
\usage{
getEst(df, iso3, year)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{iso3}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
