% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPaths.R
\name{getPaths}
\alias{getPaths}
\title{FUNCTION_TITLE}
\usage{

  getPaths(pathfile = "~/Z/ABRAID/prevalence modelling/under five mortality/paths_for_nick.csv")
}
\arguments{
\item{pathfile}{PARAM_DESCRIPTION, Default: '~/Z/ABRAID/prevalence modelling/under five mortality/paths_for_nick.csv'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
