% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPointsWrapper.R
\name{getPointsWrapper}
\alias{getPointsWrapper}
\title{FUNCTION_TITLE}
\usage{
getPointsWrapper(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
