% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPointsWrapper_w_dict.R
\name{getPointsWrapper.w.dict}
\alias{getPointsWrapper.w.dict}
\title{FUNCTION_TITLE}
\usage{
getPointsWrapper.w.dict(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
