% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPoly.R
\name{getPoly}
\alias{getPoly}
\title{FUNCTION_TITLE}
\usage{
getPoly(level, code, admin1, admin2)
}
\arguments{
\item{level}{PARAM_DESCRIPTION}

\item{code}{PARAM_DESCRIPTION}

\item{admin1}{PARAM_DESCRIPTION}

\item{admin2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
