% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpolys.R
\name{getpolys}
\alias{getpolys}
\title{FUNCTION_TITLE}
\usage{
getpolys(i)
}
\arguments{
\item{i}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
