% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_stacker.R
\name{glm_stacker}
\alias{glm_stacker}
\title{FUNCTION_TITLE}
\usage{
glm_stacker(df, model_names = c("gam", "gbm"), indicator = indicator,
  indicator_family = indicator_family)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{model_names}{PARAM_DESCRIPTION, Default: c("gam", "gbm")}

\item{indicator}{PARAM_DESCRIPTION, Default: indicator}

\item{indicator_family}{PARAM_DESCRIPTION, Default: indicator_family}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
