% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_time_hours.R
\name{grab_time_hours}
\alias{grab_time_hours}
\title{FUNCTION_TITLE}
\usage{
grab_time_hours(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
