% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_run_summary.R
\name{graph_run_summary}
\alias{graph_run_summary}
\title{FUNCTION_TITLE}
\usage{
graph_run_summary(run_date, indicator, indicator_group, return_graph = F)
}
\arguments{
\item{run_date}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{return_graph}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
