% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_wt_mean.R
\name{group_wt_mean}
\alias{group_wt_mean}
\title{FUNCTION_TITLE}
\usage{
group_wt_mean(i)
}
\arguments{
\item{i}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
