% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_quadtree_leaf.R
\name{id.quadtree.leaf}
\alias{id.quadtree.leaf}
\title{FUNCTION_TITLE}
\usage{
id.quadtree.leaf(q, ...)
}
\arguments{
\item{q}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
