% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_aggregate_admin.R
\name{input_aggregate_admin}
\alias{input_aggregate_admin}
\title{Correctly assign and then collapse input data to the admin0/ admin1/ admin2 level
to be used in the visualization function and add input data to plots}
\usage{
input_aggregate_admin(indicator, indicator_group, regions = c("cssa",
  "wssa", "essa", "sssa", "name"), run_date = NULL, input_data = NULL,
  indicator_family = "binomial", svy_id = "nid",
  sample_column = "sum_of_sample_weights", subnational_nids = NULL,
  shapefile_version = "current", use_kish = FALSE)
}
\arguments{
\item{indicator}{indicator name used in file structure for mbg}

\item{indicator_group}{indicator group}

\item{regions}{Regions specified from your model}

\item{run_date}{model run date}

\item{input_data}{If specified, provides the preloaded input data so the function does not look in your model directory}

\item{indicator_family}{If specified as Gaussian, this makes sure to not divide the prevalence by N which is required for binomial indicators}

\item{svy_id}{This is the unique survey id, usually labeled "nid" but other teams might use different terminology}

\item{sample_column}{This is the name of the column that contains the sum of the sample weights for the collapsed points.}

\item{subnational_nids}{If specified, these are the nids of surveys that are not nationally representative, and appear on the data plots with a different shape}

\item{shapefile_version}{String indicating version of shapefile to pull}

\item{use_kish}{boolean - If TRUE, admin aggregated N's are calculated using Kish approx}
}
\value{
a list containing three data tables that correspond to collapsed admin levels, named ad0/ad1/ad2. This list should then
be assigned to parameters ad0_data/ad1_data/ad2_data for the visualization function if plot_data == T. The admin data will have columns
named outcome
}
\description{
This function pulls the input data from the model directory, however it requires that you include
a column that retains the sum of the sample weights when collapsing to lat/long or shapefiles in your collapse
code in order to correctly recollapse to the specified admin. For this reason, it is important to include this
information in the collapse code before using this function. You can name the column whatever you want, though the
default name is sum_of_sample_weights. This is an argument to the function, sample_column. The input data must also have
a column labeled "point" that is an indicator for whether the data is point data (1) or polygon data (0).

You may also specify the input
data yourself by passing the loaded input data as an argument to the function, do this if your input data is not specified
in the model directory, though this should be the case for most using the mbg pipeline.

This function uses the sf package, which will currently only work when running on the lbd_singularity image or in
RStudio on the cluster.
}
\examples{
\dontrun{
# Set up directories and files
# Set `run_date`, `indicator`, `indicator_group`, `sample_column` as you wish

admin_data <- input_aggregate_admin(indicator = indicator, indicator_group, regions = c("cssa", "wssa", "essa", "sssa"), indicator_family = "binomial", sample_column = "sum_of_sample_weights")
ad0_data <- admin_data$ad0
ad1_data <- admin_data$ad1
ad2_data <- admin_data$ad2

## Running the plotting code with plot_data = T
subnational_ts_plots(
  ad0_df = ad0_df,
  ad1_df = ad1_df,
  ad2_df = ad2_df,
  ad0_shape = ad0_shape,
  ad1_shape = ad1_shape,
  ad2_shape = ad2_shape,
  ind_title = "My Indicator",
  out_dir = out_dir,
  highisbad = F,
  val_range = c(0, 1),
  ad0_map_regions = c("cssa", "essa", "name", "sssa", "wssa"),
  ad0_map_region_titles = c(
    "Central Sub-Saharan Africa",
    "Eastern Sub-Saharan Africa",
    "Northern Africa",
    "Southern Sub-Saharan Africa",
    "Western Sub-Saharan Africa"
  ),
  verbose = T,
  plot_data = T,
  ad0_data = ad0_data,
  ad1_data = ad1_data,
  ad2_data = ad2_data
)
}
}
\author{
Michael Cork, \email{mcork23\@uw.edu}
}
