% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_clean.R
\name{input_clean}
\alias{input_clean}
\title{FUNCTION_TITLE}
\usage{
input_clean(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
