% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertRaster.R
\name{insertRaster}
\alias{insertRaster}
\title{FUNCTION_TITLE}
\usage{
insertRaster(raster, new_vals, idx = NULL)
}
\arguments{
\item{raster}{PARAM_DESCRIPTION}

\item{new_vals}{PARAM_DESCRIPTION}

\item{idx}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{brick}}
}
