% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gbd.R
\name{interpolate_gbd}
\alias{interpolate_gbd}
\title{FUNCTION_TITLE}
\usage{
interpolate_gbd(gbd)
}
\arguments{
\item{gbd}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
