% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr.R
\name{iqr}
\alias{iqr}
\title{FUNCTION_TITLE}
\usage{
iqr(x, quantile_low = 25, quantile_high = 75)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{quantile_low}{PARAM_DESCRIPTION, Default: 25}

\item{quantile_high}{PARAM_DESCRIPTION, Default: 75}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
