% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_error.R
\name{is.error}
\alias{is.error}
\title{FUNCTION_TITLE}
\usage{
is.error(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
