% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_admin_shapefile_string.R
\name{is_admin_shapefile_string}
\alias{is_admin_shapefile_string}
\title{Return logical indicating if string is an admin shapefile version string.}
\usage{
is_admin_shapefile_string(s)
}
\arguments{
\item{s}{String the string to check.}
}
\description{
Checks strings to see if they are equal to "current" or match a YYYY_MM_DD format.
If so, returns TRUE. Else FALSE. No checking is done to ensure that the date string
has a corresponding subdirectory in the admin_shapefiles directory.
}
