% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer.R
\name{is_integer}
\alias{is_integer}
\title{Tests if object is an integer}
\usage{
is_integer(i)
}
\arguments{
\item{i}{An object}
}
\value{
TRUE/FALSE
}
\description{
\code{is_integer()} Tests if object is an integer since \code{is.integer()}
does not.
}
\details{
We need to be able to test if an object is an integer or not. Unfortunately
R's \code{is.integer()} does not actually tell you if an object is an integer
(`is.integer(5)` returns FALSE for example).  This function will return FALSE
if something that is not an integer is passed in, including something like a
string that R likes to coerce into a NA which when compared with something
else resulting in NA instead of TRUE or FALSE, ex.:
`x <- 'a'; x == as.integer(x)`
}
\examples{
\dontrun{
is_integer(1.2) # returns FALSE
is_integer(1.) # returns TRUE
is_integer(-1) # returns TRUE
is_integer("a") # returns FALSE
is_integer(5) # returns TRUE
is_integer(c(1, 1.2, 3)) # returns TRUE FALSE TRUE
is_integer(c("a", 1.2, 3)) # returns FALSE
}
}
\seealso{
\code{\link{setmklthreads}}
\code{\link{setompthreads}}

Other MBG setup functions: \code{\link{is_lbd_core_repo}},
  \code{\link{is_lbd_singularity}},
  \code{\link{is_rstudio}}, \code{\link{is_singularity}},
  \code{\link{load_R_packages}},
  \code{\link{load_mbg_functions}},
  \code{\link{load_setthreads}}, \code{\link{mbg_setup}},
  \code{\link{set_original_threads}},
  \code{\link{set_serial_threads}},
  \code{\link{source_functions}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
