% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_lbd_singularity.R
\name{is_lbd_singularity}
\alias{is_lbd_singularity}
\title{Tests if within LBD specific Singularity image}
\usage{
is_lbd_singularity()
}
\value{
TRUE/FALSE
}
\description{
\code{is_lbd_singularity} convenience function to check if a Singularity
container was spun up specifically from a LBD image.
}
\details{
This could be used in multiple places and only wanted to change the path
to where the Singularity images live (if it ever changes) in one place and
not many.
}
\seealso{
This function is used by:
\code{\link{load_R_packages}}

Other MBG setup functions: \code{\link{is_integer}},
  \code{\link{is_lbd_core_repo}}, \code{\link{is_rstudio}},
  \code{\link{is_singularity}},
  \code{\link{load_R_packages}},
  \code{\link{load_mbg_functions}},
  \code{\link{load_setthreads}}, \code{\link{mbg_setup}},
  \code{\link{set_original_threads}},
  \code{\link{set_serial_threads}},
  \code{\link{source_functions}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
