% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_new_cluster.R
\name{is_new_cluster}
\alias{is_new_cluster}
\title{Is this the new cluster?}
\usage{
is_new_cluster()
}
\value{
logical indicating if program is on the new cluster.
}
\description{
Returns logical indicating if program is running on the new cluster.
}
