% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_oos_preds.R
\name{is_oos_preds}
\alias{is_oos_preds}
\title{FUNCTION_TITLE}
\usage{
is_oos_preds(rd = run_date, all.data = df,
  cell_draws_filename = "\%s_cell_draws_eb_bin\%i_\%s_\%i.RData",
  holdouts = 5, reg, years = 2000:2015, indic = indicator,
  indic_group = indicator_group, holdoutlist = NULL, fun = "mean",
  shapefile_version = "current", ...)
}
\arguments{
\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{all.data}{PARAM_DESCRIPTION, Default: df}

\item{cell_draws_filename}{PARAM_DESCRIPTION, Default: '\%s_cell_draws_eb_bin\%i_\%s_\%i.RData'}

\item{holdouts}{PARAM_DESCRIPTION, Default: 5}

\item{reg}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION, Default: \code{2000:2015}}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}

\item{indic_group}{PARAM_DESCRIPTION, Default: indicator_group}

\item{holdoutlist}{PARAM_DESCRIPTION, Default: NULL}

\item{fun}{PARAM_DESCRIPTION, Default: 'mean'}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
