% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_oos_preds_testing.R
\name{is_oos_preds_testing}
\alias{is_oos_preds_testing}
\title{FUNCTION_TITLE}
\usage{
is_oos_preds_testing(rd = run_date, all.data = df,
  cell_draws_filename = "\%s_cell_draws_eb_bin\%i_\%s_\%i.RData",
  holdouts = 5, reg, years = 2000:2015, indic = indicator,
  indic_group = indicator_group, holdoutlist = NULL)
}
\arguments{
\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{all.data}{PARAM_DESCRIPTION, Default: df}

\item{cell_draws_filename}{PARAM_DESCRIPTION, Default: '\%s_cell_draws_eb_bin\%i_\%s_\%i.RData'}

\item{holdouts}{PARAM_DESCRIPTION, Default: 5}

\item{reg}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION, Default: 2000:2015}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}

\item{indic_group}{PARAM_DESCRIPTION, Default: indicator_group}

\item{holdoutlist}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
