% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_singularity.R
\name{is_singularity}
\alias{is_singularity}
\title{Tests if within a Singularity container}
\usage{
is_singularity(image_dir = "")
}
\arguments{
\item{image_dir}{A path to a directory containing Singularity images
[default = ""]. Must be a valid path.}
}
\value{
TRUE/FALSE
}
\description{
\code{is_singularity} returns TRUE if within a Singularity container, FALSE
otherwise.
}
\details{
This function can determine if it is being run in a Singularity container.
If the 'image_dir' argument is left blank, it will return TRUE if it is
within any Singularity container. If a path is provided for 'image_dir', this
function will test if the container was spun up from an image that lives in
the specified directory. This will help us determine if the container was
spun up from a specific team's image, which can be important to distinguish
where load packages from or if one should attempt to fit a model with a
special package like TMB.
}
\examples{
\dontrun{
# return TRUE if in any Singularity container / FALSE otherwise
is_singularity()
# TRUE only if container is spun up from image in "/share/singularity-images/lbd"
is_singularity(image_dir = "/share/singularity-images/lbd")
# Will \\code{stop()} because an invalid path was supplied
is_singularity(image_dir = "/path/to/nowhere")
}
}
\seealso{
This function is used by:
  \code{\link{is_lbd_singularity}}

Other MBG setup functions: \code{\link{is_integer}},
  \code{\link{is_lbd_core_repo}},
  \code{\link{is_lbd_singularity}},
  \code{\link{is_rstudio}}, \code{\link{load_R_packages}},
  \code{\link{load_mbg_functions}},
  \code{\link{load_setthreads}}, \code{\link{mbg_setup}},
  \code{\link{set_original_threads}},
  \code{\link{set_serial_threads}},
  \code{\link{source_functions}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
