% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_worldpop_release.R
\name{is_valid_worldpop_release}
\alias{is_valid_worldpop_release}
\title{Tests if worldpop release is valid.}
\usage{
is_valid_worldpop_release(ignore_me_or_get_error)
}
\arguments{
\item{ignore_me_or_get_error}{an unused parameter}
}
\description{
Tests if worldpop release is valid.
}
\details{
ignore_me_or_get_error parameter provided by ez_evparse() from set_up_config. DO NOT
 try to use this parameter in any way or R will crash with an "unxpected input" error because the string cannot
 be evaluated into anything (unlike e.g., "142" which can be evaluated into the numeric 142)
}
