% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janky_covariate_fix.R
\name{janky_covariate_fix}
\alias{janky_covariate_fix}
\title{FUNCTION_TITLE}
\usage{
janky_covariate_fix(df, var)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
## 4749, 44533 fix to create if missing
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
