% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_year_map.R
\name{last_year_map}
\alias{last_year_map}
\title{Labeled map}
\usage{
last_year_map(shape, centroids, admin = 0, val_range = c(0, 1),
  distiller_direction = 1, year = 2015)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame specified to the region/country/admin1 level that your model was run over}

\item{centroids}{Data table that species the x and y coordinates of the centroids for the shape of the map
and has the attached labels to display using ggrepel.}

\item{admin}{The specified admin level called in subnational_ts_function}

\item{val_range}{The specified value range for mean value of your indicator}

\item{distiller_direction}{Either 1 or -1, if high is bad, distiller direction is -1 with the red tones at high values of indicator}

\item{year}{This specifies the last year of your time series, which is used for creating a legend for the function}
}
\value{
This function returns a ggplot object
}
\description{
Create a labeled map of the admin levels included in the region/country/admin1 level. The labels included are the
specified admin names unless over 25 geographies are included, in which case they are labeled by ihme loc id. This is done so larger regions
like Africa do not print out the entire country name, as it would take up too much space.

This map comes in two flavors:
1) last_year_map: This map colors the region/country/admin1 level by your indicator with the scale specified by val_range (default 0-1)
2) last_year_map_multiple : When multiple model indicators/runs are included, the map is no longer colored by indicator and instead
all grey.

This function is called in the subnational_ts_function and is pretty specific to this function. However, it could also serve as a
template for generalizing a labeled map function that also colors by a specific indicator.
}
\examples{

\dontrun{
## If ad0_shape_reg_df is a fortified data frame over a specific region, and centroids are the centroids for the defined
## countries in that region labeled by country name.
gg_lastyear_map <- last_year_map(ad0_shape_reg_df, centroids, admin = 0)
}
}
\author{
Michael Cork, \email{mcork23@uw.edu}
}
