% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_year_map_multiple.R
\name{last_year_map_multiple}
\alias{last_year_map_multiple}
\title{FUNCTION_TITLE}
\usage{
last_year_map_multiple(shape, centroids, admin = 0, year = 2015)
}
\arguments{
\item{shape}{PARAM_DESCRIPTION}

\item{centroids}{PARAM_DESCRIPTION}

\item{admin}{PARAM_DESCRIPTION, Default: 0}

\item{year}{PARAM_DESCRIPTION, Default: 2015}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
