% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_quadtree.R
\name{lines.quadtree}
\alias{lines.quadtree}
\title{FUNCTION_TITLE}
\usage{
lines.quadtree(q, xylim, ...)
}
\arguments{
\item{q}{PARAM_DESCRIPTION}

\item{xylim}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
