% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2paths.R
\name{list2paths}
\alias{list2paths}
\title{FUNCTION_TITLE}
\usage{
list2paths(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
