% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_admin_raster.R
\name{load_admin_raster}
\alias{load_admin_raster}
\title{FUNCTION_TITLE}
\usage{
load_admin_raster(admin_level, simple_raster, disag_fact = NULL,
  shapefile_version = "current")
}
\arguments{
\item{admin_level}{PARAM_DESCRIPTION}

\item{simple_raster}{PARAM_DESCRIPTION}

\item{disag_fact}{PARAM_DESCRIPTION, Default: NULL}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
