% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_crop_covariates.R
\name{load_and_crop_covariates}
\alias{load_and_crop_covariates}
\title{FUNCTION_TITLE}
\usage{
load_and_crop_covariates(fixed_effects, simple_polygon, agebin = 1)
}
\arguments{
\item{fixed_effects}{PARAM_DESCRIPTION}

\item{simple_polygon}{PARAM_DESCRIPTION}

\item{agebin}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
