% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_crop_covariates_annual.R
\name{load_and_crop_covariates_annual}
\alias{load_and_crop_covariates_annual}
\title{FUNCTION_TITLE}
\usage{
load_and_crop_covariates_annual(covs, measures, simple_polygon,
  start_year = 1998, end_year = 2017, interval_mo = 60, agebin = 1)
}
\arguments{
\item{covs}{PARAM_DESCRIPTION}

\item{measures}{PARAM_DESCRIPTION}

\item{simple_polygon}{PARAM_DESCRIPTION}

\item{start_year}{PARAM_DESCRIPTION, Default: 1998}

\item{end_year}{PARAM_DESCRIPTION, Default: 2017}

\item{interval_mo}{PARAM_DESCRIPTION, Default: 60}

\item{agebin}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{mask}}
}
