% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_preds.R
\name{load_cell_preds}
\alias{load_cell_preds}
\title{FUNCTION_TITLE}
\usage{
load_cell_preds(indicator_group, indicator, rd = run_date, region,
  agebin, u5m = FALSE, other = "", ageasindic = TRUE)
}
\arguments{
\item{indicator_group}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{region}{PARAM_DESCRIPTION}

\item{agebin}{PARAM_DESCRIPTION}

\item{u5m}{PARAM_DESCRIPTION, Default: FALSE}

\item{other}{PARAM_DESCRIPTION, Default: ''}

\item{ageasindic}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
