% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_preds_stack.R
\name{load_cell_preds_stack}
\alias{load_cell_preds_stack}
\title{FUNCTION_TITLE}
\usage{
load_cell_preds_stack(indicator_group, indicator, rd = run_date, region,
  agebin)
}
\arguments{
\item{indicator_group}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{region}{PARAM_DESCRIPTION}

\item{agebin}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
