% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_custom_admin_raster.R
\name{load_custom_admin_raster}
\alias{load_custom_admin_raster}
\title{Load Custom Admin Raster}
\usage{
load_custom_admin_raster(shapefile_path, field, simple_raster,
  shapefile = NULL)
}
\arguments{
\item{shapefile_path}{character string - Full file path to shapefile to be used to make admin raster.}

\item{field}{character string - Field in shapefile to be used as value in raster.}

\item{simple_raster}{Simple raster to set extent and resolution of admin raster.}

\item{shapefile}{option to pass in shapefile rather than path to save some time. Default `NULL`}
}
\value{
Raster with resolution and extent the same as simple_raster, with shapefile field as value
}
\description{
takes in a shapefile path or an spdf object and converts it into a admin raster using the field variable
}
\examples{
\dontrun{
custom_admin_raster <- load_custom_admin_raster(shapefile_path, field, simple_raster)
}

}
