% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_from_parallelize.R
\name{load_from_parallelize}
\alias{load_from_parallelize}
\title{Load from parallelize()}
\usage{
load_from_parallelize(fname = as.character(commandArgs()[4]),
  rownumber = as.numeric(commandArgs()[5]))
}
\arguments{
\item{fname}{filename of the temp file created by \code{parallelize()}}

\item{rownumber}{which row of the `lv` object should this particular
instance of the child script load from}
}
\value{
nothing; assigns objects to child script global environment.
}
\description{
This function takes the two things passed in a qsub created by
\code{parallelize()} - the temp file name and which row in loopvars the current iteration
of the child script should load from - and loads the appropriate \code{save_objs} and
\code{expand_vars} from \code{parallelize()} into the environment of the child script.
}
\note{
This is meant to be run from the \code{child script}; by default
both \code{fname} and \code{rownumber} should be loaded appropriately from
\code{commandArgs()}
}
\examples{
# Note: this is within the CHILD SCRIPT, not the master script

# A good place to put this is right after you've sourced all the
# mbg_central function scripts.  Then simply run the
# function to set up your environment:
\dontrun{
load_from_parallelize()
}

}
