% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gbd_covariates.R
\name{load_gbd_covariates}
\alias{load_gbd_covariates}
\title{Load GBD Covariates}
\usage{
load_gbd_covariates(covs, measures, year_ids, age_ids, template,
  use_subnationals = F, simple_polygon, interval_mo, shapefile_version)
}
\arguments{
\item{covs}{A character vector listing GBD covariates/outputs to extract. For covariates, this
should be indicated by covariate_name_short, while for outputs, this should be indicated by
acause. Usually fulfilled by gbd_fixed_effects.}

\item{measures}{A character vector coresponding to 'covs' listing the type of GBD quantity for
each item. Options are 'covariate' and 'outcome'. Usually fulfilled by gbd_fixed_effects_measures.}

\item{year_ids}{A numeric vector of year_ids. Usually fulfilled by year_list.}

\item{age_ids}{A string of age_ids. Usually fulfilled by gbd_fixed_effects_age.}

\item{template}{A raster layer of the buffered modelling area. usually it will be cov_layers[[1]][[1]].
If NULL, a default template is loaded using load_and_crop_covariates_annual()}

\item{use_subnationals}{Logical. If true, the function will replace admin0 with a subnational units
where possible. Use with caution because it's not been tested outside of Africa. It might not
work for countries with multiple levels of subnational units (e.g. UK or India).}

\item{simple_polygon}{simple_polygon object used for the modeling region. made in load_simple_polygon}

\item{interval_mo}{number of months in a time unit. usually 12 to correspond 'annual' year_loadxs}
}
\value{
A list of covariates
}
\description{
A faster version of load_gbd_covariates. I've left the old one for backwards capability (dccasey 8/23/2017)
}
