% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gbd_data.R
\name{load_gbd_data}
\alias{load_gbd_data}
\title{FUNCTION_TITLE}
\usage{
load_gbd_data(gbd_type, gbd_name, gaul_list, measure_id = 6,
  age_group_id = 1, metric_id = 3, year_ids = c(2000:2017),
  return_by_age_sex = "no", collapse_age_sex = FALSE,
  shapefile_version = "current", gbd_round_id = 5,
  named_location_field = "GAUL_CODE", ...)
}
\arguments{
\item{gbd_type}{PARAM_DESCRIPTION}

\item{gbd_name}{PARAM_DESCRIPTION}

\item{gaul_list}{PARAM_DESCRIPTION}

\item{measure_id}{PARAM_DESCRIPTION, Default: 6}

\item{age_group_id}{PARAM_DESCRIPTION, Default: 1}

\item{metric_id}{PARAM_DESCRIPTION, Default: 3}

\item{year_ids}{PARAM_DESCRIPTION, Default: c(2000:2017)}

\item{return_by_age_sex}{PARAM_DESCRIPTION, Default: 'no'}

\item{collapse_age_sex}{PARAM_DESCRIPTION, Default: FALSE}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}

\item{gbd_round_id}{PARAM_DESCRIPTION, Default: 5}

\item{named_location_field}{PARAM_DESCRIPTION, Default: 'GAUL_CODE'}

\item{...}{any other fields to be passed to all of the get_* functions from GBD CC}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
