% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{load_input_data}
\alias{load_input_data}
\title{FUNCTION_TITLE}
\usage{
load_input_data(indicator, simple = NULL, agebin = 0,
  removeyemen = FALSE, pathaddin = "", withdate = FALSE, date = "",
  years = "five_year", range = 5, update_run_date = FALSE,
  withtag = FALSE, datatag = "", use_share = FALSE, yl = year_list)
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{simple}{PARAM_DESCRIPTION, Default: NULL}

\item{agebin}{PARAM_DESCRIPTION, Default: 0}

\item{removeyemen}{PARAM_DESCRIPTION, Default: FALSE}

\item{pathaddin}{PARAM_DESCRIPTION, Default: ''}

\item{withdate}{PARAM_DESCRIPTION, Default: FALSE}

\item{date}{PARAM_DESCRIPTION, Default: ''}

\item{years}{PARAM_DESCRIPTION, Default: 'five_year'}

\item{range}{PARAM_DESCRIPTION, Default: 5}

\item{update_run_date}{PARAM_DESCRIPTION, Default: FALSE}

\item{withtag}{PARAM_DESCRIPTION, Default: FALSE}

\item{datatag}{PARAM_DESCRIPTION, Default: ''}

\item{use_share}{PARAM_DESCRIPTION, Default: FALSE}

\item{yl}{PARAM_DESCRIPTION, Default: year_list}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_match}}
}
