% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{load_libs}
\alias{load_libs}
\title{Load libraries}
\usage{
load_libs(packages, stop = F)
}
\arguments{
\item{packages}{A vector or list of packages}

\item{stop}{Boolean; if T will stop code if not all packages load. if F (default),
the code will contiue and will throw a warning.}
}
\value{
None
}
\description{
A function to load libraries based on whether the
code is being run on cluster prod or the geos nodes. Tries
to load each package, and returns errors or warnings if the package fails to
load rather than breaking the loop.
}
\note{
TO BE DEPRECATED BECAUSE LBDCore HAS PACKAGE LIST in dependencies
}
\author{
Rebecca Stubbs
}
