% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mbg_covariates.R
\name{load_mbg_covariates}
\alias{load_mbg_covariates}
\title{FUNCTION_TITLE}
\usage{
load_mbg_covariates(mbg_fixed_effects, simple_polygon)
}
\arguments{
\item{mbg_fixed_effects}{PARAM_DESCRIPTION}

\item{simple_polygon}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
