% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mbg_functions.R
\name{load_mbg_functions}
\alias{load_mbg_functions}
\title{Loads all of the other MBG functions from a specified repo (directory)}
\usage{
load_mbg_functions(repo)
}
\arguments{
\item{repo}{Path to repository directory to attempt to load functions from}
}
\value{
None
}
\description{
\code{load_mbg_functions} finds all of the *.R files with 'functions' in the
naming convention and sources them with a message.
}
\details{
Makes sure that the provided repo directory exists and if so, attempts to
load scripts of R functions from within that directory. Note that this
function will *only* load R scripts with 'functions' in the naming convention
of the R script file name it finds after searching recursively from the
directory provided for 'repo'.
}
\note{
There is an exception added to not pull from the 'testing' subdirectory
}
\seealso{
This function is used by:
\code{\link{mbg_setup}}

Other MBG setup functions: \code{\link{is_integer}},
  \code{\link{is_lbd_core_repo}},
  \code{\link{is_lbd_singularity}},
  \code{\link{is_rstudio}}, \code{\link{is_singularity}},
  \code{\link{load_R_packages}},
  \code{\link{load_setthreads}}, \code{\link{mbg_setup}},
  \code{\link{set_original_threads}},
  \code{\link{set_serial_threads}},
  \code{\link{source_functions}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
