% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pipeline_image.R
\name{load_pipeline_image}
\alias{load_pipeline_image}
\title{Load Pipeline Pre-saved Image}
\usage{
load_pipeline_image(indicator_group, indicator, run_date, pathaddin,
  file_format = ".RData")
}
\arguments{
\item{indicator_group}{Indicator Group}

\item{indicator}{Indicator Name}

\item{run_date}{A pre-saved run-date}

\item{pathaddin}{Additional thing to add to path, quite often: \code{paste0("_bin", age, "_", reg, "_", holdout)}}
}
\value{
The pre-saved environment from launch script
}
\description{
Load Pipeline Pre-saved image into the global environment
}
