% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_populations_cov.R
\name{load_populations_cov}
\alias{load_populations_cov}
\title{Load Populations For Covariates}
\usage{
load_populations_cov(reg, pop_measure, measure = "prevalence",
  simple_polygon, simple_raster, year_list, interval_mo, outputdir,
  pixel_id)
}
\arguments{
\item{reg}{modeling region over which you are operating}

\item{pop_measure}{the world pop covariate measure used in your model}

\item{measure}{needs to be depricated, currently a flag that if set to 'prevalence' this function also grabs teh covatiate stakers}

\item{simple_polygon}{the simple polygon for your region}

\item{simple_raster}{the simple raster for your model}

\item{year_list}{The years you are modeling over, will be used to pull populations}

\item{interval_mo}{the number of months between time steps}

\item{outputdir}{currently not used, there is a commented out line of code which can save the population raster if you want}
}
\value{
used internally above, returns a data frame with the population in each cell-year so rates can be turned into counts

for fractional raking and aggregation
}
\description{
This gets the population in each cell which you need to convert the rates to counts for fractional raking
it can also get the stackers if you would like
}
