% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_region_input_data.R
\name{load_region_input_data}
\alias{load_region_input_data}
\title{FUNCTION_TITLE}
\usage{
load_region_input_data(indicator_group, indicator, run_date, reg)
}
\arguments{
\item{indicator_group}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{reg}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
