% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_simple_polygon.R
\name{load_simple_polygon}
\alias{load_simple_polygon}
\title{FUNCTION_TITLE}
\usage{
load_simple_polygon(gaul_list, buffer, tolerance = 0.2,
  subset_only = F, makeplots = F, use_premade = F,
  custom_shapefile_path = NULL, custom_shapefile = NULL, raking = F,
  shapefile_version = "current")
}
\arguments{
\item{gaul_list}{PARAM_DESCRIPTION}

\item{buffer}{PARAM_DESCRIPTION}

\item{tolerance}{PARAM_DESCRIPTION, Default: 0.2}

\item{subset_only}{PARAM_DESCRIPTION, Default: F}

\item{makeplots}{PARAM_DESCRIPTION, Default: F}

\item{use_premade}{PARAM_DESCRIPTION, Default: F}

\item{custom_shapefile_path}{PARAM_DESCRIPTION, Default: NULL}

\item{custom_shapefile}{PARAM_DESCRIPTION, Default: NULL}

\item{raking}{PARAM_DESCRIPTION, Default: F}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[rgeos]{gIntersection}}
 \code{\link[raster]{area}}
}
