% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_worldpop_covariate.R
\name{load_worldpop_covariate}
\alias{load_worldpop_covariate}
\title{Load worldpop covariate raster and return}
\usage{
load_worldpop_covariate(template_raster, covariate = "worldpop",
  pop_measure, pop_release, start_year = min(year_list),
  end_year = max(year_list), interval = interval_mo)
}
\arguments{
\item{template_raster}{the raster template which all returned data will match.}

\item{covariate}{the covariate to load. Defaults to "worldpop"}

\item{pop_measure}{the covariate measure to load.}

\item{pop_release}{the covariate measure release to use.}

\item{start_year}{the first year to locate data for. Defaults the minimum value in the configuration value \code{year_list}}

\item{end_year}{the last year to locate data for. defaults the maximum value in the configuration value \code{year_list}}

\item{interval}{the number of months between data readings. Defaults to the global interval_mo}
}
\value{
list with named value containing your covariate data as a raster.
}
\description{
Loads a covariate raster (worldpop by default, can be overridden) and returns.
}
\examples{
\dontrun{
worldpop <- load_worldpop_covariate(simple_polygon, measure = 'a0004t', release = '2019_06_10')

raked_worldpop <- load_worldpop_covariate(simple_polygon,
                                          covariate = 'worldpop_raked',
                                          measure = 'a0004t',
                                          release = '2019_06_10',
                                          start_year = 2000,
                                          end_year = 2017,
                                          interval = 12)
}
}
