% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.inla.spde2.matern.new.R
\name{local.inla.spde2.matern.new}
\alias{local.inla.spde2.matern.new}
\title{SPDE priors function}
\usage{
local.inla.spde2.matern.new(mesh, alpha = 2, prior.pc.rho, prior.pc.sig)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 2}

\item{prior.pc.rho}{PARAM_DESCRIPTION}

\item{prior.pc.sig}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[INLA]{character(0)}}
}
