% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_map_draw.R
\name{location_map_draw}
\alias{location_map_draw}
\title{Create a locator map that highlights the admin level in question in red within a map background}
\usage{
location_map_draw(admin_shape, surround_shape)
}
\arguments{
\item{admin_shape}{SpatialPolygonsDataFrame specified to the specific admin unit, colored in red}

\item{surround_shape}{SpatialPolygonsDataFrame of surrounding geography to place admin unit.
If highlighted admin_shape is a country, then the surround_shape is typically the region it falls in}
}
\value{
This function returns a ggplot object
}
\description{
This function is called in the subnational_ts_function and is a way to generalize locator maps
that are useful to contextualize where the admin unit lies within a reigon or country.
}
\examples{
\dontrun{
## ad0_shape_simple is the simplified ad0 shapefile for Africa, and ad0_reg_codes defines the
## admin 0 codes for a specified region in Africa
admin_shape <- subset(ad0_shape_simple, ADM0_CODE \%in\% ad0_reg_codes)
surround_shape <- ad0_shape_simple
location_map <- location_map_draw(admin_shape, surround_shape)
}
}
\author{
Michael Cork, \email{mcork23@uw.edu}
}
