% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_extrap_gbd_input_by_loc.R
\name{loess_extrap_gbd_input_by_loc}
\alias{loess_extrap_gbd_input_by_loc}
\title{Extrapolate GBD measure to future}
\usage{
loess_extrap_gbd_input_by_loc(gbd, year_list, year_forecast_end = 2030)
}
\arguments{
\item{gbd}{The data.table with name, year, and mean values (output from \code{get_gbd_estimates})}

\item{year_list}{List of in sample years (config argument)}

\item{year_forecast_end}{Year to forecast to. Default: 2030}
}
\value{
A data.table with name, year and mean columns extrapolated to \code{year_forecast_end}
}
\description{
Use a simple loess model to extrapolate GBD input data to desired year
}
