% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log.R
\name{Log}
\alias{Log}
\title{FUNCTION_TITLE}
\usage{
Log(text, ...)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
