% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_rake.R
\name{logit_rake}
\alias{logit_rake}
\title{FUNCTION_TITLE}
\usage{
logit_rake(country_year, rake_dt, gbd)
}
\arguments{
\item{country_year}{PARAM_DESCRIPTION}

\item{rake_dt}{PARAM_DESCRIPTION}

\item{gbd}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
