% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitFindK.R
\name{LogitFindK}
\alias{LogitFindK}
\title{FUNCTION_TITLE}
\usage{
LogitFindK(gbdval, pixelval, weightval, MaxIter = 40, MaxJump = 10,
  FunTol = 1e-05, approx_0_1)
}
\arguments{
\item{gbdval}{PARAM_DESCRIPTION}

\item{pixelval}{PARAM_DESCRIPTION}

\item{weightval}{PARAM_DESCRIPTION}

\item{MaxIter}{PARAM_DESCRIPTION, Default: 40}

\item{MaxJump}{PARAM_DESCRIPTION, Default: 10}

\item{FunTol}{PARAM_DESCRIPTION, Default: 1e-05}

\item{approx_0_1}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
