% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat3D.R
\name{lonlat3D}
\alias{lonlat3D}
\title{Converts from long-lat coords to corresponding 3D
  coords, (x, y, z), on a sphere with radius 1}
\usage{
lonlat3D(lon, lat)
}
\arguments{
\item{lon}{numeric vector of longitude coords}

\item{lat}{numeric vector of latitude coords}
}
\value{
3 column numeric matrix where each row is a (x,y,z) of the
  transformed (long, lat) coords
}
\description{
This function takes in a vector of longitude and a vector of
latitude and returns coordinates on the S2 sphere (globe living in
3D) in (x, y, z) coords
}
