% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstype.R
\name{lstype}
\alias{lstype}
\title{FUNCTION_TITLE}
\usage{
lstype(type = "closure")
}
\arguments{
\item{type}{PARAM_DESCRIPTION, Default: 'closure'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
