% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2000_2015_aroc.R
\name{make.2000.2015.aroc}
\alias{make.2000.2015.aroc}
\title{FUNCTION_TITLE}
\usage{
make.2000.2015.aroc(pred.rast)
}
\arguments{
\item{pred.rast}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
