% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast.R
\name{make.forecast}
\alias{make.forecast}
\title{FUNCTION_TITLE}
\usage{
make.forecast(start.rast, aroc.rast, n.yrs = 15, uselogit = FALSE)
}
\arguments{
\item{start.rast}{PARAM_DESCRIPTION}

\item{aroc.rast}{PARAM_DESCRIPTION}

\item{n.yrs}{PARAM_DESCRIPTION, Default: 15}

\item{uselogit}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
