% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_admin_pred_summary.R
\name{make_admin_pred_summary}
\alias{make_admin_pred_summary}
\title{Make admin prediction summary}
\usage{
make_admin_pred_summary(admin_pred, sp_hierarchy_list,
  summary_stats = "mean", ...)
}
\arguments{
\item{admin_pred}{the admin draws object (e.g. admin_1, etc)}

\item{sp_hierarchy_list}{spatial hierarchy object from the admin draws RData files}

\item{...}{any other arguments to pass to the `summary_stats` functions
(note that currently will be passed to all functions)}

\item{summary_stat}{summary statistic to apply}
}
\value{
data table of summary stats and admin info
}
\description{
Take an admin pred object and an sp_hierarchy list and generate a sorted,
cleaned table for a given set of summary statistics
}
\examples{
\dontrun{
# load("indicator_raked_admin_draws_eb_bin0.RData") # Replace with your admin draws
make_admin_pred_summary(
  admin_2,
  sp_hierarchy_list,
  c("mean", "cirange", "upper", "lower")
)
}
}
