% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_africa_raster.R
\name{make_africa_raster}
\alias{make_africa_raster}
\title{FUNCTION_TITLE}
\usage{
make_africa_raster(gaul)
}
\arguments{
\item{gaul}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
