% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aroc_weights.R
\name{make_aroc_weights}
\alias{make_aroc_weights}
\title{Make AROC weights}
\usage{
make_aroc_weights(weighting_res, weighting_type, year_list, pow = 1,
  input_data, mult_emp_exp)
}
\arguments{
\item{weighting_res}{Spatial resolution to apply weights. Can be done by either:
'domain' (e.g. Africa),
'country'.
In the future this could be expanded to adm* levels. Currently this only affects
weighting if you also select weighting_type==empirical}

\item{weighting_type}{Method to generate the weighting used in the weighted-mean of the
rate of changes calculated between all adjacent pairs of years in year_list.
Can be: 'exponential' (see pow argument for details) or 'empirical'.
Empirical generates weights proportional to the amount of data in each year.
  - The total sample size across all years divided by num_years is added to every year
    to ensure no pairs of years get zero weight
  - If empirical and country are both selected and there are countries with no data then
    exponential weights are used}

\item{year_list}{Vector (integer) of years included in the model run / cell pred object}

\item{pow}{Power used in year-weighting scheme:
exponential weight used in determining year_wt_i:
     (yr_i - yr_1)/(yr_n - yr_i))^pow
if pow==0, get uniform wts
if pow==1, get linear wts
if pow > 1, get exponential wts, etc.}

\item{input_data}{Required to derive emprical weights. A set of cleaned and collapsed micro
data (e.g. input to MBG models) that will be used to derive the emprical weights.
Required Columns:
   - country (iso3)
   - N
   - year}

\item{mult_emp_exp}{if TRUE and you are calculating emprical weights, the sample-size
driven empirical weights are also multipled by the exponential weighting scheme such
that both the exponential weight and the empirical weight contribute to the final weights}
}
\value{
writes cell- and admin- level aroc objects to standard directories and formats
  in the 'pred_derivatives' folder of the model run.  cell-level objects are in the
  cell_pred indexed format. Both cell- and admin- aroc objects are matrices wide by draw.
}
\description{
Generates weights to use in the weighted mean calculation of rate of changes
between adjacent model output years.
}
\examples{
\dontrun{
make_aroc_weights(
  weighting_res = "country",
  weighting_type = "empirical",
  pow = 1,
  input_data = data,
  mult_emp_exp = TRUE
)
}
}
