% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_beta_plot.R
\name{make_beta_plot}
\alias{make_beta_plot}
\title{FUNCTION_TITLE}
\usage{
make_beta_plot(indicator_group, indicator, run_date, region)
}
\arguments{
\item{indicator_group}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
