% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cell_pred_summary.R
\name{make_cell_pred_summary}
\alias{make_cell_pred_summary}
\title{FUNCTION_TITLE}
\usage{
make_cell_pred_summary(draw_level_cell_pred, mask = simple_raster,
  return_as_raster = TRUE, summary_stat = "mean", ...)
}
\arguments{
\item{draw_level_cell_pred}{PARAM_DESCRIPTION}

\item{mask}{PARAM_DESCRIPTION, Default: simple_raster}

\item{return_as_raster}{PARAM_DESCRIPTION, Default: TRUE}

\item{summary_stat}{PARAM_DESCRIPTION, Default: 'mean'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
