% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_condSim.R
\name{make_condSim}
\alias{make_condSim}
\title{FUNCTION_TITLE}
\usage{
make_condSim(pop_wts_object, cell_pred, admin_level, years = c(2000,
  2005, 2010, 2015), gaul_list = gaul_list, fun = "mean",
  summarize = FALSE, shapefile_version = "current")
}
\arguments{
\item{pop_wts_object}{PARAM_DESCRIPTION}

\item{cell_pred}{PARAM_DESCRIPTION}

\item{admin_level}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION, Default: c(2000, 2005, 2010, 2015)}

\item{gaul_list}{PARAM_DESCRIPTION, Default: gaul_list}

\item{fun}{PARAM_DESCRIPTION, Default: 'mean'}

\item{summarize}{PARAM_DESCRIPTION, Default: FALSE}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
