% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_country_list.R
\name{make_country_list}
\alias{make_country_list}
\title{FUNCTION_TITLE}
\usage{
make_country_list(gaul, reg_simple_raster)
}
\arguments{
\item{gaul}{PARAM_DESCRIPTION}

\item{reg_simple_raster}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
