% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_country_list_allAdmins.R
\name{make_country_list_allAdmins}
\alias{make_country_list_allAdmins}
\title{FUNCTION_TITLE}
\usage{
make_country_list_allAdmins(gaul)
}
\arguments{
\item{gaul}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
