% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folds.R
\name{make_folds}
\alias{make_folds}
\title{FUNCTION_TITLE}
\usage{
make_folds(data, n_folds, spat_strat = NULL, temp_strat = NULL,
  spte_strat = NULL, save.file = paste0("/share/geospatial/mbg/",
  indicator_group, "/", indicator, "/output/", run_date, "/stratum.rds"),
  ..., strat_cols, seed)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{n_folds}{PARAM_DESCRIPTION}

\item{spat_strat}{PARAM_DESCRIPTION, Default: NULL}

\item{temp_strat}{PARAM_DESCRIPTION, Default: NULL}

\item{spte_strat}{PARAM_DESCRIPTION, Default: NULL}

\item{save.file}{PARAM_DESCRIPTION, Default: paste0("/share/geospatial/mbg/", indicator_group, "/", indicator,
"/output/", run_date, "/stratum.rds")}

\item{...}{PARAM_DESCRIPTION}

\item{strat_cols}{PARAM_DESCRIPTION}

\item{seed}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
