% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folds_by_poly.R
\name{make_folds_by_poly}
\alias{make_folds_by_poly}
\title{FUNCTION_TITLE}
\usage{
make_folds_by_poly(cts_in_polys, pt_poly_map, n_folds, ...)
}
\arguments{
\item{cts_in_polys}{PARAM_DESCRIPTION}

\item{pt_poly_map}{PARAM_DESCRIPTION}

\item{n_folds}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
