% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gaul_table.R
\name{make_gaul_table}
\alias{make_gaul_table}
\title{FUNCTION_TITLE}
\usage{
make_gaul_table(rr)
}
\arguments{
\item{rr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
