% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gg_map.R
\name{make_gg_map}
\alias{make_gg_map}
\title{FUNCTION_TITLE}
\usage{
make_gg_map(rbrick, title, i)
}
\arguments{
\item{rbrick}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION}

\item{i}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
