% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_period_map.R
\name{make_period_map}
\alias{make_period_map}
\title{FUNCTION_TITLE}
\usage{
make_period_map(modeling_periods)
}
\arguments{
\item{modeling_periods}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
