% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_period_matrix.R
\name{make_period_matrix}
\alias{make_period_matrix}
\title{FUNCTION_TITLE}
\usage{
make_period_matrix(period)
}
\arguments{
\item{period}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
