% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive__make_population_weights.R,
%   R/make_population_weights.R
\name{archive__make_population_weights}
\alias{archive__make_population_weights}
\alias{make_population_weights}
\title{FUNCTION_TITLE}
\usage{
archive__make_population_weights(admin_level, simple_raster, pop_raster,
  gaul_list)

make_population_weights(admin_level, simple_raster, pop_raster, gaul_list,
  custom_admin_raster = NULL)
}
\arguments{
\item{admin_level}{PARAM_DESCRIPTION}

\item{simple_raster}{PARAM_DESCRIPTION}

\item{pop_raster}{PARAM_DESCRIPTION}

\item{gaul_list}{PARAM_DESCRIPTION}

\item{custom_admin_raster}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[seegSDM]{character(0)}}
 \code{\link[raster]{extract}}
}
