% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proj.R
\name{make_proj}
\alias{make_proj}
\title{Make projection}
\usage{
make_proj(ind_gp, ind, rd, type, proj_years = c(2020, 2025),
  measure = "mortality", skip_cols = NULL, raked,
  matrix_pred_name = NULL, year_list = c(2000:2015),
  uselogit = FALSE, extra_file_tag = "",
  shapefile_version = "current")
}
\arguments{
\item{ind_gp}{indicator group}

\item{ind}{indicator}

\item{rd}{run_date}

\item{type}{Type of aroc to create. Options include \code{cell}, \code{admin},
or \code{c('cell', 'admin')} for both.}

\item{proj_years}{Vector (integer) of years that you want to project to.  Note
that this is different from \code{year_list}, which is the list of years that
were included in the model run / are included in the aroc object.}

\item{measure}{prevalence, incidence, mortality, etc}

\item{skip_cols}{columns to skip when reading in the cell preds
For example, if the first two columns store non-pred information in your
file format, \code{skip_cols = 2} will read in all columns from 3 onwards}

\item{raked}{Should we do this with raked cell/admin preds?}

\item{matrix_pred_name}{In \code{sprintf} notation. The one object passed into
the string should will be a region name. this allows different regions to be
passed to different named matrix_preds (pixel level, ad0, ad1, ad2, ...)
e.g. 'had_diarrhea_cell_draws_eb_bin0_\%s_diarrhea2_0.RData' which
will be passed to sprintf('had_diarrhea_cell_draws_eb_bin0_\%s_0.RData', reg)}

\item{year_list}{Vector (integer) of years included in the model run / cell pred object}

\item{uselogit}{Should this be done in logit space?}

\item{extra_file_tag}{Appended at the end of all files generated from this run of the function. Useful if you're comparing different weights/resolutions to calcualte AROC and project}
}
\value{
writes cell- and admin- level projection objects to standard directories and formats
  in the 'pred_derivatives' folder of the model run.  cell-level objects are in the
  cell_pred indexed format. Both cell- and admin- projection objects are matrices wide by draw.
NOTE: admin_preds are sorted such that the order of the rows is by year and then admin_code
}
\description{
Generates a set of draw-level projection objects from aroc objects for
a given set of target years
}
\examples{
\dontrun{
make_proj(
  ind_gp = indicator_group,
  ind = indicator,
  rd = run_date,
  type = c("cell", "admin"),
  proj_years = c(2020, 2025, 2030),
  measure = "prevalence",
  skip_cols = NULL,
  year_list = c(2000:2015),
  uselogit = TRUE
)
}
}
