% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qsub.R
\name{make_qsub}
\alias{make_qsub}
\title{Make qsub}
\usage{
make_qsub(user = Sys.info()["user"], code, cores = slots,
  memory = 100, proj = NULL, ig = indicator_group,
  indic = indicator, reg = "test", age = 0, vallevel = "",
  hard_set_sbh_wgt = TRUE, pct_sbh_wgt = 100, rd = run_date,
  log_location = "sgeoutput", saveimage = FALSE, test = FALSE,
  holdout = 0, use_base_covariates = FALSE,
  test_pv_on_child_models = TRUE, constrain_children_0_inf = FALSE,
  child_cv_folds = 5, fit_stack_cv = TRUE, shell = "r_shell.sh",
  modeltype = "full", corerepo = core_repo, geo_nodes = FALSE,
  use_c2_nodes = FALSE, singularity = singularity_version,
  singularity_opts = NULL)
}
\arguments{
\item{proj}{Can pass in a project name to submit your job under. If default
and the 'geo_nodes' argument is left as its default of 'FALSE', jobs
will be submitted to the prod cluster under the default project
'proj_geospatial'. If default and with 'geos_nodes = TRUE', jobs will be
submitted to the geos (LBD) nodes under the default project
'proj_geo_nodes'. If a project name is passed in for 'proj' the job will
be submitted under that project. Note that this function does not check for
valid project names since these are likely to change often and likely
valid project names are different on each cluster. [default = NULL]}

\item{geo_nodes}{If TRUE, your job will be submitted to the geos (LBD)
cluster, if FALSE, it will be submitted to the prod cluster. Note that if
using the 'proj' argument, make sure to use project name which is valid on
the cluster you are submitting to. [default = FALSE]}

\item{use_c2_nodes}{If TRUE, your job will be submitted to the C2 nodes on
the prod cluster, if FALSE, the C2 nodes are not specified. Note that if
FALSE, your job may land on a node with much less memory or your node may
still land on a C2 node anyway. If both the 'use_c2_nodes' and 'geo_nodes'
arguments are set to TRUE, then the code will issue a warning and default
to the geos nodes. [default = FALSE]}

\item{singularity}{Launch R from a Singularity image. The default is
  the default image. You may also provide a string which can be either a complete
  path to a Singularity image that is not located at the default image
  location, or just the name of the Singularity image that is assumed located
  at the default image location. NULL is also accepted, which will launch R
  using the default R installation on the geos or prod nodes, but this is
  no longer recommended and will likely be deprecated at some point in the
  future.

  If 'default' is chosen, the default image is defined in the shell script
  executed by this R script ('shell_sing.sh') so that no R code need be
  updated when the default image is updated. Different versions of a
  Singularity image or test versions may be specified by providing the name
  or path of the image. Currently, all standard images for LBD are kept at
  the default location of /share/singularity-images/lbd.
  [default = 'default']}

\item{singularity_opts}{pass in a named list of environmental variables.
\code{qsub_sing_envs} will check that the names of the list members passed
in match the environmental variables that the shell_sing.sh script knows
about: 'SET_OMP_THREADS' and/or 'SET_MKL_THREADS'. Passing in other
environmental names in the list will result in an error. If this is left
as 'NULL' and a Singularity image is used, SET_OMP_THREADS and
SET_MKL_THREADS will remain unset and the shell_sing.sh script will use
the default setting of SET_OMP_THREADS=1 and SET_MKL_THREADS={max_threads}
(see shell_sing.sh comments). For example SET_OMP_THREADS=1 and
SET_MKL_THREADS=4 can be achieved by passing in
  \code{envs = list(SET_OMP_THREADS=1, SET_MKL_THREADS=4)}
[default = NULL]}
}
\value{
Returns a qsub string
}
\description{
Constructs a qsub string and returns it
}
