% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qsubs_aggregation.R
\name{make_qsubs_aggregation}
\alias{make_qsubs_aggregation}
\title{FUNCTION_TITLE}
\usage{
make_qsubs_aggregation(qsubs_to_make, stderr_log, stdout_log, project,
  resources, singularity_str, queue = NULL, priority = 0, slots, shell,
  code, indicator, indicator_group, run_date, pop_measure, overwrite,
  corerepo, raking_shapefile_version, modeling_shapefile_version)
}
\arguments{
\item{qsubs_to_make}{PARAM_DESCRIPTION}

\item{stderr_log}{PARAM_DESCRIPTION}

\item{stdout_log}{PARAM_DESCRIPTION}

\item{project}{PARAM_DESCRIPTION}

\item{resources}{PARAM_DESCRIPTION}

\item{singularity_str}{PARAM_DESCRIPTION}

\item{queue}{PARAM_DESCRIPTION. Default: NULL.}

\item{slots}{PARAM_DESCRIPTION}

\item{shell}{PARAM_DESCRIPTION}

\item{code}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{pop_measure}{PARAM_DESCRIPTION}

\item{overwrite}{PARAM_DESCRIPTION}

\item{corerepo}{PARAM_DESCRIPTION}

\item{raking_shapefile_version}{PARAM_DESCRIPTION}

\item{modeling_shapefile_version}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
