% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_quilt_matrix.R
\name{make_quilt_matrix}
\alias{make_quilt_matrix}
\title{FUNCTION_TITLE}
\usage{
make_quilt_matrix(this_year)
}
\arguments{
\item{this_year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fields]{quilt.plot}}
}
