% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regions_map.R
\name{make_regions_map}
\alias{make_regions_map}
\title{FUNCTION_TITLE}
\usage{
make_regions_map(regions, subset_shape, shapefile_version = "current")
}
\arguments{
\item{regions}{PARAM_DESCRIPTION}

\item{subset_shape}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
