% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_square.R
\name{make_square}
\alias{make_square}
\title{FUNCTION_TITLE}
\usage{
make_square(location_set_version_id, year_start, year_end, by_sex = 1,
  by_age = 1, custom_sex_id = NULL, custom_age_group_id = NULL,
  covariates = NULL, population = FALSE)
}
\arguments{
\item{location_set_version_id}{PARAM_DESCRIPTION}

\item{year_start}{PARAM_DESCRIPTION}

\item{year_end}{PARAM_DESCRIPTION}

\item{by_sex}{PARAM_DESCRIPTION, Default: 1}

\item{by_age}{PARAM_DESCRIPTION, Default: 1}

\item{custom_sex_id}{PARAM_DESCRIPTION, Default: NULL}

\item{custom_age_group_id}{PARAM_DESCRIPTION, Default: NULL}

\item{covariates}{PARAM_DESCRIPTION, Default: NULL}

\item{population}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
