% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stack_rasters.R
\name{make_stack_rasters}
\alias{make_stack_rasters}
\title{FUNCTION_TITLE}
\usage{
make_stack_rasters(covariate_layers = all_cov_layers, period = NULL,
  child_models = list(), indicator_family = "binomial",
  rd = run_date, re = reg, ind_gp = indicator_group,
  ind = indicator, ho = holdout, centre_scale_df = NULL, ...)
}
\arguments{
\item{covariate_layers}{PARAM_DESCRIPTION, Default: all_cov_layers}

\item{period}{PARAM_DESCRIPTION, Default: NULL}

\item{child_models}{PARAM_DESCRIPTION, Default: list()}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{re}{PARAM_DESCRIPTION, Default: reg}

\item{ind_gp}{PARAM_DESCRIPTION, Default: indicator_group}

\item{ind}{PARAM_DESCRIPTION, Default: indicator}

\item{ho}{PARAM_DESCRIPTION, Default: holdout}

\item{centre_scale_df}{PARAM_DESCRIPTION, Default: NULL}

\item{stacker_model}{PARAM_DESCRIPTION, Default: stacked_results[[2]]}

\item{return_children}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
