% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_rasters.R
\name{make_summary_rasters}
\alias{make_summary_rasters}
\title{Makes summary rasters from cell pred}
\usage{
make_summary_rasters(cell_pred, simple_raster,
  summary_measures = c("mean", "cirange", "lower", "upper"), raked = T)
}
\arguments{
\item{cell_pred}{cell pred object}

\item{simple_raster}{admin raster matching the cell pred object}

\item{summary_measures}{charater vector - functions to summarize by. see default below}

\item{raked}{boolean default `T` - adds "raked" or "unraked" to name of output}
}
\value{
a named list with a raster brick for each summary measure
}
\description{
wrapper function for make_cell_pred_summary
}
