% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time_stamp.R
\name{make_time_stamp}
\alias{make_time_stamp}
\title{FUNCTION_TITLE}
\usage{
make_time_stamp(time_stamp)
}
\arguments{
\item{time_stamp}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
