% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_weights_vector.R
\name{make_weights_vector}
\alias{make_weights_vector}
\title{FUNCTION_TITLE}
\usage{
make_weights_vector(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
