% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_year_table.R
\name{make_year_table}
\alias{make_year_table}
\title{FUNCTION_TITLE}
\usage{
make_year_table(this_year)
}
\arguments{
\item{this_year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
