% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matAgeRate.R
\name{matAgeRate}
\alias{matAgeRate}
\title{FUNCTION_TITLE}
\usage{
matAgeRate(age_group, groups = c("15-19", "20-24", "25-29", "30-34"))
}
\arguments{
\item{age_group}{PARAM_DESCRIPTION}

\item{groups}{PARAM_DESCRIPTION, Default: c("15-19", "20-24", "25-29", "30-34")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
