% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg_get_nodeenv.R
\name{mbg_get_nodeenv}
\alias{mbg_get_nodeenv}
\title{Get image of node in the MBG DAG}
\usage{
mbg_get_nodeenv(node, ig, indic, rd, reg, age, holdout)
}
\arguments{
\item{node}{The name(s) of the node in the DAG}

\item{age}{the age for the model which uses this data.}

\item{holdout}{numeric holdout value (0 for no holdout).}

\item{indicator_group}{the indicator group e.g., "hiv" or "u5m".}

\item{indicator}{the indicator e.g., "wasting_mod_b" or "male_circumcision".}

\item{run_date}{string run date in YYYY_MM_DD_HH_MM_SS format.}

\item{region}{str region name.}
}
\value{
the path to save the file.
}
\description{
Get model data for each node in DAG
}
