% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg_job_marker.R
\name{mbg_job_marker}
\alias{mbg_job_marker}
\title{Create an empty marker file for tracking start or end of jobs}
\usage{
mbg_job_marker(type, tmpdir = "~/.mbgjob")
}
\arguments{
\item{type}{Either \code{"start"} or \code{"end"}.}

\item{tmpdir}{Temporary folder to store job markers in.
Defaults to a folder in your home directory.}
}
\description{
Create an empty marker file for tracking start or end of jobs
}
\note{
If \code{JOB_ID} is not found in the environment variable, and
the R session is not interactive, then this function will break.
}
