% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg_save_nodeenv.R
\name{mbg_save_nodeenv}
\alias{mbg_save_nodeenv}
\title{Save image of node in the MBG DAG}
\usage{
mbg_save_nodeenv(node, ig, indic, rd, reg, age, holdout, objs = ls(envir
  = .GlobalEnv))
}
\arguments{
\item{node}{The name of the node in the DAG}

\item{age}{the age for the model which uses this data.}

\item{holdout}{numeric holdout value (0 for no holdout).}

\item{objs}{Objects to save out.
Default: everything in environment (\code{ls()})}

\item{indicator_group}{the indicator group e.g., "hiv" or "u5m".}

\item{indicator}{the indicator e.g., "wasting_mod_b" or "male_circumcision".}

\item{run_date}{string run date in YYYY_MM_DD_HH_MM_SS format.}

\item{region}{str region name.}
}
\value{
the path to save the file, but the file itself is saved out.
}
\description{
Save model data for each node in DAG
}
