% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg_setup.R
\name{mbg_setup}
\alias{mbg_setup}
\title{Loads R packages and sources functions for MBG modeling}
\usage{
mbg_setup(package_list, repos)
}
\arguments{
\item{package_list}{A vector of R package names to attempt to load}

\item{repos}{A vector of paths to repository directories. At a
minimum, this should contain the 'core_repo', but may also contain an
indicator repo and one or more additional repos. Functions will be loaded
from each repo in the order they appear in this vector. If any path has
'lbd_core' as the last directory in the path, only it's subdirectory
'mbg_central' will be recursively searched for functions to load as is the
custom. A single, complete path (or object storing a single path) can also
be provided as a string.}
}
\value{
None
}
\description{
\code{mbg_setup()} Loads R packages and sources functions for MBG modeling
with helper functions in 'MBG setup functions' family
}
\details{
This function makes use of some other helper functions in the 'MBG setup
functions' family to load all the necessary packages in a provided vector
as well as functions for the core_repo and optionally additional repos such
as indicator repos and/or development repos. Functions will be read in from
repos in the order they appear in the 'repos' vector argument. The examples
below show how this could be run with one or multiple repos.
}
\examples{
\dontrun{
mbg_setup(package_list = "ggplot2", repos = "/share/code/geospatial/lbd_core")
mbg_setup(package_list = package_list, repos = c(core_repo, indic_repo))
}
}
\seealso{
This function uses:
\code{\link{load_R_packages}}
\code{\link{load_mbg_functions}}
\code{\link{is_lbd_core_repo}}

Other MBG setup functions: \code{\link{is_integer}},
  \code{\link{is_lbd_core_repo}},
  \code{\link{is_lbd_singularity}},
  \code{\link{is_rstudio}}, \code{\link{is_singularity}},
  \code{\link{load_R_packages}},
  \code{\link{load_mbg_functions}},
  \code{\link{load_setthreads}},
  \code{\link{set_original_threads}},
  \code{\link{set_serial_threads}},
  \code{\link{source_functions}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
