% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MbgStandardCovariateLoader.R
\docType{data}
\name{MbgStandardCovariateLoader}
\alias{MbgStandardCovariateLoader}
\title{Covariate loader for standard Model-based Geostatistics.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MbgStandardCovariateLoader
}
\arguments{
\item{start_year}{A numeric indicating the earliest year of data to attempt to retrieve.}

\item{end_year}{A numeric indicating the latest year of data to attempt to retrieve.}

\item{interval}{A numeric number of months that the data is provided in. Annnual data
would have an interval of 12}

\item{covariate_config}{A data.table with covariate loading information. This must contain
columns "covariate", "measure", and "release". All three should be string values with
the "release" being a timestamp in the form of YYYY_MM_DD e.g., "2019_06_13".}
}
\description{
Loads covariate data in bulk and returns suitable raster/brick objects.
}
\details{
Covariates are stored in a 00_MBG_STANDARD directory. See
\code{MbgStandardCovariateLoader$public_fields$cov_dir} or an instances $cov_dir attribute
as .tif files and loaded as raster objects (either as a layer or a brick of layers).
}
\examples{
\dontrun{
config.table <- data.table(
  covariate = c("access", "evi"),
  measure = c("mean", "median"),
  release = c("2019_06_10", "2019_06_10")
)

start_year <- 1998
end_year <- 2017
interval_mo <- 60
template_raster <- suppressWarnings(empty_world_raster())

loader <- MbgStandardCovariateLoader$new(
  start_year = start_year,
  end_year = end_year,
  interval = interval_mo,
  covariate_config = config.table
)
# all loaded covariates will be cropped and masked to the provided template raster.
lcc <- loader$get_covariates(template_raster)
}

}
\keyword{datasets}
