% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sp_hierarchy.R
\name{merge_sp_hierarchy}
\alias{merge_sp_hierarchy}
\title{Merge shape hierarchy}
\usage{
merge_sp_hierarchy(df, admin_level, idx_col, sp_h = NULL,
  shapefile_version = "current")
}
\arguments{
\item{df}{Data frame or data table with a column of admin codes}

\item{admin_level}{The administrative level of the codes in the
\code{df} object.  Integer; 0, 1, or 2.}

\item{idx_col}{which column in \code{df} contains the administrative
codes?  A character.}

\item{sp_h}{Spatial hierarchy object generated with
\code{get_sp_hierarchy()}. If \code{NULL}, then will be generated
within the function. If using this function several times, it may
save time to generate the sp_h object once, then pass it to this
function}

\item{shapefile_version}{String specifying shapefile version to pull of is.null(sp_h)}
}
\value{
Data table with sp_hierarchy added on.  Note that the
  original admin_code column will be renamed to conform with
  the "ADMX_CODE" convention.
}
\description{
Takes a data frame or data table with a column of administrative codes
and merges on the accompanying admin code names, as well as any higher
order administrative codes within which that admin code is nested
}
\examples{
\dontrun{
# add spatial hierarchy information to df, which contains
# admin 2 codes in the column "spatial_idx", using a pre-made
# spatial hierarchy object stored in sp_h
df2 <- merge_sp_hierarchy(
  df = df,
  admin_level = 2,
  idx_col = "spatial_idx",
  sp_h = sp_h
)
}

}
