% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_with_ihme_loc.R
\name{merge_with_ihme_loc}
\alias{merge_with_ihme_loc}
\title{FUNCTION_TITLE}
\usage{
merge_with_ihme_loc(d, re = Regions, shapefile_version = "current")
}
\arguments{
\item{d}{PARAM_DESCRIPTION}

\item{re}{PARAM_DESCRIPTION, Default: Regions}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
