% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_table.R
\name{model_fit_table}
\alias{model_fit_table}
\title{FUNCTION_TITLE}
\usage{
model_fit_table(lv = loopvars[loopvars[, 3] == 0, ], rd = run_date,
  nullmodel = "", indicator = indicator,
  indicator_group = indicator_group, holdout = 0, coefs.sum1,
  use_gp = use_gp, spde_prior = spde_prior,
  use_stacking_covs = use_stacking_covs, stacker_name_vec)
}
\arguments{
\item{lv}{PARAM_DESCRIPTION, Default: loopvars[loopvars[, 3] == 0, ]}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{nullmodel}{PARAM_DESCRIPTION, Default: ''}

\item{indicator}{PARAM_DESCRIPTION, Default: indicator}

\item{indicator_group}{PARAM_DESCRIPTION, Default: indicator_group}

\item{holdout}{PARAM_DESCRIPTION, Default: 0}

\item{coefs.sum1}{PARAM_DESCRIPTION}

\item{use_gp}{PARAM_DESCRIPTION, Default: use_gp}

\item{spde_prior}{PARAM_DESCRIPTION, Default: spde_prior}

\item{use_stacking_covs}{PARAM_DESCRIPTION, Default: use_stacking_covs}

\item{stacker_name_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[INLA]{character(0)}}
}
