% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_table_list.R
\name{model_fit_table_list}
\alias{model_fit_table_list}
\title{FUNCTION_TITLE}
\usage{
model_fit_table_list(regions, rd = run_date, holdout = 0, age = 0,
  ind = indicator, ind_gp = indicator_group,
  sharedir = sprintf("/share/geospatial/mbg/\%s/\%s", indicator_group,
  indicator))
}
\arguments{
\item{regions}{PARAM_DESCRIPTION}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{holdout}{PARAM_DESCRIPTION, Default: 0}

\item{age}{PARAM_DESCRIPTION, Default: 0}

\item{ind}{PARAM_DESCRIPTION, Default: indicator}

\item{ind_gp}{PARAM_DESCRIPTION, Default: indicator_group}

\item{sharedir}{PARAM_DESCRIPTION, Default: sprintf("/share/geospatial/mbg/\%s/\%s", indicator_group, indicator)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
