% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{must_haves}
\alias{must_haves}
\title{Configuration must-haves for MBG}
\format{A vector of necessary configuration names}
\usage{
must_haves
}
\description{
Configuration must-haves for MBG
}
\examples{
\dontrun{
data(must_haves)
}
}
\keyword{datasets}
