% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_breaks.R
\name{my_breaks}
\alias{my_breaks}
\title{FUNCTION_TITLE}
\usage{
my_breaks(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
