% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_strata.R
\name{name_strata}
\alias{name_strata}
\title{FUNCTION_TITLE}
\usage{
name_strata(all_strat)
}
\arguments{
\item{all_strat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
