% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPixelVal.R
\name{NewPixelVal}
\alias{NewPixelVal}
\title{FUNCTION_TITLE}
\usage{
NewPixelVal(K, vals)
}
\arguments{
\item{K}{PARAM_DESCRIPTION}

\item{vals}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
