% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nid_folds.R
\name{nid_folds}
\alias{nid_folds}
\title{FUNCTION_TITLE}
\usage{
nid_folds(nids, yr, ss, n_folds)
}
\arguments{
\item{nids}{PARAM_DESCRIPTION}

\item{yr}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION}

\item{n_folds}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
