% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntv_plot_function.R
\name{ntv_plot_function}
\alias{ntv_plot_function}
\title{FUNCTION_TITLE}
\usage{
ntv_plot_function(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
