% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_qcd.R
\name{odds_qcd}
\alias{odds_qcd}
\title{FUNCTION_TITLE}
\usage{
odds_qcd(x, quantile_low = 5, quantile_high = 95)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{quantile_low}{PARAM_DESCRIPTION, Default: 5}

\item{quantile_high}{PARAM_DESCRIPTION, Default: 95}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
