% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_above.R
\name{p_above}
\alias{p_above}
\title{FUNCTION_TITLE}
\usage{
p_above(x, value, equal_to = F)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION}

\item{equal_to}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
