% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_region_name.R
\name{parse_region_name}
\alias{parse_region_name}
\title{FUNCTION_TITLE}
\usage{
parse_region_name(i, regions)
}
\arguments{
\item{i}{PARAM_DESCRIPTION}

\item{regions}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
