% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{percent.over}
\alias{percent.over}
\title{Fitted model objects for estimating the memory of parallel_model}
\format{An object of class \code{numeric} of length 4.}
\usage{
percent.over
}
\description{
Fitted model objects for estimating the memory of parallel_model
}
\keyword{datasets}
