% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_postamble.R
\name{pipeline_postamble}
\alias{pipeline_postamble}
\title{Pipeline Postable}
\usage{
pipeline_postamble(headnode = FALSE, addl_objs_to_save = NULL)
}
\arguments{
\item{headnode}{If TRUE, then all of the objects are saved out, and we assume
that this is the head node in the graph. Otherwise, we save a diff of
\code{ls()} and the \code{current_env}
Default: \code{FALSE}}

\item{addl_objs_to_save}{Add additional objects to save if desired.
Default: \code{NULL}.}
}
\description{
Set up start job marker, load the DAG object, and assign
the correct constants in the global environment for the node
}
