% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preamble.R
\name{pipeline_preamble}
\alias{pipeline_preamble}
\title{Pipeline Preamble}
\usage{
pipeline_preamble(headnode = FALSE)
}
\arguments{
\item{headnode}{If TRUE, then the previous node's environment is not loaded
and we assume that this is the head node in the graph.
Default: \code{FALSE}}
}
\description{
Set up start job marker, load the DAG object, and assign
the correct constants in the global environment for the node
}
