% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cov_wts.R
\name{plot.cov.wts}
\alias{plot.cov.wts}
\title{FUNCTION_TITLE}
\usage{
plot.cov.wts(rd, ind, ind_gp, reg, plot.inla.col = TRUE, age = 0,
  holdout = 0)
}
\arguments{
\item{rd}{PARAM_DESCRIPTION}

\item{ind}{PARAM_DESCRIPTION}

\item{ind_gp}{PARAM_DESCRIPTION}

\item{reg}{PARAM_DESCRIPTION}

\item{plot.inla.col}{PARAM_DESCRIPTION, Default: TRUE}

\item{age}{PARAM_DESCRIPTION, Default: 0}

\item{holdout}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[reshape2]{melt}}
}
