% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_preds.R
\name{plot.preds}
\alias{plot.preds}
\title{FUNCTION_TITLE}
\usage{
plot.preds(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
