% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rfs.R
\name{plot.rfs}
\alias{plot.rfs}
\title{FUNCTION_TITLE}
\usage{
plot.rfs(ind.gp = indicator_group, ind = indicator, rd = run_date,
  output.dir = si.fig.dir, shapefile_version = "current",
  title = "Comparison to GBD 2016 in\\n")
}
\arguments{
\item{ind.gp}{PARAM_DESCRIPTION, Default: indicator_group}

\item{ind}{PARAM_DESCRIPTION, Default: indicator}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{output.dir}{PARAM_DESCRIPTION, Default: si.fig.dir}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}

\item{title}{PARAM_DESCRIPTION, Default: 'Comparison to GBD 2016 in \\n'. Region gets pasted on to this}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
