% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_5q0_summed.R
\name{plot_5q0_summed}
\alias{plot_5q0_summed}
\title{FUNCTION_TITLE}
\usage{
plot_5q0_summed(data = df, smooth = 1, age_col = "age",
  yr_col = "year", died_col = "died", N_col = "N",
  weight_col = "weight", long_col = "longitude",
  lat_col = "latitude", dt_col = "data_type", save_dir = "~",
  dim = 10)
}
\arguments{
\item{data}{PARAM_DESCRIPTION, Default: df}

\item{smooth}{PARAM_DESCRIPTION, Default: 1}

\item{age_col}{PARAM_DESCRIPTION, Default: 'age'}

\item{yr_col}{PARAM_DESCRIPTION, Default: 'year'}

\item{died_col}{PARAM_DESCRIPTION, Default: 'died'}

\item{N_col}{PARAM_DESCRIPTION, Default: 'N'}

\item{weight_col}{PARAM_DESCRIPTION, Default: 'weight'}

\item{long_col}{PARAM_DESCRIPTION, Default: 'longitude'}

\item{lat_col}{PARAM_DESCRIPTION, Default: 'latitude'}

\item{dt_col}{PARAM_DESCRIPTION, Default: 'data_type'}

\item{save_dir}{PARAM_DESCRIPTION, Default: '~'}

\item{dim}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[base]{sample}}
}
