% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abs_errors.R
\name{plot_abs_errors}
\alias{plot_abs_errors}
\title{FUNCTION_TITLE}
\usage{
plot_abs_errors(gaul_list = gaul_list, df, sample = "BOTH",
  subset_shape = subset_shape, ind = indicator,
  ind_gp = indicator_group, rd = run_date, save.dir,
  year_col = "original_year")
}
\arguments{
\item{gaul_list}{PARAM_DESCRIPTION, Default: gaul_list}

\item{df}{PARAM_DESCRIPTION}

\item{sample}{PARAM_DESCRIPTION, Default: 'BOTH'}

\item{subset_shape}{PARAM_DESCRIPTION, Default: subset_shape}

\item{ind}{PARAM_DESCRIPTION, Default: indicator}

\item{ind_gp}{PARAM_DESCRIPTION, Default: indicator_group}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{save.dir}{PARAM_DESCRIPTION}

\item{year_col}{PARAM_DESCRIPTION, Default: 'original_year'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[base]{colSums}}
 \code{\link[scales]{muted}}
}
