% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_custom_regions.R
\name{plot_custom_regions}
\alias{plot_custom_regions}
\title{Plot the world with custom region lists}
\usage{
plot_custom_regions(region_list, plot_title = NULL, verbose = T,
  shapefile_version = "current")
}
\arguments{
\item{region_list}{vector of region names (defined in `pull_custom_modeling_regions()`)}
}
\value{
ggplot object of the map of the custom regions
}
\description{
Make a world plot of custom region lists
}
\examples{
\dontrun{
png(
  file = "/path/to/file.png",
  width = 14,
  height = 6,
  units = "in",
  res = 300
)

plot_custom_regions(region_list, plot_title = "Vaccine Regions")

dev.off()
}
}
