% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_nid_raked.R
\name{plot_each_nid_raked}
\alias{plot_each_nid_raked}
\title{FUNCTION_TITLE}
\usage{
plot_each_nid_raked(this_nid)
}
\arguments{
\item{this_nid}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
