% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam_models.R
\name{plot_gam_models}
\alias{plot_gam_models}
\title{Creates plots of component smooth objects from fitted gam model}
\usage{
plot_gam_models(child_model_list, regions, o_dir)
}
\arguments{
\item{child_model_list}{Nested list of child model objects (models within regions)
such that \code{child_model_list[["cssa"]][["gam"]]} returns the
appropriate gam model object for region cssa}

\item{regions}{character vector of regions}

\item{o_dir}{output directory}
}
\value{
writes png files in standardized format to `o_dir\\gam\\[region]\\`
}
\description{
This function creates plots in a standardized 4x3 format, saved to
standardized directories and filenames
}
\examples{
\dontrun{
message("Loading child models for each region")
child_model_list <- lapply(Regions, function(reg) {
  child_model_file <- paste0(sharedir, "child_model_list_", reg, "_0.RData")
  load(child_model_file, verbose = F)
  return(child_models)
})

names(child_model_list) <- Regions

# Plot GAM models
plot_gam_models(
  child_model_list = child_model_list,
  regions = Regions,
  o_dir = out_dir
)
}

}
