% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_of_plots.R
\name{plot_of_plots}
\alias{plot_of_plots}
\title{FUNCTION_TITLE}
\usage{
plot_of_plots(run_date, reg, age, nfolds = 5)
}
\arguments{
\item{run_date}{PARAM_DESCRIPTION}

\item{reg}{PARAM_DESCRIPTION}

\item{age}{PARAM_DESCRIPTION}

\item{nfolds}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
