% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordered_admins_colors.R
\name{plot_ordered_admins_colors}
\alias{plot_ordered_admins_colors}
\title{"Traffic light" plots and maps to help visualize categories of administrative estimates}
\usage{
plot_ordered_admins_colors(df, ad0_name, plot_year, indicator_title,
  pop_title, cutoffs, error_bars = "yaxis", ad2_shp = NULL,
  ad0_shp = NULL, verbose = F, shapefile_version = "current")
}
\arguments{
\item{df}{data.frame or data.table from the aggregation code.  Needs to have
columns for \code{mean}, \code{upper}, \code{lower}, \code{ADM2_CODE}, \code{ADM0_CODE}, \code{year}, and also a
also a \code{pop} column which can be obtained as in the example (from merging on the
pops from the admin draws object)}

\item{ad0_name}{the \code{ADM0_NAME} that you're interested in plotting. Can loop over this
in a wrapper function to be able to produce plots for many countries
sequentially}

\item{plot_year}{which year in df would you like to plot?}

\item{pop_title}{title for your population, e.g. "Population: children < 5 years"}

\item{cutoffs}{vector of cutoffs for your plot. For instance \code{c(0, 0.6, 0.9, 1)}.
Must be 4 items (3 bins) for now}

\item{error_bars}{would you like error bars? Either "yaxis" or "none"}

\item{ad2_shp}{admin2 shape for plotting}

\item{ad0_shp}{admin0 shape for national borders}

\item{verbose}{more updates from the function (Boolean)}

\item{shapefile_version}{string indicating version of shapefile to pull}

\item{indicator}{title for your indicator, e.g. "DPT3 Coverage"}
}
\value{
a grob object for the full plot. Note: need to use \code{grid.draw()} to plot grobs!
}
\description{
This is currently written to work for three broad categories (in the original case,
vaccine coverage 0-60%, 60-90%, and 90-100%) which are colored red, yellow, and blue.
Borderline estimates (i.e. with CIs that overlap 60% or 90% in the original example) are
colored orange and green.  If CIs extend into all 3 categories, then the estimate
is colored gray.

Currently, this is relatively inflexible - ideally would be extended to include
other category/color schemes and work with admin1s.  Feel free to submit a PR!
}
\examples{
\dontrun{
# Set up directories and files
# Set `run_date`, `indicator`, `indicator_group`, `out_dir` as you wish

share_dir <- paste0(
  "/share/geospatial/mbg/", indicator_group, "/",
  indicator, "/output/", run_date, "/"
)
in_dir <- paste0(share_dir, "pred_derivatives/admin_summaries/")
in_file_ad2 <- paste0(in_dir, indicator, "_admin_2_raked_summary.csv")
ad2_df <- fread(in_file_ad2)

# Drop Ma'tan al-Sarra if present
ad2_df <- subset(ad2_df, ADM0_CODE != 40762)

# Merge on populations
load(paste0(share_dir, indicator, "_raked_admin_draws_eb_bin0_0.RData"))
pops <- subset(admin_2, year \%in\% unique(ad2_df$year), select = c("ADM2_CODE", "pop", "year"))
ad2_df <- merge(ad2_df, pops, by = c("ADM2_CODE", "year"), all.x = T, all.y = F)
rm(admin_2)

# Use GAUL shapefiles as defaults
ad0_shp <- readRDS("/share/geospatial/rds_shapefiles/background_map_africa/background_map_africa.rds")
ad2_shp <- readOGR(get_admin_shapefile(admin_level = 2))

a_grob <- plot_ordered_admins_colors(
  df = ad2_df,
  ad0_name = "Senegal",
  plot_year = 2016,
  indicator_title = "DPT3 Coverage",
  pop_title = "Population: Children < 5 years",
  cutoffs = c(0, 0.6, 0.9, 1),
  error_bars = "yaxis",
  ad2_shp = ad2_shp,
  ad0_shp = ad0_shp,
  verbose = T
)

grid.draw(a_grob)
}

}
\author{
Jon Mosser, \email{jmosser@uw.edu}
}
