% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other_params.R
\name{plot_other_params}
\alias{plot_other_params}
\title{Plot other parameters}
\usage{
plot_other_params(model_results, other_params)
}
\arguments{
\item{model_results}{table from [whatever]_model_results_table.csv}

\item{other_params}{names of other parameters (not stackers)}
}
\value{
ggplot object plotting parametersby region
}
\description{
Create a plot of other parameters from INLA fit
}
\examples{
\dontrun{
# Plot stacker betas
# Plot other parameters
gg_other_params <- plot_other_params(
  model_results = model_results,
  other_params = other_params
)
}
}
