% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlay.R
\name{plot_overlay}
\alias{plot_overlay}
\title{FUNCTION_TITLE}
\usage{
plot_overlay(plot_data, multiple_runs)
}
\arguments{
\item{plot_data}{PARAM_DESCRIPTION}

\item{multiple_runs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
