% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quilt.R
\name{plot_quilt}
\alias{plot_quilt}
\title{FUNCTION_TITLE}
\usage{
plot_quilt(gaul_list, df, sample, subset_shape,
  shapefile_version = "current")
}
\arguments{
\item{gaul_list}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION}

\item{sample}{PARAM_DESCRIPTION}

\item{subset_shape}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
