% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_raster}
\alias{plot_raster}
\title{FUNCTION_TITLE}
\usage{
plot_raster(ras, indicator, region = "africa", return_map = T,
  out_dir = NULL, highisbad = T, min_value = NULL,
  mid_value = NULL, max_value = NULL, legend_title = NULL,
  plot_title = NULL, layer_names = NULL, cores = 12,
  individual_layers = T)
}
\arguments{
\item{ras}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION, Default: 'africa'}

\item{return_map}{PARAM_DESCRIPTION, Default: T}

\item{out_dir}{PARAM_DESCRIPTION, Default: NULL}

\item{highisbad}{PARAM_DESCRIPTION, Default: T}

\item{min_value}{PARAM_DESCRIPTION, Default: NULL}

\item{mid_value}{PARAM_DESCRIPTION, Default: NULL}

\item{max_value}{PARAM_DESCRIPTION, Default: NULL}

\item{legend_title}{PARAM_DESCRIPTION, Default: NULL}

\item{plot_title}{PARAM_DESCRIPTION, Default: NULL}

\item{layer_names}{PARAM_DESCRIPTION, Default: NULL}

\item{cores}{PARAM_DESCRIPTION, Default: 12}

\item{individual_layers}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[pacman]{p_load}}
}
