% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacker_betas.R
\name{plot_stacker_betas}
\alias{plot_stacker_betas}
\title{Plot stacker betas}
\usage{
plot_stacker_betas(model_results, stackers, xaxis = "stacker")
}
\arguments{
\item{model_results}{table from [whatever]_model_results_table.csv}

\item{stackers}{vector containing names of stackers}
}
\value{
ggplot object plotting stackers by region
}
\description{
Create a plot of stacker betas
}
\examples{
\dontrun{
# Plot stacker betas
gg_betas <- plot_stacker_betas(
  model_results = model_results,
  stackers = stackers
)
}
}
