% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stackers.R
\name{plot_stackers}
\alias{plot_stackers}
\title{Plot Stackers}
\usage{
plot_stackers(reg, ig = indicator_group, ind = indicator,
  rd = run_date, ss = stackers, yl = year_list, zmin = NULL,
  zmax = NULL, sh_dir = sharedir, highisbad = F, o_dir = out_dir,
  individual_countries = T, shapefile_version = "current")
}
\arguments{
\item{reg}{region}

\item{ig}{indicator group}

\item{ind}{indicator}

\item{rd}{run date}

\item{ss}{vector of stacker names}

\item{yl}{year list (in vector form, e.g. `c(2000:2015)`)}

\item{zmin}{minimum value for color scheme (if NULL will calculate from data)}

\item{zmax}{maximum value for color scheme (if NULL will calculate from data)}

\item{sh_dir}{`/share` directory, including run date}

\item{highisbad}{should high values be colored in red ("bad")? Logical.}

\item{o_dir}{output directory}

\item{individual_countries}{should individual countries be graphed as well? Logical.}

\item{shapefile_version}{string specifies shapefile version to be used}
}
\value{
Writes a series of image fileswith maps of each stacker, mean covariate raster, and
        a map of input data for each year-region combination (and year-country if individual_countries = T)
        in prespecified folder structure
}
\description{
Plot maps of stackers, mean raster of covariate, and data informing model
}
\examples{
\dontrun{
mclapply(Regions, function(r) {
  message(paste0("Making stacker maps for region: ", r))
  plot_stackers(reg = r, highisbad = F, individual_countries = T)
}, mc.cores = 5)
}
}
