% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_target.R
\name{plot_target}
\alias{plot_target}
\title{FUNCTION_TITLE}
\usage{
plot_target(gaul_list, df, sample, subset_shape, target, target_type,
  region = NULL, shapfile_version = "current")
}
\arguments{
\item{gaul_list}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION}

\item{sample}{PARAM_DESCRIPTION}

\item{subset_shape}{PARAM_DESCRIPTION}

\item{target}{PARAM_DESCRIPTION}

\item{target_type}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION, Default: NULL}

\item{shapfile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[Hmisc]{cut2}}
}
