% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varios_and_violins.R
\name{plot_varios_and_violins}
\alias{plot_varios_and_violins}
\title{FUNCTION_TITLE}
\usage{
plot_varios_and_violins(indicator, indicator_group, run_date, this_gaul,
  shapefile_version = "current")
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{this_gaul}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
