% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violins.R
\name{plot_violins}
\alias{plot_violins}
\title{FUNCTION_TITLE}
\usage{
plot_violins(indicator, indicator_proper_name, indicator_group, run_date,
  this_gaul, shapefile_verison = "current")
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_proper_name}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{this_gaul}{PARAM_DESCRIPTION}

\item{shapefile_verison}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
