% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_quadtree.R, R/points_quadtree_leaf.R
\name{points_quadtree}
\alias{points_quadtree}
\alias{points.quadtree}
\alias{points_quadtree_leaf}
\alias{points.quadtree.leaf}
\title{FUNCTION_TITLE}
\usage{
points_quadtree(q, alpha = 0.1, ...)

\method{points}{quadtree}(...)

points_quadtree_leaf(q, alpha = 0.1, ...)

\method{points}{quadtree.leaf}(...)
}
\arguments{
\item{q}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.1}

\item{...}{PARAM_DESCRIPTION}

\item{q}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.1}

\item{...}{PARAM_DESCRIPTION}

\item{q}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.1}

\item{...}{PARAM_DESCRIPTION}

\item{q}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
