% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_load_combine_save.R
\name{post_load_combine_save}
\alias{post_load_combine_save}
\title{FUNCTION_TITLE}
\usage{
post_load_combine_save(regions = strata, summstats = c("mean",
  "cirange", "upper", "lower"), raked = c("raked", "unraked"),
  rf_table = TRUE, run_summ = TRUE, indic = indicator,
  ig = indicator_group, sdir = sharedir, proj = FALSE,
  proj_folder = NULL)
}
\arguments{
\item{regions}{PARAM_DESCRIPTION, Default: strata}

\item{summstats}{PARAM_DESCRIPTION, Default: c("mean", "cirange", "upper", "lower")}

\item{raked}{PARAM_DESCRIPTION, Default: c("raked", "unraked")}

\item{rf_table}{PARAM_DESCRIPTION, Default: TRUE}

\item{run_summ}{PARAM_DESCRIPTION, Default: TRUE}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}

\item{ig}{PARAM_DESCRIPTION, Default: indicator_group}

\item{sdir}{PARAM_DESCRIPTION, Default: sharedir}

\item{proj}{Look at projection files? Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{merge}}
}
