% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_run_image_path.R
\name{pre_run_image_path}
\alias{pre_run_image_path}
\title{Set up pre-run image path}
\usage{
pre_run_image_path(indicator_group, indicator, run_date, age, region,
  holdout, file_format = ".RData", prefix = "pre_run_tempimage", ...)
}
\arguments{
\item{indicator_group}{the indicator group e.g., "hiv" or "u5m".}

\item{indicator}{the indicator e.g., "wasting_mod_b" or "male_circumcision".}

\item{run_date}{string run date in YYYY_MM_DD_HH_MM_SS format.}

\item{age}{the age for the model which uses this data.}

\item{region}{str region name.}

\item{holdout}{numeric holdout value (0 for no holdout).}

\item{file_format}{Default: ".RData".}

\item{prefix}{File prefix. Default: "pre_run_tempimage"}

\item{...}{All other relevant info}
}
\value{
the path to save the file.
}
\description{
Return the path to save model data prior to running parallel script.
}
