% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predGLM.R
\name{predGLM}
\alias{predGLM}
\title{FUNCTION_TITLE}
\usage{
predGLM(result, intercept = "(Intercept)", fixed_continuous = NULL,
  fixed_group = NULL, random_group = "cluster_id")
}
\arguments{
\item{result}{PARAM_DESCRIPTION}

\item{intercept}{PARAM_DESCRIPTION, Default: '(Intercept)'}

\item{fixed_continuous}{PARAM_DESCRIPTION, Default: NULL}

\item{fixed_group}{PARAM_DESCRIPTION, Default: NULL}

\item{random_group}{PARAM_DESCRIPTION, Default: 'cluster_id'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
