% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mbg.R
\name{predict_mbg}
\alias{predict_mbg}
\title{Predict MBG (INLA) draws}
\usage{
predict_mbg(res_fit, cs_df, mesh_s, mesh_t, cov_list, samples,
  simple_raster, transform, ind = indicator, indg = indicator_group,
  rd = run_date, region = reg, full_df = df, coefs.sum1 = FALSE,
  pred_gp = TRUE, fixed_effects = all_fixed_effects, yl = year_list,
  nperiod = ifelse(is.character(year_list), length(eval(parse(text =
  year_list))), length(year_list)), stacker_names = child_model_names,
  shapefile_version = "current", simple_raster_subnats = NULL,
  subnat_country_to_get = NULL, use_space_only_gp = FALSE,
  use_time_only_gmrf = FALSE)
}
\arguments{
\item{res_fit}{A fitted INLA model object}

\item{cs_df}{A dataframe containing mean and standard deviations used to scale the covariates}

\item{mesh_s}{An 'inla.mesh' object used to fit the spatial SPDE GP approximation in res_fit}

\item{mesh_t}{An 'inla.mesh' object used to fit the temporal correlation structure in res_fit}

\item{cov_list}{A named list containing covariate rasters used in the model fit}

\item{samples}{Integer number of draws to take from the fitted posterior distribution}

\item{simple_raster}{A simple_raster defining the modeling domain pixels}

\item{transform}{String name of transformation to be applied to
draws. Possible options are 'inverse-logit'. Any other string
will result in the identity transform.}

\item{ind}{indicator string}

\item{indg}{indicator_group string}

\item{rd}{run_date string}

\item{region}{region string}

\item{full_df}{data.frame/table containing observational data used to fit model}

\item{coefs.sum1}{Logical. Were the coefficients constrained to sum-to-1 in the model fit}

\item{pred_gp}{Logical. Should gaussian process draws be added to the draw objects.}

\item{fixed_effects}{string containing fixed effects used in res_fit separated by " + "}

\item{yl}{numeric vector of years in model}

\item{nperiod}{number of temporal periods in res_fit}

\item{stacker_names}{string vector containing names of child models used for stacking in res_fit}

\item{shapefile_version}{Shapefile version}

\item{simple_raster_subnats}{simple raster for the subnational country}

\item{subnat_country_to_get}{ISO code for subnational country}
}
\value{
list containing mean raster, sd raster, and a matrix of
  space-time pixel cell predictions. The cell prediction matrix is
  wide on draws and long on space-time locations. The space-time
  rows are formatted such that all pixels in the first period take
  up rows 1:num.space.pixels, Then period 2 space pixels are next
  and so on.
}
\description{
Generate draws from a fitted INLA MBG model across a
  space-time modeling domain. Draws are generated at the same
  resolution as simple_raster (probably 5x5km).
}
