% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mbg_tmb.R
\name{predict_mbg_tmb}
\alias{predict_mbg_tmb}
\title{Predict MBG from a TMB Model}
\usage{
predict_mbg_tmb(samples, seed = NULL, tmb_input_stack = input_data,
  model_fit_object = model_fit, fes = all_fixed_effects,
  sr = simple_raster, yl = year_list, zl = z_list,
  transform = "inverse-logit", covs_list = cov_list,
  clamp_covs = FALSE)
}
\arguments{
\item{samples}{Number of draws to take}

\item{seed}{Seed to set for RNG}

\item{model_fit_object}{object output from function fit_mbg_tmb()}

\item{fes}{a string of model fixed effects of form 'var1 + var2 + var3' or as string vector
corresponding to column names in d}

\item{sr}{simple raster object}

\item{yl}{a vector of years for analysis (i.e. c(2001,2002,2003))}

\item{zl}{a vector of z for analysis (i.e. c(1,2,3,4,5)). Config item z_list. No z-dimension it should just be zero}

\item{covs_list}{named list of covariate bricks, each should be length(yl) long}

\item{tmb_input_data}{input_data output of build_mbg_data_stack_tmb}
}
\value{
a cell_preds object
}
\description{
Project out to a full sample space defined by the sr argument
}
