% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model_raster.R
\name{predict_model_raster}
\alias{predict_model_raster}
\title{FUNCTION_TITLE}
\usage{
predict_model_raster(model_call, covariate_layers, constants = NULL,
  indicator_family = "binomial", centre_scale_df = NULL)
}
\arguments{
\item{model_call}{PARAM_DESCRIPTION}

\item{covariate_layers}{PARAM_DESCRIPTION}

\item{constants}{PARAM_DESCRIPTION, Default: NULL}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}

\item{centre_scale_df}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
