% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cell_pred.R
\name{prep_cell_pred}
\alias{prep_cell_pred}
\title{Pred cell pred}
\usage{
prep_cell_pred(cell_pred = cell_pred, cell_ids = cell_ids,
  pixel_id = pixel_id, covdt = covdt)
}
\arguments{
\item{cell_pred}{the cell pred object that will be raked and aggregated}

\item{cell_ids}{the cell ids from the global simple raster, used to connect the data to the link table}

\item{covdt}{the data table containing the populations and possibly the staker covariates as well}

\item{pixel_ids}{the cell ids from the simple raster}
}
\value{
returns the cell_pred as a dt with additional columns for cell_id and pixel_id wich allows for
merging to the link table.  Also adds the population and covariate columns if you would like.
}
\description{
This converts the cell pred to a data.table and sets up the variables needed for merging
}
