% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gbd_pops_for_fraxrake.R
\name{prep_gbd_pops_for_fraxrake}
\alias{prep_gbd_pops_for_fraxrake}
\title{Get GBD population for use in fractional raking}
\usage{
prep_gbd_pops_for_fraxrake(pop_measure = "a0004t", reg, year_list, ...)
}
\arguments{
\item{pop_measure}{WorldPop measure, default: "a0004t"}

\item{reg}{Region}

\item{year_list}{the years to get population for}
}
\value{
A data.table with \code{c('location_id', 'year_id', 'sex_id', 'run_id', 'age_group_id', 'population')}
}
\description{
Get GBD population for use in fractional raking
}
\note{
the \code{age_group_id} returned is equal to the value of \code{pop_measure}
}
