% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_link_table.R
\name{prep_link_table}
\alias{prep_link_table}
\title{Prep Link Table}
\usage{
prep_link_table(link_table = link_table, simple_raster = simple_raster,
  pixel_id = pixel_id)
}
\arguments{
\item{link_table}{the link table for your region}

\item{simple_raster}{the simple raster for your region}

\item{pixel_ids}{a list of all valif pixel ids in your simple raster}
}
\value{
a link table such that cell fractions that fall either in the ocean or in a country outside of your region are dropped
and the areas of the remain cell fractions for those cells are rescaled to equal 1 so no data is lost.  The result is cells that
lie on the border of the region are zippered into one region or the other.
}
\description{
This prepares the link table for merging and does the zippering operation on the region boundary.
The zippering means that a given cell is only in one modeled reagion.
}
