% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_postest.R
\name{prep_postest}
\alias{prep_postest}
\title{FUNCTION_TITLE}
\usage{
prep_postest(indicator, indicator_group, run_date, save_objs)
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{save_objs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
