% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_shapes_for_raking.R
\name{prep_shapes_for_raking}
\alias{prep_shapes_for_raking}
\title{Create rasters for raking and aggregation}
\usage{
prep_shapes_for_raking(reg, modeling_shapefile_version,
  raking_shapefile_version, field = "loc_id")
}
\arguments{
\item{reg}{Region}

\item{modeling_shapefile_version}{Modeling shapefile version}

\item{raking_shapefile_version}{Raking shapefile version}

\item{field}{Location field. Default: "loc_id"}
}
\value{
A list with objects: "simple_raster", "simple_polygon", "subset_shape", "new_simple_raster", "new_simple_polygon", "new_subset_shape"
}
\description{
This snippet of code is taken from \code{rake_cell_pred} so that
we can feed this into making summary statistics later on in the fractional raking code
}
