% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepLines.R
\name{prepLines}
\alias{prepLines}
\title{FUNCTION_TITLE}
\usage{
prepLines(rate, year, ylab = "", title = "", xlim = c(1999, 2016),
  ylim = c(0, max(rate)), line_years = c(2000, 2005, 2010, 2015))
}
\arguments{
\item{rate}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{ylab}{PARAM_DESCRIPTION, Default: ''}

\item{title}{PARAM_DESCRIPTION, Default: ''}

\item{xlim}{PARAM_DESCRIPTION, Default: c(1999, 2016)}

\item{ylim}{PARAM_DESCRIPTION, Default: c(0, max(rate))}

\item{line_years}{PARAM_DESCRIPTION, Default: c(2000, 2005, 2010, 2015)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
