% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_betas.R
\name{process_betas}
\alias{process_betas}
\title{FUNCTION_TITLE}
\usage{
process_betas(reg)
}
\arguments{
\item{reg}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
