% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_stack_rasters.R
\name{produce_stack_rasters}
\alias{produce_stack_rasters}
\title{Produce Stack Rasters}
\usage{
produce_stack_rasters(covariate_layers = all_cov_layers, period = 1,
  child_models = list(), indicator_family = "binomial",
  centre_scale_df = NULL)
}
\arguments{
\item{covariate_layers}{list of raster objects that consititue the covs. Must share names with the DT that the models were fit on, Default: all_cov_layers}

\item{period}{what period should be used. 1:N -- as of 11/2 we assume 4, Default: 1}

\item{child_models}{a list of model objects from the child models, Default: list()}

\item{indicator_family}{PARAM_DESCRIPTION, Default: 'binomial'}

\item{centre_scale_df}{PARAM_DESCRIPTION, Default: NULL}

\item{stacker_model}{the model object from the stacker, Default: stacker}

\item{return_children}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Create a stacked prediction for each period and child model. Function returns a list of raster objects. First object is a brick of stacked results. Additional objects refer to the child models
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
