% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_area_map.R
\name{proportional_area_map}
\alias{proportional_area_map}
\title{Proportional Area Map}
\usage{
proportional_area_map(data, ad_level, value_col, main_title = NULL,
  legend_title, fill_col, alpha = 0.25, size_scale_max = 6,
  size_scale_breaks = NULL, size_scale_labels = NULL,
  lake_river_color = "lightblue", out_file = NULL,
  out_file_height = 12, out_file_width = 12, out_file_res = 300,
  out_file_pointsize = 16, legend_text_size = 16,
  legend_title_size = 18, shapefile_version = "current")
}
\arguments{
\item{data}{data frame or data table with at least a value and ADMX_CODE column}

\item{ad_level}{admin level to map (0,1,2)}

\item{value_col}{column in `data` holding the value to plot (proportional to size)}

\item{main_title}{main title for the map (default = NULL)}

\item{legend_title}{legend title for the map}

\item{alpha}{transparency (alpha) of the proportional area circles}

\item{lake_river_color}{controls the color of the lakes/rivers}

\item{out_file}{output file name (png)}

\item{out_file_height}{height of output file (inches)}

\item{out_file_width}{width of ouput file (inches)}

\item{out_file_res}{resolution of output file (dpi)}

\item{out_file_pointsize}{pointsize of text in output file (generally would first
adjust legend text/title size as below with the relevant
options, and then use this to modify line spacing, etc.)}

\item{legend_text_size}{adjust the size of the legend text}

\item{legend_title_size}{adjust the size of the legend title}

\item{fill_color}{color to use to fill in the proportional area circles}

\item{scale_size_max}{max size of the proportional area circles
(passed to `scale_size_area()`)}

\item{scale_size_breaks}{breaks to specify for size scale legend
(passed to `scale_size_area()`)}

\item{scale_size_breaks}{labels to specify for size scale legend
(passed to `scale_size_area()`)}
}
\value{
ggplot object containing the map object; writes png to `out_file` if not NULL
}
\description{
Create proportional area maps for count data at various admin levels
}
\examples{

\dontrun{
a1_df <- fread("/share/geospatial/mbg/[ig]/[indic]/output/[rd]/pred_derivatives/admin_summaries/[indic]_admin_1_raked_summary.csv")
a1_df <- subset(a1_df, year == 2015)

gg_obj <- proportional_area_map(
  data = a1_df,
  ad_level = 1,
  value_col = "mean",
  main_title = NULL,
  legend_title = "DPT3: 2015",
  fill_col = "red",
  alpha = 0.25,
  size_scale_max = 5,
  out_file = "/my/out/file.png"
)
}
}
