% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptime_folds_combine.R
\name{proptime_folds_combine}
\alias{proptime_folds_combine}
\title{FUNCTION_TITLE}
\usage{
proptime_folds_combine(yr, n_folds, ss, ts, ...)
}
\arguments{
\item{yr}{PARAM_DESCRIPTION}

\item{n_folds}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION}

\item{ts}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
