% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_country_draws.R
\name{pull_country_draws}
\alias{pull_country_draws}
\title{FUNCTION_TITLE}
\usage{
pull_country_draws(reg, periods, raked = "", in_dir, pop_measure,
  start_year, end_year, admin2_shapes, all_region_pops,
  shapefile_version = "current")
}
\arguments{
\item{reg}{PARAM_DESCRIPTION}

\item{periods}{PARAM_DESCRIPTION}

\item{raked}{PARAM_DESCRIPTION, Default: ''}

\item{in_dir}{PARAM_DESCRIPTION}

\item{pop_measure}{PARAM_DESCRIPTION}

\item{start_year}{PARAM_DESCRIPTION}

\item{end_year}{PARAM_DESCRIPTION}

\item{admin2_shapes}{PARAM_DESCRIPTION}

\item{all_region_pops}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[seegSDM]{character(0)}}
 \code{\link[raster]{extract}}
}
