% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_covariate.R
\name{pull_covariate}
\alias{pull_covariate}
\title{FUNCTION_TITLE}
\usage{
pull_covariate(cov, ci = FALSE)
}
\arguments{
\item{cov}{PARAM_DESCRIPTION}

\item{ci}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
