% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_custom_modeling_regions.R
\name{pull_custom_modeling_regions}
\alias{pull_custom_modeling_regions}
\title{Pull custom modeling regions}
\usage{
pull_custom_modeling_regions(custom_regions)
}
\arguments{
\item{custom_region}{character vector of custom modeling regions}
}
\value{
Returns a named list of custom modeling regions with associated
   "standard" (non-custom) modeling regions that can be directly interpreted
   by get_adm0_codes().
}
\description{
Define modeling regions that are not simple combinations of the
  default MBG regions (in other words, regions that are not combinations of
  four-letter MBG regions such as "wssa" or "seas+ocea" or ISO-3 codes such
  as 'ZAF' or 'CHN').
}
