% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_df.R
\name{pull_df}
\alias{pull_df}
\title{FUNCTION_TITLE}
\usage{
pull_df(region)
}
\arguments{
\item{region}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
