% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_period_pops.R
\name{pull_period_pops}
\alias{pull_period_pops}
\title{FUNCTION_TITLE}
\usage{
pull_period_pops(period)
}
\arguments{
\item{period}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{extract}}
}
