% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_polys_foreach.R
\name{pull_polys_foreach}
\alias{pull_polys_foreach}
\title{FUNCTION_TITLE}
\usage{
pull_polys_foreach(cores, shapefiles, shapes, shp_path)
}
\arguments{
\item{cores}{PARAM_DESCRIPTION}

\item{shapefiles}{PARAM_DESCRIPTION}

\item{shapes}{PARAM_DESCRIPTION}

\item{shp_path}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
