% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_result.R
\name{pull_result}
\alias{pull_result}
\title{FUNCTION_TITLE}
\usage{
pull_result(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
