% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullupstream.R
\name{pullupstream}
\alias{pullupstream}
\title{FUNCTION_TITLE}
\usage{
pullupstream()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
