% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_to_global_env.R
\name{push_to_global_env}
\alias{push_to_global_env}
\title{Push to global environment}
\usage{
push_to_global_env(x)
}
\arguments{
\item{x}{a \code{named} list}
}
\value{
None
}
\description{
Takes a list and assigns its variables to the global scope
}
\examples{
x <- list(a = 1, b = "C")
push_to_global_env(x)
"a" \%in\% ls()
"b" \%in\% ls()
}
