% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_sing_envs.R
\name{qsub_sing_envs}
\alias{qsub_sing_envs}
\title{Adds environmental variables to a qsub string}
\usage{
qsub_sing_envs(qsub_stub, envs, image)
}
\arguments{
\item{qsub_stub}{A string containing the initial qsub string to which
environmental variables will be concatenated to in the form of
'-v ENV=VALUE'.}

\item{envs}{This should be a named list of environmental variables.
\code{qsub_sing_envs} will check that the names of the list members passed
in match the environmental variables that the shell_sing.sh script knows
about: 'SET_OMP_THREADS' and/or 'SET_MKL_THREADS'. Passing in other
environmental names in the list will result in an error. If this is left
as 'NULL' and a Singularity image is used, SET_OMP_THREADS and
SET_MKL_THREADS will remain unset and the shell_sing.sh script will use
the default setting of SET_OMP_THREADS=1 and SET_MKL_THREADS={max_threads}
(see shell_sing.sh comments). For example SET_OMP_THREADS=1 and
SET_MKL_THREADS=4 can be achieved by passing in
  \code{envs = list(SET_OMP_THREADS=1, SET_MKL_THREADS=4)}}

\item{image}{The keyword (e.g. 'default') or path to the Singularity image.
This should have been defined by \code{get_singularity} so likely
\code{get_singularity} should have been run on the 'singularity' argument
(in \code{make_qsub_share} or \code{parallelize} for example) before this
function is run.}
}
\value{
Returns a string with at least '-v sing_image=image' and possibly
  other environmental variables values if they were passed into the
  'singularity_opts' argument of functions like \code{make_qsub_share}.
}
\description{
\code{qsub_sing_envs} assumes that a qsub string is being built to launch a
Singularity container. It always adds in the '-v sing_image=sing_image' as
expected by lbd_core/mbg_central/shell_sing.sh script that ultimately
launches the container. Optionally, users may want to pass the additional
environmental variables 'SET_OMP_THREADS' and/or 'SET_MKL_THREADS' to
shell_sing.sh. If one or both of those are passed into \code{qsub_sing_envs}
it will add those environmental variables and their values as additional
'-v' flags in the construction of the qsub command.
}
\seealso{
The function \code{\link{get_singularity}} should likely be run
  before this function is run. This function is used by:
    \code{\link{parallelize}}
    \code{\link{make_qsub}}
    \code{\link{make_qsub_share}}
    \code{\link{make_qsub_postest}}
    \code{\link{submit_aggregation_script}}
}
