% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad1.R
\name{quad1}
\alias{quad1}
\title{FUNCTION_TITLE}
\usage{
quad1(xy, i, id = 1, k)
}
\arguments{
\item{xy}{PARAM_DESCRIPTION}

\item{i}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION, Default: 1}

\item{k}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
