% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad2.R
\name{quad2}
\alias{quad2}
\title{FUNCTION_TITLE}
\usage{
quad2(xy, i, id = 1, ss, target_ss, min_in_bin = 5, rand = TRUE,
  d = dim(xy))
}
\arguments{
\item{xy}{PARAM_DESCRIPTION}

\item{i}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION, Default: 1}

\item{ss}{PARAM_DESCRIPTION}

\item{target_ss}{PARAM_DESCRIPTION}

\item{min_in_bin}{PARAM_DESCRIPTION. Default: 5}

\item{rand}{PARAM_DESCRIPTION. Default: TRUE}

\item{d}{}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
