% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadtree_ct.R
\name{quadtree_ct}
\alias{quadtree_ct}
\title{FUNCTION_TITLE}
\usage{
quadtree_ct(xy, ss, target_ss, min_in_bin = 5, rand = T)
}
\arguments{
\item{xy}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION}

\item{target_ss}{PARAM_DESCRIPTION}

\item{min_in_bin}{PARAM_DESCRIPTION, Default: 5}

\item{rand}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
