% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadtree_folds.R
\name{quadtree_folds}
\alias{quadtree_folds}
\title{FUNCTION_TITLE}
\usage{
quadtree_folds(xy, ss, mb, ts, n_folds, plot_fn = NULL,
  plot_shp = NULL, save_qt = TRUE, ..., t_folds = 1, stratum = 1)
}
\arguments{
\item{xy}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION}

\item{mb}{PARAM_DESCRIPTION}

\item{ts}{PARAM_DESCRIPTION}

\item{n_folds}{PARAM_DESCRIPTION}

\item{plot_fn}{PARAM_DESCRIPTION, Default: NULL}

\item{plot_shp}{PARAM_DESCRIPTION, Default: NULL}

\item{save_qt}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}

\item{t_folds}{PARAM_DESCRIPTION, Default: 1}

\item{stratum}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
