% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_config_db.R
\name{query_config_db}
\alias{query_config_db}
\title{Query config info to SQLite database}
\usage{
query_config_db(run_date)
}
\arguments{
\item{run_date}{Run date datestamp}
}
\description{
This function will query the SQLite database with all the config records using run_date
}
