% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_cell_pred.R
\name{rake_cell_pred}
\alias{rake_cell_pred}
\title{Standard Raking Function}
\usage{
rake_cell_pred(cell_pred, rake_to, reg, year_list, pop_measure,
  rake_method = "linear", rake_subnational = T, crosswalk = F,
  shapefile_path = get_admin_shapefile(admin_level = 0, raking = T,
  version = modeling_shapefile_version), field = "loc_id",
  zero_heuristic = F, approx_0_1 = F, simple_raster = NULL,
  simple_polygon = NULL, pop_raster = NULL,
  modeling_shapefile_version = "current",
  raking_shapefile_version = "current", if_no_gbd = "return_na",
  custom_raking_shapefile = NULL, countries_not_to_subnat_rake = NULL,
  ...)
}
\arguments{
\item{cell_pred}{Cell pred object to be raked.}

\item{rake_to}{Df with `name`, `year`, and `mean` columns. values in name must match up with values in `field`.}

\item{reg}{character string - Region used to produce cell pred object.}

\item{year_list}{integer vector - vector of years}

\item{pop_measure}{character string - population measure (can be found in config of model run)}

\item{rake_method}{character string - must be either `linear` or `logit`}

\item{rake_subnational}{boolean default `T`. If true, uses the subnational raking shapefile, otherwise uses the adm0 shapefile. Make sure that values in `rake_to` contain the codes for level of raking chosen.}

\item{crosswalk}{Boolean default `T`, for models run before the new gaul shapefiles ready on 7/6/18, the cell_pred needs to be crosswalked to match the subnational raking raster. Introduces NAs into raster where the new simple raster has values but the original does not.}

\item{shapefile_path}{character string -Path to shapefile that will be used for raking. Preset to subnational raking shapefile, don't change.}

\item{field}{character string - Field in shapefile that has admin identifiers that match with rake_to. Preset to ihme_loc_ids at the lowest level in the shapefile. Don't change.}

\item{zero_heuristic}{Boolean default `F`.  If logit raking, this will automatically set rf = -9999 for any country-year with a target value of 0.  This produces a raked value of 0 for all pixels.  Raking to a target of zero in logit space is very time-consuming and the algorithm
can only approach zero asymptotically.  For situations where the target is truly zero (most useful for, say, an intervention pre-introduction) this will both speed up the process and ensure that zeros are returned.}

\item{approx_0_1}{Boolean default `F`. If logit raking, any values of zero will be replaced with 1e-10 and values of 1 will be replaced with (1-(1e-10)).  Otherwise, logit transformations will fail in `NewFindK`. Useful if some areas have very low or high predicted values in `cell_pred`,
such that some draws are either 0 or 1 (or extremely close to these values).}

\item{simple_raster}{default `NULL`, option to pass in simple raster to function if it's been loaded already. NOTE: if the pop raster is not being passed in as well, the simple_polygon needs to be supplied as well. There is a check to ensure this.}

\item{simple_polygon}{default `NULL`, option to pass in simple polygon if its been loaded already. This is necessary if the simple raster is being passed in, but the pop raster is not. The covariate loading function requires a simple polygon.}

\item{pop_raster}{default `NULL`, option to pass in pop raster if its been loaded already.

Additional Parameters (...)}

\item{modeling_shapefile_version}{string identifying version of of shapefile used in modeling}

\item{raking_shapefile_version}{string identifying version of of shapefile to use in raking}

\item{if_no_gbd}{default `return_na`, other option `return_unraked`. If return_na, any location-years without gbd estimates will return NA raking factors. if return_unraked, will any location-years without gbd estimates will return 1 for linear raking, 0 for logit raking.}

\item{custom_raking_shapefile}{SPDF object -shapefile that will be used for raking. Used for passing in custom raking shapefile for choosing subnational countries to rake to. See `make_custom_raking_shapefile()`}

\item{countries_not_to_subnat_rake}{as it sounds. Used for constructing raking raster. Default: NULL}

\item{MaxJump}{default `10`. Maximum size of a jump to the answer (for logit raking).}

\item{MaxIter}{default `80`. Number of jumps towards the solution (for logit raking)}

\item{FunTol}{default `1e-5`. Maximum allowed difference between the raking target and raked results (for logit raking)}

\item{iterate}{default `F`. If logit raking for a location-year fails, try again with `MaxJump` and `MaxIter` times 10. If that fails, try again times 100. For circumstances where raking target is very far from estimate and raking does not converge.}

\item{if_no_gbd}{default `return_na`, other option `return_unraked`. If return_na, any location-years without gbd estimates will return NA raking factors. if return_unraked, will any location-years without gbd estimates will return 1 for linear raking, 0 for logit raking.}
}
\value{
Returns a named list with a raked cell pred object, simple raster used for raking, raking factors, and (optional) rasters of mean, lower, upper, and cirange for years in year list
}
\description{
A function used to rake mbg ouputs to GBD estimates. Can rake to either national level or subnational level (where available). Supports linear and logit raking. Optionally outputs summary rasters of raked cell pred object.
}
\author{
Michael Collison
}
