% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_predictions.R
\name{rake_predictions}
\alias{rake_predictions}
\title{Raking predictions}
\usage{
rake_predictions(raking_factors, pop_wts_object, cell_pred,
  logit_rake = FALSE)
}
\arguments{
\item{raking_factors}{output from calc_raking_factors}

\item{pop_wts_object}{output from make_population_weights}

\item{cell_pred}{Cells by Draws matrix which is output from predict_mbg()}

\item{logit_rake}{do logit raking?, Default: FALSE}
}
\value{
eturns cell_preds raked
}
\description{
Rakes estimates to gold standard estimates (usually GBD)
}
