% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raked_frax_counts_save.R
\name{raked_frax_counts_save}
\alias{raked_frax_counts_save}
\title{Save outputs from fractional counts raking}
\usage{
raked_frax_counts_save(output_list, sharedir, indicator, age, reg, holdout)
}
\arguments{
\item{output_list}{The list of outputs from \code{fractionally_rake_counts} call.}

\item{sharedir}{Output to directory}

\item{indicator}{Indicator}

\item{age}{Age}

\item{reg}{Region}

\item{holdout}{Holdout}
}
\description{
Save outputs from fractional count raking
coming out of \code{fractionally_rake_counts} call.
}
\examples{
\dontrun{
raked_frax_counts_save(
  output_list = outputs,
  sharedir = "/share/geospatial/mbg/training/tr_had_diarrhea",
  indicator = "tr_had_diarrhea",
  age = 0,
  reg = "tza",
  holdout = 0
)
}

}
