% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking_factors_map.R
\name{raking_factors_map}
\alias{raking_factors_map}
\title{FUNCTION_TITLE}
\usage{
raking_factors_map(indicator, indicator_group, run_date, gaul,
  shapefile_version, year_plot_list)
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{gaul}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION}

\item{year_plot_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
