% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_folds.R
\name{rand_folds}
\alias{rand_folds}
\title{FUNCTION_TITLE}
\usage{
rand_folds(data, n_folds = 5, strat_cols = NULL, ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{n_folds}{PARAM_DESCRIPTION, Default: 5}

\item{strat_cols}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
