% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_s_folds.R
\name{rand_s_folds}
\alias{rand_s_folds}
\title{FUNCTION_TITLE}
\usage{
rand_s_folds(xy, ss = 1, n_folds = 5, ...)
}
\arguments{
\item{xy}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION, Default: 1}

\item{n_folds}{PARAM_DESCRIPTION, Default: 5}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
