% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_draws.R
\name{rank_draws}
\alias{rank_draws}
\title{rank_draws}
\usage{
rank_draws(df, ordr, columns)
}
\arguments{
\item{df}{data frame with columns for each draw}

\item{'high'}{means a high value is rank 1, 'low' means a
low value should be rank 1}

\item{columns:}{vector of column names that contain the draws}
}
\description{
function to transform values to ranks by draw
}
