% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_covariate_config.R
\name{read_covariate_config}
\alias{read_covariate_config}
\title{Reads covariate configuration file.}
\usage{
read_covariate_config(path_or_text, ...)
}
\arguments{
\item{path_or_text}{the file path or literal CSV text to read in.}

\item{...}{any additional arguments to pass to \code{data.table::fread}}
}
\value{
configuration as a data.table
}
\description{
Read covariate configuration file providing sensible default values \code{header = TRUE, fill = TRUE}
}
