% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_reg_list}
\alias{ref_reg_list}
\title{Full List of All Reference Regions}
\format{An object of class \code{list} of length 78.}
\usage{
ref_reg_list
}
\description{
Define modeling regions that are not simple combinations of the
  default MBG regions (in other words, regions that are not combinations of
  four-letter MBG regions such as "wssa" or "seas+ocea" or ISO-3 codes such
  as 'ZAF' or 'CHN').
If you need to add a new custom region, add it to this list
}
\examples{
\dontrun{
data(ref_reg_list)
}
}
\keyword{datasets}
