% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_violin.R
\name{region_violin}
\alias{region_violin}
\title{FUNCTION_TITLE}
\usage{
region_violin(indicator, indicator_group, run_date, output_file)
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{output_file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
