% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_aarons_shapefiles.R
\name{rename_aarons_shapefiles}
\alias{rename_aarons_shapefiles}
\title{FUNCTION_TITLE}
\usage{
rename_aarons_shapefiles(len = (length(Regions) * 4), prefix)
}
\arguments{
\item{len}{PARAM_DESCRIPTION, Default: (
length(Regions) * 4}

\item{prefix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
