% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_each_gaul.R
\name{report_each_gaul}
\alias{report_each_gaul}
\title{FUNCTION_TITLE}
\usage{
report_each_gaul(gaul, quilt_plot_list, target_plot_list,
  shapefile_version = "current")
}
\arguments{
\item{gaul}{PARAM_DESCRIPTION}

\item{quilt_plot_list}{PARAM_DESCRIPTION}

\item{target_plot_list}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
