% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTime.R
\name{reportTime}
\alias{reportTime}
\title{FUNCTION_TITLE}
\usage{
reportTime()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
