% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res.sd}
\alias{res.sd}
\title{Fitted model objects for estimating the memory of parallel_model}
\format{An object of class \code{numeric} of length 1.}
\usage{
res.sd
}
\description{
Fitted model objects for estimating the memory of parallel_model
}
\keyword{datasets}
