% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_polygons.R
\name{resample_polygons}
\alias{resample_polygons}
\title{Function to perform polygon resampling}
\usage{
resample_polygons(data,
  shp_path = "/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory",
  ignore_warnings = TRUE, cores = as.numeric(slots),
  indic = indicator, unique_vars = NULL, density = 0.001,
  perpixel = TRUE, prob = TRUE, use_1k_popraster = TRUE,
  pull_poly_method = "fast", gaul_list = get_adm0_codes("africa",
  shapefile_version = "current"), seed = NULL,
  shapefile_version = "current", custom_ras = NULL, years = NULL)
}
\arguments{
\item{data}{a data frame or data table with the following columns: cluster_id, exposed,
<indicator>, latitude, longitude, shapefile, location code. Location_code refers to
the shapefile polygon that the data. Data should alreaday be aggregated up
to the smallest geog unit (if this is point, then lat long should have values).
If lat long don't have values then shapefile and location_id should}

\item{shp_path}{Path to shapefiles (standard = our shapefile directory)}

\item{ignore_warnings}{if TRUE, and there are polys that are not found the function will warn but not stop.}

\item{cores}{Number of cores to use for parallel processing bits}

\item{indic}{Indicator being run (e.g. `died`, `dpt3_cov`, `has_lri`.  Note that this is only used for some very specific
renaming at the very end of the function - this code will actually preserve all columns passed to it.}

\item{unique_vars}{character values of additional names that uniquely ID rows in data, for example "age_bin" for
u5m data. Defaults to NULL}

\item{density}{Passed to `seegMBG::getPoints()` as the `n` argument, which determines the # of integration points.
Determines how many pseudopoints are laid down for each polygon. Default = 0.001.
This can be interpreted as points per pixel if `perpixel = T`, and as points per square km if `perpixel = T`
and `use_1k_popraster = T` (default settings)}

\item{perpixel}{Passed to `seegMBG::getPoints()` as the `perpixel` argument  This determines whether the `density`
arguments represents an absolute number, or if it is the number of integration points per pixel (so
that the total # of integration points = `density` * `[total number of pixels]`).  By default this is
TRUE with a density of 0.001.}

\item{prob}{Passed to `seegMBG::getPoints()` as the `prob` argument. Determines whether to weight the integration points
by the values of the raster (here, the population raster). TRUE by default.}

\item{use_1k_popraster}{Should a 1km x 1km population raster be used (disaggregates the 5x5 km population raster)?
Logical, default = TRUE.  Otherwise will use 5k pop raster}

\item{pull_poly_method}{How should we get the polygons for resampling?  Options are parallel processing with `mclapply()`
(`pull_poly_method = "mclapply"`), parallel processing with `foreach()` (`pull_poly_method = "foreach"`)
or fast polygon pulling using RDS files (`pull_poly_method = "fast"` -- the default)}

\item{gaul_list}{List of gaul codes covering your modeling area, for use in pulling the population rasters}

\item{seed}{Optional seed that is set just prior to resampling to gaurantee the same results when rerunning on the same data.
By default this is `NULL`, in which case no seed is set.}

\item{shapefile_version}{string indicating which admin shapefile to use to generate population raster}
}
\value{
A data table with your input `data` object, now with polygons duplicated and weighed according to the resampling
        algorithm
}
\description{
Resample polygons using the k-means clustering algorithm developed by RB et al
}
\examples{
\dontrun{
# Will not run unless you have vaccine data
df_pointpoly <- resample_polygons(
  data = df_input,
  cores = 20,
  indic = "dpt3_cov",
  density = 0.001
)
}
}
