% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_polygons_bysamplingtype.R
\name{resample_polygons_bysamplingtype}
\alias{resample_polygons_bysamplingtype}
\title{FUNCTION_TITLE}
\usage{
resample_polygons_bysamplingtype(data,
  shp_path = "/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory",
  ignore_warnings = TRUE, cores = as.numeric(slots),
  indic = indicator, unique_vars = NULL, density = 0.001,
  perpixel = TRUE, prob = TRUE, use_1k_popraster = TRUE,
  raster_dict = NULL, raster_col = NULL, regions = NULL,
  shapefile_version = "current")
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{shp_path}{PARAM_DESCRIPTION, Default: '/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory'}

\item{ignore_warnings}{PARAM_DESCRIPTION, Default: TRUE}

\item{cores}{PARAM_DESCRIPTION, Default: as.numeric(slots)}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}

\item{unique_vars}{PARAM_DESCRIPTION, Default: NULL}

\item{density}{PARAM_DESCRIPTION, Default: 0.001}

\item{perpixel}{PARAM_DESCRIPTION, Default: TRUE}

\item{prob}{PARAM_DESCRIPTION, Default: TRUE}

\item{use_1k_popraster}{PARAM_DESCRIPTION, Default: TRUE}

\item{raster_dict}{PARAM_DESCRIPTION, Default: NULL}

\item{raster_col}{PARAM_DESCRIPTION, Default: NULL}

\item{regions}{PARAM_DESCRIPTION, Default: NULL}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
