% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmlike.R
\name{rmlike}
\alias{rmlike}
\title{FUNCTION_TITLE}
\usage{
rmlike(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
