% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormMatrix.R
\name{rnormMatrix}
\alias{rnormMatrix}
\title{FUNCTION_TITLE}
\usage{
rnormMatrix(n, mean, sd)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{mean}{PARAM_DESCRIPTION}

\item{sd}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
