% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimizerPy.R
\name{run_optimizerPy}
\alias{run_optimizerPy}
\title{Run BRT Optimizer in Python}
\usage{
run_optimizerPy(funcs.file_path, funcs.file, bounds.file_path, bounds.file,
  data.file_path, data.loc, optimizer, learner, cv_folds, n_calls, jobnum,
  col_start, return_cmd_instead_of_running = FALSE)
}
\arguments{
\item{funcs.file_path}{directory containing the optimizer script}

\item{funcs.file}{file name of optimizer script file}

\item{bounds.file_path}{directory containing the search space boundaries csv}

\item{bounds.file}{file name of search space boundaries csv file}

\item{data.file_path}{model data file path.}

\item{data.loc}{model data file sub dir -- i.e. -- data.loc <- paste0('/data_train_', jobnum, '.csv'),}

\item{optimizer}{choice of non-parametric model to optimize with, either: 'rf' (Random Forest), 'gp' (Gaussian Process), 'brt' (Boosted Regression Tree)}

\item{learner}{choice of BRT model, either: brtR, brtC, xgbR, xgbC -- R=regression, C=classification, xgb=use XGBoost BRT, BRT=use sklearn brt}

\item{cv_folds}{number of cross-validation folds / set of parameters}

\item{n_calls}{number of draws to make to inform posterior distribution over parameter space}

\item{jobnum}{The resulting file is saved to: data.file_path + "best_pars_" + jobnum + ".csv"}

\item{col_start}{numeric index. Assumes passed array has covariate data from here until the last column}

\item{return_cmd_instead_of_running}{if TRUE returns the string to run the program instead of calling system(cmd). You probably don't want this unless you're testing or want to embed this command in a qsub call.}
}
\value{
The string of the qsub command, or nothing if the job is chosen to be submitted
}
\description{
Tune hyperparameters for BRT and save tuning results with shrinkage, intersection depth, and the number of trees.
}
