% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{run_query}
\alias{run_query}
\title{FUNCTION_TITLE}
\usage{
run_query(dbname, host, query)
}
\arguments{
\item{dbname}{PARAM_DESCRIPTION}

\item{host}{PARAM_DESCRIPTION}

\item{query}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
