% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sql_query.R
\name{run_sql_query}
\alias{run_sql_query}
\title{FUNCTION_TITLE}
\usage{
run_sql_query(conn, query)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{query}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[DBI]{dbGetQuery}},\code{\link[DBI]{character(0)}}
 \code{\link[RMySQL]{dbNextResult}}
}
