% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_best_model.R
\name{save_best_model}
\alias{save_best_model}
\title{FUNCTION_TITLE}
\usage{
save_best_model(indicator_group, indicator, run_date, measure)
}
\arguments{
\item{indicator_group}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{measure}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
