% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_brt_contributions.R
\name{save_brt_contributions}
\alias{save_brt_contributions}
\title{FUNCTION_TITLE}
\usage{
save_brt_contributions(brt_mod_obj = trans_covs[[1]], a = age,
  r = reg, pa = pathaddin, returnx = FALSE)
}
\arguments{
\item{brt_mod_obj}{PARAM_DESCRIPTION, Default: trans_covs[[1]]}

\item{a}{PARAM_DESCRIPTION, Default: age}

\item{r}{PARAM_DESCRIPTION, Default: reg}

\item{pa}{PARAM_DESCRIPTION, Default: pathaddin}

\item{returnx}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
save a csv of contributions of different variables in brt
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
