% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_custom_raking_outputs.R
\name{save_custom_raking_outputs}
\alias{save_custom_raking_outputs}
\title{Save outputs from custom raking function}
\usage{
save_custom_raking_outputs(custom_rake_output, outdir, indicator, prefix,
  reg)
}
\arguments{
\item{custom_rake_output}{Output list from custom raking function}

\item{outdir}{Directory for files to be saved to}

\item{indicator}{Name of indicator being modeled}

\item{prefix}{Character string to be added to files to avoid overwriting non-custom files in folder}

\item{reg}{Name of region that was modeled}

\item{age_group}{Name of age group}
}
\description{
This function saves outputs from \code{custom_rake}
}
\examples{
\dontrun{
save_custom_raking_outputs(custom_rake_output,
  outdir = sprintf("/share/geospatial/mbg/u5m/\%s_\%s/output/\%s", indicator, age_group, run_date),
  indicator,
  age_group,
  prefix = "custom_india",
  reg = "south_asia"
)
}
}
