% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_maps.R
\name{save_maps}
\alias{save_maps}
\title{FUNCTION_TITLE}
\usage{
save_maps(input_df, stacker_list, master_shape, result_brick, zmin, zmax,
  yl, ind, ig, sh_dir, highisbad, o_dir, ctry = NULL, shapefile_version)
}
\arguments{
\item{input_df}{PARAM_DESCRIPTION}

\item{stacker_list}{PARAM_DESCRIPTION}

\item{master_shape}{PARAM_DESCRIPTION}

\item{result_brick}{PARAM_DESCRIPTION}

\item{zmin}{PARAM_DESCRIPTION}

\item{zmax}{PARAM_DESCRIPTION}

\item{yl}{PARAM_DESCRIPTION}

\item{ind}{PARAM_DESCRIPTION}

\item{ig}{PARAM_DESCRIPTION}

\item{sh_dir}{PARAM_DESCRIPTION}

\item{highisbad}{PARAM_DESCRIPTION}

\item{o_dir}{PARAM_DESCRIPTION}

\item{ctry}{PARAM_DESCRIPTION, Default: NULL}

\item{shapefile_version}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
