% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_mbg_input.R
\name{save_mbg_input}
\alias{save_mbg_input}
\title{FUNCTION_TITLE}
\usage{
save_mbg_input(indicator = indicator,
  indicator_group = indicator_group, df = df,
  simple_raster = simple_raster, simple_raster2 = NULL,
  mesh_s = mesh_s, mesh_t = mesh_t, cov_list = all_cov_layers,
  run_date = NULL, pathaddin = "", child_model_names = NULL,
  all_fixed_effects = NULL, period_map = NULL, centre_scale = T)
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION, Default: indicator}

\item{indicator_group}{PARAM_DESCRIPTION, Default: indicator_group}

\item{df}{PARAM_DESCRIPTION, Default: df}

\item{simple_raster}{PARAM_DESCRIPTION, Default: simple_raster}

\item{mesh_s}{PARAM_DESCRIPTION, Default: mesh_s}

\item{mesh_t}{PARAM_DESCRIPTION, Default: mesh_t}

\item{cov_list}{PARAM_DESCRIPTION, Default: all_cov_layers}

\item{run_date}{PARAM_DESCRIPTION, Default: NULL}

\item{pathaddin}{PARAM_DESCRIPTION, Default: ''}

\item{child_model_names}{PARAM_DESCRIPTION, Default: NULL}

\item{all_fixed_effects}{PARAM_DESCRIPTION, Default: NULL}

\item{period_map}{PARAM_DESCRIPTION, Default: NULL}

\item{centre_scale}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[raster]{mask}}
}
