% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_mbg_preds.R
\name{save_mbg_preds}
\alias{save_mbg_preds}
\title{FUNCTION_TITLE}
\usage{
save_mbg_preds(config, time_stamp, run_date, mean_ras, sd_ras, res_fit,
  cell_pred, df, pathaddin = "")
}
\arguments{
\item{config}{PARAM_DESCRIPTION}

\item{time_stamp}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{mean_ras}{PARAM_DESCRIPTION}

\item{sd_ras}{PARAM_DESCRIPTION}

\item{res_fit}{PARAM_DESCRIPTION}

\item{cell_pred}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION}

\item{pathaddin}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
