% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_post_est.R
\name{save_post_est}
\alias{save_post_est}
\title{FUNCTION_TITLE}
\usage{
save_post_est(x, filetype, filename, indic = indicator)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{filetype}{PARAM_DESCRIPTION}

\item{filename}{PARAM_DESCRIPTION}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
