% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_raking_outputs.R
\name{save_raking_outputs}
\alias{save_raking_outputs}
\title{Save Outputs from Raking Function}
\usage{
save_raking_outputs(rake_output, outdir, indicator, prefix, reg, year_list,
  save_vec = NULL, save_draws_as = "RDS")
}
\arguments{
\item{rake_output}{Output list from subnational raking function}

\item{outdir}{character string - Directory for files to be saved to}

\item{indicator}{charater string - Name of indicator being modeled}

\item{prefix}{Character string to be added to start of file name to avoid overwriting non-custom files in folder}

\item{reg}{character string - Name of region that was modeled}

\item{year_list}{integer vector of years}

\item{save_vec}{default `NULL`. If null, uses names from subnational_rake_output. Otherwise takes a character vector with the names of objects in subnational_rake_output to save.}

\item{save_draws_as}{character string default `RDS`. additional supported arguments `rds` and `Rdata`. The file type to save raked draws. If using `RDS` or `rds`, the file extension will be .RDS or .rds accordingly.}

\item{age_group}{character string - Name of age group}
}
\description{
Save Outputs from Raking Function
}
\examples{
\dontrun{
save_custom_raking_outputs(custom_rake_output,
  outdir = sprintf("/share/geospatial/mbg/u5m/\%s_\%s/output/\%s", indicator, age_group, run_date),
  indicator,
  age_group,
  prefix = "custom_india",
  reg = "south_asia"
)
}
}
