% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_shapefile_as_rds.R
\name{save_shapefile_as_rds}
\alias{save_shapefile_as_rds}
\title{FUNCTION_TITLE}
\usage{
save_shapefile_as_rds(shapefile_path, verbose = F,
  in_dir = "/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory/",
  out_dir = "/share/geospatial/rds_shapefiles/")
}
\arguments{
\item{shapefile_path}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: F}

\item{in_dir}{PARAM_DESCRIPTION, Default: '/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory/'}

\item{out_dir}{PARAM_DESCRIPTION, Default: '/share/geospatial/rds_shapefiles/'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_match}}
}
