% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_standard_covariate.R
\name{save_standard_covariate}
\alias{save_standard_covariate}
\title{Save Standard Covariate}
\usage{
save_standard_covariate(cov_name, ig = indicator_group, run_date,
  ind = NULL, raked = T, year_list = c(2000:2015),
  measure = "mean")
}
\arguments{
\item{cov_name}{Name for the covariate (how will it be called in the
fixed effects parameter?).  In most cases will be the same
as indicator.}

\item{ig}{Indicator groups}

\item{run_date}{Run date to pull from to make this covariate}

\item{ind}{Indicator (if different from cov name).  If `indicator` is NULL,
(default) then `indicator` is set to `cov_name`.}

\item{raked}{Do you want to use the raked version?}

\item{year_list}{Vector (numeric) of years, with length the same as the number
of layers in your modeled output .tif}

\item{measure}{Covariate measure (e.g. mean, median, etc.)}
}
\value{
Creates a series of .tif files, one for each year, in the standard MBG
        covariate directory using appropriate naming convention.  Note that for
        consistency with other covariates, these are expanded to global rasters.
        Also saves a readme.txt file with details of the covariate creation.
}
\description{
Save an mbg output as a standard covariate
}
\examples{
\dontrun{
save_standard_covariate(
  cov_name = "dpt3_cov_test",
  ig = "vaccine",
  run_date = "2017_10_27_12_55_26",
  raked = T,
  year_list = c(2000:2015),
  measure = "mean"
)
}
}
