% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_original_threads.R
\name{set_original_threads}
\alias{set_original_threads}
\title{Sets MKL/OMP threads back to original setting}
\usage{
set_original_threads()
}
\value{
None
}
\description{
\code{set_original_threads()} Uses functions in the "setthreads.so" shared
library built into LBD Singularity images to set MKL/OMP back to an original
setting.
}
\details{
Uses functions in the "setthreads.so" shared library built into LBD
Singularity images to set MKL/OMP threads back to an original setting. The
shared library should exist in the LBD Singularity image and should have
already been loaded by \code{mbg_setup()}. This function checks to make sure
that it is loaded and if not, attempts to use \code{load_setthreads()} to do
so. A warning is generated if this is run outside of an LBD Singularity image
and no thread adjustment is done.

This function is meant to be used after \code{set_threads_serial()} was used
to set MKL/OMP operations to serial before a call to a process that forks
(like \code{mclapply()}) which is known to have issues with threaded
processes. This function will try and retrieve the original thread settings
from the MKL_NUM_THREADS or OMP_NUM_THREADS environmental variable and again
set the threads back to those values. If those environmentals are not set,
it sets them to 1.
}
\examples{
\dontrun{
set_serial_threads()
mclapply(...)
set_original_threads()
}

}
\seealso{
This function depends on:
\code{\link{load_setthreads}}
and is usually coupled with:
\code{\link{set_serial_threads}}

Other MBG setup functions: \code{\link{is_integer}},
  \code{\link{is_lbd_core_repo}},
  \code{\link{is_lbd_singularity}},
  \code{\link{is_rstudio}}, \code{\link{is_singularity}},
  \code{\link{load_R_packages}},
  \code{\link{load_mbg_functions}},
  \code{\link{load_setthreads}}, \code{\link{mbg_setup}},
  \code{\link{set_serial_threads}},
  \code{\link{source_functions}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
