% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_serial_threads.R
\name{set_serial_threads}
\alias{set_serial_threads}
\title{Sets MKL/OMP threads to serial}
\usage{
set_serial_threads()
}
\value{
None
}
\description{
\code{set_serial_threads()} Uses functions in the "setthreads.so" shared
library built into LBD Singularity images to set MKL/OMP threads to serial.
}
\details{
Uses functions in the "setthreads.so" shared ' library built into LBD
Singularity images to set MKL/OMP threads to serial. The shared library
should exist in ' the LBD Singularity image and should have already been
loaded by \code{mbg_setup()}. This function checks to make sure that it is
loaded and if not, attempts to use \code{load_setthreads()} to do so. A
warning is generated if this is run outside of an LBD Singularity image and
no thread adjustment is done.

This can be used to set any multi-threaded operations to serial before a call
to a process that forks (like \code{mclapply()}) which is known to have
issues with threaded processes. Then \code{set_original_threads()} can be
used to set the threads back to their original setting.
}
\examples{
\dontrun{
set_serial_threads()
mclapply(...)
set_original_threads()
}

}
\seealso{
This function depends on:
\code{\link{load_setthreads}}
and is usually coupled with:
\code{\link{set_original_threads}}

Other MBG setup functions: \code{\link{is_integer}},
  \code{\link{is_lbd_core_repo}},
  \code{\link{is_lbd_singularity}},
  \code{\link{is_rstudio}}, \code{\link{is_singularity}},
  \code{\link{load_R_packages}},
  \code{\link{load_mbg_functions}},
  \code{\link{load_setthreads}}, \code{\link{mbg_setup}},
  \code{\link{set_original_threads}},
  \code{\link{source_functions}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
