% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setmklthreads.R
\name{setmklthreads}
\alias{setmklthreads}
\title{Sets number of MKL threads}
\usage{
setmklthreads(threads = 1)
}
\arguments{
\item{threads}{A single value indicating the number of threads to use
[default = 1]}
}
\value{
None
}
\description{
\code{setmklthreads()} Uses a function in the "setthreads.so" shared library
built into LBD Singularity images to set the number of MKL threads to a user
defined value
}
\details{
Uses a function in the "setthreads.so" shared library built into LBD
Singularity images to set the number of MKL threads to a user defined value.
The shared library should exist in the LBD Singularity image and should have
already been loaded by \code{mbg_setup()}. This function checks to make sure
that it is loaded and if not, attempts to use \code{load_setthreads()} to do
so. A warning is generated if this is run outside of an LBD Singularity image
and no thread adjustment is done.
}
\examples{
\dontrun{
setmklthreads(2) # sets the MKL threads to 2
}

}
\seealso{
This function depends on:
\code{\link{load_setthreads()}}
\code{\link{is_integer()}}
Is used by:
\code{\link{set_serial_threads()}}
\code{\link{set_original_threads()}}
And is related to:
\code{\link{get_mkl_threads()}}
OMP thread setting is done with \code{\link{setompthreads}}

Other Mutlti-threading Functions: \code{\link{get_max_forked_threads}},
  \code{\link{get_mkl_threads}},
  \code{\link{get_omp_threads}},
  \code{\link{get_total_threads}},
  \code{\link{setmkldynamic}}, \code{\link{setompnested}},
  \code{\link{setompthreads}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{Mutlti-threading Functions}
