% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_design.R
\name{setup_design}
\alias{setup_design}
\title{FUNCTION_TITLE}
\usage{
setup_design(df, var)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
