% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_log_location.R
\name{setup_log_location}
\alias{setup_log_location}
\title{Set up log location}
\usage{
setup_log_location(log_location, user, indicator, indicator_group,
  run_date, make_dirs = TRUE)
}
\arguments{
\item{log_location}{the location of the log files OR 'sgeoutput' OR
'sharedir'. 'sgeoutput' and 'sharedir' will appropriately substitute the
user's default sgeoutput directory or the models default output directory
(respectively).}

\item{user}{the name of the user.}

\item{indicator}{the indicator e.g., "wasting_mod_b" or "male_circumcision".}

\item{run_date}{string run date in YYYY_MM_DD_HH_MM_SS format.}

\item{make_dirs}{whether to make the directories. [default = TRUE]}

\item{age}{the age for the model which uses this data.}
}
\description{
Determines error/output file path and optionally ensures dirs exist.
}
