% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_parallel_model_inputs.R
\name{setup_parallel_model_inputs}
\alias{setup_parallel_model_inputs}
\title{Grabbing parallel model inputs}
\usage{
setup_parallel_model_inputs(use_argparse = FALSE,
  load_prerun_image = TRUE, file_format = ".rds")
}
\arguments{
\item{use_argparse}{Boolean. Was argparse used in the qsub? Default: FALSE}

\item{load_prerun_image}{Boolean. Load pre-saved image? Default: TRUE}

\item{file_format}{File format of pre-run image. Default = ".rds"}
}
\value{
A named \code{list} with the arguments: \code{reg, age, run_date, test, holdout, indicator, indicator_group, pathaddin, outputdir}
}
\description{
Get parallel model inputs from the command line arguments
}
