% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_to_spointsdf.R
\name{shape_to_spointsdf}
\alias{shape_to_spointsdf}
\title{Shapefile to SpatialPointsDataFrame}
\usage{
shape_to_spointsdf(sp.df, field)
}
\arguments{
\item{sp.df}{SpatialPolygonDataFrame. Input admin shapefile}

\item{field}{String. The field describing the shapes, for e.g. ADM0_CODE}
}
\value{
A SpatialPointsDataFrame with coords and value of field
}
\description{
Converts a shapefile into a SpatialPointsDataFrame for a given field
}
