% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_cov_layers.R
\name{shiny_cov_layers}
\alias{shiny_cov_layers}
\title{FUNCTION_TITLE}
\usage{
shiny_cov_layers(fixed_effects, gaul_list, run_date, indicator,
  indicator_group)
}
\arguments{
\item{fixed_effects}{PARAM_DESCRIPTION}

\item{gaul_list}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
