% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_data_scatter.R
\name{shiny_data_scatter}
\alias{shiny_data_scatter}
\title{FUNCTION_TITLE}
\usage{
shiny_data_scatter(df, run_date, indicator, indicator_group, year_var)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{year_var}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
