% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_raked.R
\name{shiny_raked}
\alias{shiny_raked}
\title{FUNCTION_TITLE}
\usage{
shiny_raked(gaul_list, run_date, indicator, indicator_group, pred_file,
  layer_name)
}
\arguments{
\item{gaul_list}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{pred_file}{PARAM_DESCRIPTION}

\item{layer_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
