% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_raking_map.R
\name{shiny_raking_map}
\alias{shiny_raking_map}
\title{FUNCTION_TITLE}
\usage{
shiny_raking_map()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
