% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_spdf.R
\name{simplify_spdf}
\alias{simplify_spdf}
\title{FUNCTION_TITLE}
\usage{
simplify_spdf(the_spdf, ...)
}
\arguments{
\item{the_spdf}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
