% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_to_inla_copyfile.R
\name{skip_to_inla_copyfile}
\alias{skip_to_inla_copyfile}
\title{Copy Pre-prepped INLA image}
\usage{
skip_to_inla_copyfile(indicator_group, indicator, run_date,
  skiptoinla_from_rundate, pathaddin)
}
\arguments{
\item{indicator_group}{Indicator Group}

\item{indicator}{Indicator Name}

\item{run_date}{Current model run-date}

\item{skiptoinla_from_rundate}{A run-date which had INLA info saved out}

\item{pathaddin}{Additional thing to add to path, quite often: \code{paste0("_bin", age, "_", reg, "_", holdout)}}
}
\description{
Copy Pre-prepped INLA image from a specified rundate into current rundate
}
