% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{source_functions}
\alias{source_functions}
\title{Sources a list of functions}
\usage{
source_functions(functions)
}
\arguments{
\item{functions}{Character vector of functions names including COMPLETE path}
}
\value{
None
}
\description{
\code{source_functions} sources a list of functions with
with complete path names
}
\details{
A standalone function that will source in functions from a provided character
vector of functions names with complete paths.
}
\examples{
\dontrun{
mbg_functions <- c("setup.R", "mbg_functions.R", "misc_functions.R")
source_functions(paste0(
  "/share/code/geospatial/imdavis/repos/lbd_core/mbg_central/",
  mbg_functions
))
}

}
\seealso{
This function is used by:
  \code{\link{load_mbg_functions}}

Other MBG setup functions: \code{\link{is_integer}},
  \code{\link{is_lbd_core_repo}},
  \code{\link{is_lbd_singularity}},
  \code{\link{is_rstudio}}, \code{\link{is_singularity}},
  \code{\link{load_R_packages}},
  \code{\link{load_mbg_functions}},
  \code{\link{load_setthreads}}, \code{\link{mbg_setup}},
  \code{\link{set_original_threads}},
  \code{\link{set_serial_threads}}
}
\author{
Ian M. Davis, \email{imdavis@uw.edu}
}
\concept{MBG setup functions}
