% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_geo_names.R
\name{split_geo_names}
\alias{split_geo_names}
\title{FUNCTION_TITLE}
\usage{
split_geo_names(condSim_object)
}
\arguments{
\item{condSim_object}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
