% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGeoNames.R
\name{splitGeoNames}
\alias{splitGeoNames}
\title{FUNCTION_TITLE}
\usage{
splitGeoNames(geo)
}
\arguments{
\item{geo}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
