% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_link_table.R
\name{standard_link_table}
\alias{standard_link_table}
\title{Load the standard link table and return it}
\usage{
standard_link_table(admin_version)
}
\arguments{
\item{admin_version}{the administrative shapefile version. Either "current" or
a valid YYYY_MM_DD string.}
}
\value{
the data.table stored in the RDS file.
}
\description{
Loads the lbd_standard_link.rds file associated with an admin shapefile.
}
