% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_raking_link_pieces.R
\name{standard_raking_link_pieces}
\alias{standard_raking_link_pieces}
\title{Assemble and return metadata for LBD standard raking.}
\usage{
standard_raking_link_pieces(shapefile_version = "current",
  countries_not_to_subnat_rake = NULL)
}
\arguments{
\item{shapefile_version}{Shapefile version. Default: 'current'}

\item{countries_not_to_subnat_rake}{as it sounds. Replaces subnational metadata with national metadata. Default: NULL}
}
\value{
A list with admin_codes and code_locid
\item{admin_codes}{A named list with keys for all administrative levels
we rake to, likely "ADM0_CODE" and "ADM1_CODE". The values associated are
vectors containing all the admin codes we rake to.}
\item{code_locid}{A data.table with two columns: admin_code and loc_id.
This is a simplification of the loc_id, ADM0_CODE, and ADM1_CODE fields.
E.g., for geographies raked to admin 0 the admin_code column will contain
relevant ADM0_CODE corresponding to loc_id values.}
}
\description{
Assembles metadata from the LBD standard raking shapefile for
use in building raster layers based upon pixel coverage and not centroid
ownership.
}
