% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startLog.R
\name{startLog}
\alias{startLog}
\title{FUNCTION_TITLE}
\usage{
startLog(file = "full_run.log")
}
\arguments{
\item{file}{PARAM_DESCRIPTION, Default: 'full_run.log'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
