% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_time.R
\name{strip_time}
\alias{strip_time}
\title{FUNCTION_TITLE}
\usage{
strip_time(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
