% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_nat_link_merge.R
\name{sub_nat_link_merge}
\alias{sub_nat_link_merge}
\title{Merge link table and GBD geographies}
\usage{
sub_nat_link_merge(rake_subnational, link, connector,
  nat_connector = NULL, countries_not_to_subnat_rake = NULL)
}
\arguments{
\item{rake_subnational}{T/F are we using the subnational targets for raking.  What targets get used is in the raking shp}

\item{link}{the link table}

\item{connector}{the table identifying what geographies go to what raking targets}

\item{nat_connector}{a connector object from `get_gbd_locs()` with `rake_subnational = F`. Used to replace subnationals specified in `countries_not_to_subnat_rake`}

\item{countries_not_to_subnat_rake}{a character vector of iso3 codes gotten from the config. All countries in this list will be raked nationally.}
}
\value{
a link table that has not only the spatial hierarchy in it but also a column for raking geographies.
}
\description{
This merges the link table and the list of GBD geographies so that we know what aggregations to use for raking.
}
