% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_results_table.R
\name{submit_results_table}
\alias{submit_results_table}
\title{FUNCTION_TITLE}
\usage{
submit_results_table(reg, indicator, indicator_group, run_date,
  pop_measure, repo, log_dir, baseline_year, goal_threshold,
  diarrhea_measure = "", target_type = "less", geos_node = F,
  shapefile_version = "current")
}
\arguments{
\item{reg}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{pop_measure}{PARAM_DESCRIPTION}

\item{repo}{PARAM_DESCRIPTION}

\item{log_dir}{PARAM_DESCRIPTION}

\item{baseline_year}{PARAM_DESCRIPTION}

\item{goal_threshold}{PARAM_DESCRIPTION}

\item{diarrhea_measure}{PARAM_DESCRIPTION, Default: ''}

\item{target_type}{PARAM_DESCRIPTION, Default: 'less'}

\item{geos_node}{PARAM_DESCRIPTION, Default: F}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
