% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_validation_report.R
\name{submit_validation_report}
\alias{submit_validation_report}
\title{FUNCTION_TITLE}
\usage{
submit_validation_report(reg, indicator, indicator_group, run_date,
  pop_measure, repo = core_repo, log_dir, target_type = "greater",
  target = 0.5, geo_nodes = FALSE, shapefile_version = "current")
}
\arguments{
\item{reg}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{pop_measure}{PARAM_DESCRIPTION}

\item{repo}{PARAM_DESCRIPTION, Default: core_repo}

\item{log_dir}{PARAM_DESCRIPTION}

\item{target_type}{PARAM_DESCRIPTION, Default: 'greater'}

\item{target}{PARAM_DESCRIPTION, Default: 0.5}

\item{geo_nodes}{PARAM_DESCRIPTION, Default: FALSE}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
