% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_validation_report_pub.R
\name{submit_validation_report_pub}
\alias{submit_validation_report_pub}
\title{FUNCTION_TITLE}
\usage{
submit_validation_report_pub(indicator, indicator_proper_name,
  indicator_group, run_date, pop_measure, repo, region, nperiod = 16,
  log_dir, geo_nodes = FALSE, shapefile_verison = "current")
}
\arguments{
\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_proper_name}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{pop_measure}{PARAM_DESCRIPTION}

\item{repo}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION}

\item{nperiod}{PARAM_DESCRIPTION, Default: 16}

\item{log_dir}{PARAM_DESCRIPTION}

\item{geo_nodes}{PARAM_DESCRIPTION, Default: FALSE}

\item{shapefile_verison}{PARAM_DESCRIPTION, Default: 'current'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
