% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnational_nid_subset.R
\name{subnational_nid_subset}
\alias{subnational_nid_subset}
\title{FUNCTION_TITLE}
\usage{
subnational_nid_subset(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
