% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractYears.R
\name{subtractYears}
\alias{subtractYears}
\title{FUNCTION_TITLE}
\usage{
subtractYears(Date, years)
}
\arguments{
\item{Date}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
