% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_admin2.R
\name{summarize_admin2}
\alias{summarize_admin2}
\title{FUNCTION_TITLE}
\usage{
summarize_admin2(gaul, indicator, indicator_group, run_date, nperiod,
  master_list)
}
\arguments{
\item{gaul}{PARAM_DESCRIPTION}

\item{indicator}{PARAM_DESCRIPTION}

\item{indicator_group}{PARAM_DESCRIPTION}

\item{run_date}{PARAM_DESCRIPTION}

\item{nperiod}{PARAM_DESCRIPTION}

\item{master_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[seegSDM]{character(0)}}
}
