% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_admins.R
\name{summarize_admins}
\alias{summarize_admins}
\title{FUNCTION_TITLE}
\usage{
summarize_admins(ind = indicator, ig = indicator_group,
  summstats = c("mean", "lower", "upper", "cirange"), raked = c(T, F),
  ad_levels = c(0, 1, 2), file_addin = NULL, ...)
}
\arguments{
\item{ind}{PARAM_DESCRIPTION, Default: indicator}

\item{ig}{PARAM_DESCRIPTION, Default: indicator_group}

\item{summstats}{PARAM_DESCRIPTION, Default: c("mean", "lower", "upper", "cirange")}

\item{raked}{PARAM_DESCRIPTION, Default: c(T, F)}

\item{ad_levels}{PARAM_DESCRIPTION, Default: c(0, 1, 2)}

\item{file_addin}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
