% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronize_shapefile_directories.R
\name{synchronize_shapefile_directories}
\alias{synchronize_shapefile_directories}
\title{FUNCTION_TITLE}
\usage{
synchronize_shapefile_directories(cores = NULL, verbose = F,
  in_dir = "/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory/",
  out_dir = "/share/geospatial/rds_shapefiles/")
}
\arguments{
\item{cores}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: F}

\item{in_dir}{PARAM_DESCRIPTION, Default: '/home/j/WORK/11_geospatial/05_survey shapefile library/Shapefile directory/'}

\item{out_dir}{PARAM_DESCRIPTION, Default: '/share/geospatial/rds_shapefiles/'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_match}}
}
