% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetYear.R
\name{targetYear}
\alias{targetYear}
\title{FUNCTION_TITLE}
\usage{
targetYear(period, period_end, width = 60)
}
\arguments{
\item{period}{PARAM_DESCRIPTION}

\item{period_end}{PARAM_DESCRIPTION}

\item{width}{PARAM_DESCRIPTION, Default: 60}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
