% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_rho_priors.R
\name{test_rho_priors}
\alias{test_rho_priors}
\title{FUNCTION_TITLE}
\usage{
test_rho_priors(temporal_model_theta1_prior)
}
\arguments{
\item{temporal_model_theta1_prior}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
