% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_empty.R
\name{theme_empty}
\alias{theme_empty}
\title{An empty `ggplot()` theme}
\usage{
theme_empty()
}
\value{
theme object for ggplot
}
\description{
An empty `ggplot()` theme, to be added to a \code{ggplot} object
}
\examples{
\dontrun{
ggplot(data = data, aes(x = x, y = y)) +
  geom_point() +
  theme_empty()
}
}
