% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyInstall.R
\name{tidyInstall}
\alias{tidyInstall}
\title{FUNCTION_TITLE}
\usage{
tidyInstall(packages, versions = NULL, lib = .libPaths()[1])
}
\arguments{
\item{packages}{PARAM_DESCRIPTION}

\item{versions}{PARAM_DESCRIPTION, Default: NULL}

\item{lib}{PARAM_DESCRIPTION, Default: .libPaths()[1]}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[versions]{installed.versions}}
 \code{\link[withr]{with_libpaths}}
 \code{\link[devtools]{reexports}}
}
