% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Create time series plots at the specified admin level, faceting by the number of admin levels in that geography}
\usage{
time_series(model_data, admin = 0, val_range = c(0, 1),
  title_plot_size = 30, ind_title = "")
}
\arguments{
\item{model_data}{admin level data table from the standard aggregation code (typically ad0_df/ad1_df/ad2_df specified to certain geographic area)}

\item{admin}{specified admin level (0/1/2)}

\item{val_range}{Range of values for y axis}

\item{title_plot_size}{Size of plot title}

\item{input_data}{input data collapsed to admin level (typically ad0_data/ad1_data/ad2_data specified to certain geographic area)}
}
\value{
a ggplot object
}
\description{
This function is called in the subnational_ts_function and is a way to generalize making time
series for each admin unit. These functions are all very similar but come in four flavors:


1) time_series:  This plots the model predictions along with the uncertainty, faceted by the admin unit
2) time_series_data: This plots the model predictions and uncertainty but adds the data aggregated to that admin level
(using input_aggregated_admin function), and it colors the points by the data source, as well as differentiating between
survey size and point and polygon with size and shape aesthetics respectively
3) time_series_multiple: This plots multiple model predictions (either different model runs or multiple indicators, specified by model_runs = T in subnational_ts_plots)
on each time trend at the specified admin unit (0/1/2)
4) time_series_multiple_data: This plots multiple model predictions and also includes the aggregated input data to that admin unit.
I only recommend using this if comparing different model runs of the same indicator, as plotting the data and multiple indicators makes the plots too busy
}
\examples{
\dontrun{
## ad0_df_reg is the ad0_df specified to specific region, say for example 'cssa' (Central Sub-Saharan Africa)
if (plot_data == F & multiple_runs == F) ad0_time_series <- time_series(ad0_df_reg, admin = 0, val_range = c(0, 1))
}
}
\author{
Michael Cork, \email{mcork23@uw.edu}
}
