% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_multiple.R
\name{time_series_multiple}
\alias{time_series_multiple}
\title{FUNCTION_TITLE}
\usage{
time_series_multiple(model_data, admin = 0, val_range = c(0, 1),
  title_plot_size = 30, ind_title = "")
}
\arguments{
\item{model_data}{PARAM_DESCRIPTION}

\item{admin}{PARAM_DESCRIPTION, Default: 0}

\item{val_range}{PARAM_DESCRIPTION, Default: c(0, 1)}

\item{title_plot_size}{PARAM_DESCRIPTION, Default: 30}

\item{ind_title}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
