% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_generate.R
\name{title_generate}
\alias{title_generate}
\title{FUNCTION_TITLE}
\usage{
title_generate(ind_title, year_list, admin = 0, ad0_reg_title = "",
  ctry = "", ad1_df_ad1 = "")
}
\arguments{
\item{ind_title}{PARAM_DESCRIPTION}

\item{year_list}{PARAM_DESCRIPTION}

\item{admin}{PARAM_DESCRIPTION, Default: 0}

\item{ad0_reg_title}{PARAM_DESCRIPTION, Default: ''}

\item{ctry}{PARAM_DESCRIPTION, Default: ''}

\item{ad1_df_ad1}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
