% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toYear.R
\name{toYear}
\alias{toYear}
\title{FUNCTION_TITLE}
\usage{
toYear(Date)
}
\arguments{
\item{Date}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
