% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_theta.R
\name{transform_theta}
\alias{transform_theta}
\title{FUNCTION_TITLE}
\usage{
transform_theta(theta)
}
\arguments{
\item{theta}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
