% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlock_file.R
\name{unlock_file}
\alias{unlock_file}
\title{FUNCTION_TITLE}
\usage{
unlock_file(shp,
  fast_shapefile_dir = "/share/geospatial/rds_shapefiles/")
}
\arguments{
\item{shp}{PARAM_DESCRIPTION}

\item{fast_shapefile_dir}{PARAM_DESCRIPTION, Default: '/share/geospatial/rds_shapefiles/'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
