% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cov_layers.R
\name{update_cov_layers}
\alias{update_cov_layers}
\title{FUNCTION_TITLE}
\usage{
update_cov_layers(original_layers, new_layers)
}
\arguments{
\item{original_layers}{PARAM_DESCRIPTION}

\item{new_layers}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
unction to update the cov layers option
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
