% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/update_fixed_effect_config_with_missing_release.R
\name{update_fixed_effect_config_with_missing_release}
\alias{update_fixed_effect_config_with_missing_release}
\title{Update fixed effect covariate configuration with optional/missing data}
\usage{
update_fixed_effect_config_with_missing_release(fixed_effect_config)
}
\arguments{
\item{fixed_effect_config}{the fixed effect configuration, a data.table}
}
\value{
NULL (the fixed_effect_config is modified in-place)
}
\description{
Updates the fixed effect configuration with \emph{release} values for each
covariate/measure pair if it is not provided by using the most recent available release.
}
