% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lv_table.R
\name{update_lv_table}
\alias{update_lv_table}
\title{FUNCTION_TITLE}
\usage{
update_lv_table(lv)
}
\arguments{
\item{lv}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
