% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_config_to_db.R
\name{upload_config_to_db}
\alias{upload_config_to_db}
\title{Uploading config info to SQLite database}
\usage{
upload_config_to_db(config, user, core_repo, indicator_group, indicator,
  run_date, verbose = FALSE)
}
\arguments{
\item{config}{The config \code{data.table} prepped apriori.}

\item{user}{User name}

\item{core_repo}{Path to core repository}

\item{indicator_group}{Indicator group}

\item{indicator}{Indicator}

\item{run_date}{Run date datestamp}

\item{verbose}{Print config before exiting?}
}
\description{
This function will upload the MBG config info into a SQLite database. There are no primary keys set, so any number of duplicate entries can be recorded.
}
