% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_pull_adm0_codes.R
\name{utility_pull_adm0_codes}
\alias{utility_pull_adm0_codes}
\title{FUNCTION_TITLE}
\usage{
utility_pull_adm0_codes(region_codes, standard_regex)
}
\arguments{
\item{region_codes}{PARAM_DESCRIPTION}

\item{standard_regex}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
