% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_raking_factors.R
\name{val_raking_factors}
\alias{val_raking_factors}
\title{FUNCTION_TITLE}
\usage{
val_raking_factors(rfs, shapefile_version = "current", gaul_list)
}
\arguments{
\item{rfs}{PARAM_DESCRIPTION}

\item{shapefile_version}{PARAM_DESCRIPTION, Default: 'current'}

\item{gaul_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
