% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_node_option.R
\name{validate_node_option}
\alias{validate_node_option}
\title{Validate node option}
\usage{
validate_node_option(use_geo_nodes, use_c2_nodes, project)
}
\arguments{
\item{use_geo_nodes}{logical indicating whether to use geospatial nodes.}

\item{use_c2_nodes}{logical indicating whether to use "c2-" prefixed nodes.}
}
\description{
Validates node selection emitting warning message if inconsistent.
}
