% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_queue_string.R
\name{validate_queue_string}
\alias{validate_queue_string}
\title{Validate user-supplied queue string}
\usage{
validate_queue_string(queue)
}
\arguments{
\item{queue}{String. User supplied queue.}
}
\value{
the queue post-validating
}
\description{
Check the input string of queue and validate whether its a legitimate queue or not.
This checks for whether the queue values are matching with \code{all}, \code{geospatial} or \code{long}
}
