% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_singularity_options.R
\name{validate_singularity_options}
\alias{validate_singularity_options}
\title{Validate Singularity options}
\usage{
validate_singularity_options(singularity, singularity_opts)
}
\arguments{
\item{singularity}{the string name of the singularity image to use.}

\item{singularity_opts}{named list of options.}
}
\description{
Validates singularity-related options and errors if invalid.
}
