% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplayout.R
\name{vplayout}
\alias{vplayout}
\title{FUNCTION_TITLE}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
