% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_for_lock.R
\name{wait_for_lock}
\alias{wait_for_lock}
\title{FUNCTION_TITLE}
\usage{
wait_for_lock(shp,
  fast_shapefile_dir = "/share/geospatial/rds_shapefiles/", tries = 20,
  interval = 30)
}
\arguments{
\item{shp}{PARAM_DESCRIPTION}

\item{fast_shapefile_dir}{PARAM_DESCRIPTION, Default: '/share/geospatial/rds_shapefiles/'}

\item{tries}{PARAM_DESCRIPTION, Default: 20}

\item{interval}{PARAM_DESCRIPTION, Default: 30}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
