% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitforaggregation.R
\name{waitforaggregation}
\alias{waitforaggregation}
\title{FUNCTION_TITLE}
\usage{
waitforaggregation(rd = run_date, indic = indicator,
  ig = indicator_group, ages, regions, holdouts, raked,
  dir_to_search = NULL)
}
\arguments{
\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}

\item{ig}{PARAM_DESCRIPTION, Default: indicator_group}

\item{ages}{PARAM_DESCRIPTION}

\item{regions}{PARAM_DESCRIPTION}

\item{holdouts}{PARAM_DESCRIPTION}

\item{raked}{PARAM_DESCRIPTION}

\item{dir_to_search}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
