% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitformodelstofinish.R
\name{waitformodelstofinish}
\alias{waitformodelstofinish}
\title{FUNCTION_TITLE}
\usage{
waitformodelstofinish(sleeptime = 100,
  path = paste0("/share/geospatial/mbg/", indicator_group, "/",
  indicator, "/output/", run_date), rd = run_date, lv = loopvars,
  showfiles = TRUE, showcluster = FALSE)
}
\arguments{
\item{sleeptime}{PARAM_DESCRIPTION, Default: 100}

\item{path}{PARAM_DESCRIPTION, Default: paste0("/share/geospatial/mbg/", indicator_group, "/", indicator,
"/output/", run_date)}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{lv}{PARAM_DESCRIPTION, Default: loopvars}

\item{showfiles}{PARAM_DESCRIPTION, Default: TRUE}

\item{showcluster}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
