% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitformodelstofinishu5m.R
\name{waitformodelstofinishu5m}
\alias{waitformodelstofinishu5m}
\title{FUNCTION_TITLE}
\usage{
waitformodelstofinishu5m(sleeptime = 100,
  path = paste0("/share/geospatial/mbg/", indicator_group, "/",
  indicator, "/output/", run_date, "/"), rd = run_date, lv = loopvars,
  ageasindic = TRUE)
}
\arguments{
\item{sleeptime}{PARAM_DESCRIPTION, Default: 100}

\item{path}{PARAM_DESCRIPTION, Default: paste0("/share/geospatial/mbg/", indicator_group, "/", indicator,
"/output/", run_date, "/")}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{lv}{PARAM_DESCRIPTION, Default: loopvars}

\item{ageasindic}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
