% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitforpostesttofinish.R
\name{waitforpostesttofinish}
\alias{waitforpostesttofinish}
\title{FUNCTION_TITLE}
\usage{
waitforpostesttofinish(sleeptime = 100, indicator = indicator,
  indicator_group = indicator_group, run_date = run_date, strata,
  showfiles = TRUE)
}
\arguments{
\item{sleeptime}{PARAM_DESCRIPTION, Default: 100}

\item{indicator}{PARAM_DESCRIPTION, Default: indicator}

\item{indicator_group}{PARAM_DESCRIPTION, Default: indicator_group}

\item{run_date}{PARAM_DESCRIPTION, Default: run_date}

\item{strata}{PARAM_DESCRIPTION}

\item{showfiles}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
