% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitforresultstable.R
\name{waitforresultstable}
\alias{waitforresultstable}
\title{FUNCTION_TITLE}
\usage{
waitforresultstable(st = strata, rd = run_date, indic = indicator,
  ig = indicator_group, baseline_year = 2000, measure = "prevalence")
}
\arguments{
\item{st}{PARAM_DESCRIPTION, Default: strata}

\item{rd}{PARAM_DESCRIPTION, Default: run_date}

\item{indic}{PARAM_DESCRIPTION, Default: indicator}

\item{ig}{PARAM_DESCRIPTION, Default: indicator_group}

\item{baseline_year}{PARAM_DESCRIPTION, Default: 2000}

\item{measure}{PARAM_DESCRIPTION, Default: 'prevalence'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_match}}
}
