% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_rmse.R
\name{weighted.rmse}
\alias{weighted.rmse}
\title{FUNCTION_TITLE}
\usage{
weighted.rmse(error, w)
}
\arguments{
\item{error}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
