% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_folds.R
\name{yr_folds}
\alias{yr_folds}
\title{FUNCTION_TITLE}
\usage{
yr_folds(xy, yr, ss = 1, n_folds = 5, ...)
}
\arguments{
\item{xy}{PARAM_DESCRIPTION}

\item{yr}{PARAM_DESCRIPTION}

\item{ss}{PARAM_DESCRIPTION, Default: 1}

\item{n_folds}{PARAM_DESCRIPTION, Default: 5}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
