% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mbg_data_stack.R
\name{build_mbg_data_stack}
\alias{build_mbg_data_stack}
\title{Build an INLA data stack for modeling}
\usage{
build_mbg_data_stack(df, fixed_effects, mesh_s, mesh_t, exclude_cs = "",
  spde_prior = "list(type='nonpc')", coefs.sum1 = FALSE,
  use_ctry_res = FALSE, use_subnat_res = FALSE,
  remove_non_subnats = FALSE, use_nugget = FALSE,
  stacker_names = child_model_names, yl = year_list, zl = z_list,
  zcol = zcol, scale_gaussian_variance_N = TRUE, tmb = FALSE,
  use_gp = TRUE, use_space_only_gp = FALSE,
  use_time_only_gmrf = FALSE, shapefile_version = "current",
  st_gp_int_zero = FALSE, s_gp_int_zero = FALSE)
}
\arguments{
\item{df}{data.frame/table containing observational data. must have
latitude and longitude columns as well as extracted values for
all covariates that will be used in modeling}

\item{fixed_effects}{string containing fixed effects used in
res_fit separated by " + "}

\item{mesh_s}{An 'inla.mesh' object used to fit the spatial SPDE GP
approximation in res_fit}

\item{mesh_t}{An 'inla.mesh' object used to fit the temporal
correlation structure in res_fit}

\item{exclude_cs}{vector of strings detailing covariates that are
not center-scaled}

\item{spde_prior}{String containing a list. Specifies the type of prior used
for the Matern model parameters. If type = "nonpc" then use non pc 
prior. User can specify nominal prior means: spde_prior$prior$range.nominal 
and spde_prior$prior$variance.nominal. Defaults are INLA defaults, that is 
20\% of the range of the mesh and 1, respectively. If type = "pc" then use 
pc prior.  User can specify spde_prior$prior$range = (range0, Prange) i.e. 
P(range < range0) = Prange and spde_prior$prior$sigma = (sigma0, Psigma) 
i.e. P(sigma > sigma0) = Psigma. Defaults for Prange and Psigma are 0.05. 
Default for range0 is 5\% of the range of the mesh and sigma0=3.}

\item{coefs.sum1}{Logical. If TRUE, add a constraint to ensure
covariate coefs sum to 1 in fitted model}

\item{use_ctry_res}{Logical. If TRUE, include country random
effects}

\item{use_subnat_res}{Logical. If TRUE, include subnational (admin-1) random
effects}

\item{remove_non_subnats}{Logical. If TRUE, NA out all points of the subnational
random effect which are not associated with the country}

\item{use_nugget}{Logical. If TRUE, include a nugget effect}

\item{stacker_names}{string vector of child model names}

\item{yl}{numeric vector of years in model}

\item{zl}{numeric vector of third dimension in kronecker
product. Only implemented in TMB}

\item{zcol}{column name of z values associated with zl. Only
implemented in TMB}

\item{scale_gaussian_variance_N}{Logical. Do you want to scale
gaussian variance by sample size? Only implemented in TMB.}

\item{tmb}{Logical. use tmb?}

\item{use_space_only_gp}{Logical. include a space only (time stationary) gp}

\item{use_time_only_gmrf}{Logical. include a time only (time stationary) gp}

\item{st_gp_int_zero}{Logical. Should the space-time GP be forced to interate to 0? Default: FALSE}

\item{s_gp_int_zero}{Logical. Should the space GP be forced to interate to 0? Only used if use_space_only_gp=TRUE. Default: FALSE}
}
\value{
List containing 1) 'inla.stack' object, 2) 'inla.spde'
  object, 3) cs_df: a center-scale dataframe containing info on
  mean and SD used to scale covaraites in design matrix
}
\description{
Generates an INLA stack that contains information
  mapping data to various INLA effects (GP, random effects, ...)
}
